The ConsensusPartition class

The ConsensusPartition class

Methods

The ConsensusPartition-class has following methods:

consensus_partition:

constructor method, run consensus partitioning with a specified top-value method and a partitioning method.

select_partition_number,ConsensusPartition-method:

make a list of plots for selecting optimized number of subgroups.

consensus_heatmap,ConsensusPartition-method:

make heatmap of the consensus matrix.

membership_heatmap,ConsensusPartition-method:

make heatmap of the membership for individual partitions.

get_signatures,ConsensusPartition-method:

get the signature rows and make heatmap.

dimension_reduction,ConsensusPartition-method:

make dimension reduction plots.

collect_plots,ConsensusPartition-method:

make heatmaps for consensus matrix and membership matrix with different number of subgroups.

collect_classes,ConsensusPartition-method:

make heatmap with subgroups with different numbers.

get_param,ConsensusPartition-method:

get parameters for the consensus clustering.

get_matrix,ConsensusPartition-method:

get the original matrix.

get_consensus,ConsensusPartition-method:

get the consensus matrix.

get_membership,ConsensusPartition-method:

get the membership of partitions generated from random samplings.

get_stats,ConsensusPartition-method:

get statistics for the consensus partitioning.

get_classes,ConsensusPartition-method:

get the consensus subgroup labels and other columns.

suggest_best_k,ConsensusPartition-method:

guess the best number of subgroups.

test_to_known_factors,ConsensusPartition-method:

test correlation between predicted subgroups and known factors, if available.

cola_report,ConsensusPartition-method:

generate a HTML report for the whole analysis.

functional_enrichment,ConsensusPartition-method:

perform functional enrichment analysis on significant genes if rows in the matrix can be corresponded to genes.

Author

Zuguang Gu <z.gu@dkfz.de>

Examples

# There is no example
NULL
#> NULL