Subset a ConsensusPartitionList object

Subset a ConsensusPartitionList object

# S3 method for ConsensusPartitionList
[(x, i, j, drop = TRUE)

Arguments

x
A ConsensusPartitionList-class object.

i
Index for top-value methods, character or nummeric.

j
Index for partition methods, character or nummeric.

drop
Whether drop class

Details

For a specific combination of top-value method and partition method, you can also subset by e.g. x['sd:hclust'].

Value

A ConsensusPartitionList-class object or a ConsensusPartition-class object.

Examples

data(cola_rl)
cola_rl[c("sd", "MAD"), c("hclust", "kmeans")]

#> A 'ConsensusPartitionList' object with 4 methods.
#>   On a matrix with 60 rows and 60 columns.
#>   Top rows are extracted by 'sd, MAD' methods.
#>   Subgroups are detected by 'hclust, kmeans' method.
#>   Number of partitions are tried for k = 2, 3, 4, 5, 6.
#>   Performed in total 3000 partitions by row resampling.
#> 
#> Following methods can be applied to this 'ConsensusPartitionList' object:
#>  [1] "cola_report"           "collect_classes"       "collect_plots"        
#>  [4] "collect_stats"         "colnames"              "functional_enrichment"
#>  [7] "get_anno_col"          "get_anno"              "get_classes"          
#> [10] "get_matrix"            "get_membership"        "get_stats"            
#> [13] "is_best_k"             "is_stable_k"           "ncol"                 
#> [16] "nrow"                  "rownames"              "show"                 
#> [19] "suggest_best_k"        "test_to_known_factors" "top_rows_heatmap"     
#> [22] "top_rows_overlap"     
#> 
#> You can get result for a single method by, e.g. object["sd", "hclust"] or object["sd:hclust"]
#> or a subset of methods by object[c("sd", "MAD")], c("hclust", "kmeans")]

cola_rl["sd", "kmeans"] # a ConsensusPartition object

#> A 'ConsensusPartition' object with k = 2, 3, 4, 5, 6.
#>   On a matrix with 60 rows and 60 columns.
#>   Top rows (20, 30, 40) are extracted by 'sd' method.
#>   Subgroups are detected by 'kmeans' method.
#>   Performed in total 750 partitions by row resampling.
#>   Best k for subgroups seems to be 3.
#> 
#> Following methods can be applied to this 'ConsensusPartition' object:
#>  [1] "cola_report"             "collect_classes"        
#>  [3] "collect_plots"           "collect_stats"          
#>  [5] "colnames"                "compare_signatures"     
#>  [7] "consensus_heatmap"       "dimension_reduction"    
#>  [9] "functional_enrichment"   "get_anno_col"           
#> [11] "get_anno"                "get_classes"            
#> [13] "get_consensus"           "get_matrix"             
#> [15] "get_membership"          "get_param"              
#> [17] "get_signatures"          "get_stats"              
#> [19] "is_best_k"               "is_stable_k"            
#> [21] "membership_heatmap"      "ncol"                   
#> [23] "nrow"                    "plot_ecdf"              
#> [25] "rownames"                "select_partition_number"
#> [27] "show"                    "suggest_best_k"         
#> [29] "test_to_known_factors"  

cola_rl["sd:kmeans"] # a ConsensusPartition object

#> A 'ConsensusPartition' object with k = 2, 3, 4, 5, 6.
#>   On a matrix with 60 rows and 60 columns.
#>   Top rows (20, 30, 40) are extracted by 'sd' method.
#>   Subgroups are detected by 'kmeans' method.
#>   Performed in total 750 partitions by row resampling.
#>   Best k for subgroups seems to be 3.
#> 
#> Following methods can be applied to this 'ConsensusPartition' object:
#>  [1] "cola_report"             "collect_classes"        
#>  [3] "collect_plots"           "collect_stats"          
#>  [5] "colnames"                "compare_signatures"     
#>  [7] "consensus_heatmap"       "dimension_reduction"    
#>  [9] "functional_enrichment"   "get_anno_col"           
#> [11] "get_anno"                "get_classes"            
#> [13] "get_consensus"           "get_matrix"             
#> [15] "get_membership"          "get_param"              
#> [17] "get_signatures"          "get_stats"              
#> [19] "is_best_k"               "is_stable_k"            
#> [21] "membership_heatmap"      "ncol"                   
#> [23] "nrow"                    "plot_ecdf"              
#> [25] "rownames"                "select_partition_number"
#> [27] "show"                    "suggest_best_k"         
#> [29] "test_to_known_factors"  

cola_rl[["sd:kmeans"]] # a ConsensusPartition object

#> A 'ConsensusPartition' object with k = 2, 3, 4, 5, 6.
#>   On a matrix with 60 rows and 60 columns.
#>   Top rows (20, 30, 40) are extracted by 'sd' method.
#>   Subgroups are detected by 'kmeans' method.
#>   Performed in total 750 partitions by row resampling.
#>   Best k for subgroups seems to be 3.
#> 
#> Following methods can be applied to this 'ConsensusPartition' object:
#>  [1] "cola_report"             "collect_classes"        
#>  [3] "collect_plots"           "collect_stats"          
#>  [5] "colnames"                "compare_signatures"     
#>  [7] "consensus_heatmap"       "dimension_reduction"    
#>  [9] "functional_enrichment"   "get_anno_col"           
#> [11] "get_anno"                "get_classes"            
#> [13] "get_consensus"           "get_matrix"             
#> [15] "get_membership"          "get_param"              
#> [17] "get_signatures"          "get_stats"              
#> [19] "is_best_k"               "is_stable_k"            
#> [21] "membership_heatmap"      "ncol"                   
#> [23] "nrow"                    "plot_ecdf"              
#> [25] "rownames"                "select_partition_number"
#> [27] "show"                    "suggest_best_k"         
#> [29] "test_to_known_factors"  

cola_rl["sd", "kmeans", drop = FALSE] # still a ConsensusPartitionList object

#> A 'ConsensusPartitionList' object with 1 methods.
#>   On a matrix with 60 rows and 60 columns.
#>   Top rows are extracted by 'sd' methods.
#>   Subgroups are detected by 'kmeans' method.
#>   Number of partitions are tried for k = 2, 3, 4, 5, 6.
#>   Performed in total 750 partitions by row resampling.
#> 
#> Following methods can be applied to this 'ConsensusPartitionList' object:
#>  [1] "cola_report"           "collect_classes"       "collect_plots"        
#>  [4] "collect_stats"         "colnames"              "functional_enrichment"
#>  [7] "get_anno_col"          "get_anno"              "get_classes"          
#> [10] "get_matrix"            "get_membership"        "get_stats"            
#> [13] "is_best_k"             "is_stable_k"           "ncol"                 
#> [16] "nrow"                  "rownames"              "show"                 
#> [19] "suggest_best_k"        "test_to_known_factors" "top_rows_heatmap"     
#> [22] "top_rows_overlap"     
#> 
#> You can get result for a single method by, e.g. object["sd", "kmeans"] or object["sd:kmeans"]

cola_rl["sd:kmeans", drop = FALSE] # still a ConsensusPartitionList object

#> A 'ConsensusPartitionList' object with 1 methods.
#>   On a matrix with 60 rows and 60 columns.
#>   Top rows are extracted by 'sd' methods.
#>   Subgroups are detected by 'kmeans' method.
#>   Number of partitions are tried for k = 2, 3, 4, 5, 6.
#>   Performed in total 750 partitions by row resampling.
#> 
#> Following methods can be applied to this 'ConsensusPartitionList' object:
#>  [1] "cola_report"           "collect_classes"       "collect_plots"        
#>  [4] "collect_stats"         "colnames"              "functional_enrichment"
#>  [7] "get_anno_col"          "get_anno"              "get_classes"          
#> [10] "get_matrix"            "get_membership"        "get_stats"            
#> [13] "is_best_k"             "is_stable_k"           "ncol"                 
#> [16] "nrow"                  "rownames"              "show"                 
#> [19] "suggest_best_k"        "test_to_known_factors" "top_rows_heatmap"     
#> [22] "top_rows_overlap"     
#> 
#> You can get result for a single method by, e.g. object["sd", "kmeans"] or object["sd:kmeans"]

cola_rl["sd", ]

#> A 'ConsensusPartitionList' object with 5 methods.
#>   On a matrix with 60 rows and 60 columns.
#>   Top rows are extracted by 'sd' methods.
#>   Subgroups are detected by 'hclust, kmeans, skmeans, pam, mclust' method.
#>   Number of partitions are tried for k = 2, 3, 4, 5, 6.
#>   Performed in total 3750 partitions by row resampling.
#> 
#> Following methods can be applied to this 'ConsensusPartitionList' object:
#>  [1] "cola_report"           "collect_classes"       "collect_plots"        
#>  [4] "collect_stats"         "colnames"              "functional_enrichment"
#>  [7] "get_anno_col"          "get_anno"              "get_classes"          
#> [10] "get_matrix"            "get_membership"        "get_stats"            
#> [13] "is_best_k"             "is_stable_k"           "ncol"                 
#> [16] "nrow"                  "rownames"              "show"                 
#> [19] "suggest_best_k"        "test_to_known_factors" "top_rows_heatmap"     
#> [22] "top_rows_overlap"     
#> 
#> You can get result for a single method by, e.g. object["sd", "hclust"] or object["sd:hclust"]
#> or a subset of methods by object["sd"], c("hclust", "kmeans")]

cola_rl[, "hclust"]

#> A 'ConsensusPartitionList' object with 4 methods.
#>   On a matrix with 60 rows and 60 columns.
#>   Top rows are extracted by 'sd, cv, MAD, ATC' methods.
#>   Subgroups are detected by 'hclust' method.
#>   Number of partitions are tried for k = 2, 3, 4, 5, 6.
#>   Performed in total 3000 partitions by row resampling.
#> 
#> Following methods can be applied to this 'ConsensusPartitionList' object:
#>  [1] "cola_report"           "collect_classes"       "collect_plots"        
#>  [4] "collect_stats"         "colnames"              "functional_enrichment"
#>  [7] "get_anno_col"          "get_anno"              "get_classes"          
#> [10] "get_matrix"            "get_membership"        "get_stats"            
#> [13] "is_best_k"             "is_stable_k"           "ncol"                 
#> [16] "nrow"                  "rownames"              "show"                 
#> [19] "suggest_best_k"        "test_to_known_factors" "top_rows_heatmap"     
#> [22] "top_rows_overlap"     
#> 
#> You can get result for a single method by, e.g. object["sd", "hclust"] or object["sd:hclust"]
#> or a subset of methods by object[c("sd", "cv")], "hclust"]

cola_rl[1:2, 1:2]

#> A 'ConsensusPartitionList' object with 4 methods.
#>   On a matrix with 60 rows and 60 columns.
#>   Top rows are extracted by 'sd, cv' methods.
#>   Subgroups are detected by 'hclust, kmeans' method.
#>   Number of partitions are tried for k = 2, 3, 4, 5, 6.
#>   Performed in total 3000 partitions by row resampling.
#> 
#> Following methods can be applied to this 'ConsensusPartitionList' object:
#>  [1] "cola_report"           "collect_classes"       "collect_plots"        
#>  [4] "collect_stats"         "colnames"              "functional_enrichment"
#>  [7] "get_anno_col"          "get_anno"              "get_classes"          
#> [10] "get_matrix"            "get_membership"        "get_stats"            
#> [13] "is_best_k"             "is_stable_k"           "ncol"                 
#> [16] "nrow"                  "rownames"              "show"                 
#> [19] "suggest_best_k"        "test_to_known_factors" "top_rows_heatmap"     
#> [22] "top_rows_overlap"     
#> 
#> You can get result for a single method by, e.g. object["sd", "hclust"] or object["sd:hclust"]
#> or a subset of methods by object[c("sd", "cv")], c("hclust", "kmeans")]