Remove rows with low variance and impute missing values

Remove rows with low variance and impute missing values

adjust_matrix(m, sd_quantile = 0.05, max_na = 0.25)

Arguments

m
A numeric matrix.

sd_quantile
Cutoff of the quantile of standard deviation. Rows with standard deviation less than it are removed.

max_na
Maximum NA fraction in each row. Rows with NA fraction larger than it are removed.

Details

The function uses impute.knn to impute missing values, then uses adjust_outlier to adjust outliers and removes rows with low standard deviations.

Value

A numeric matrix.

Examples

set.seed(123)
m = matrix(rnorm(100), nrow = 10)
m[sample(length(m), 5)] = NA
m[1, ] = 0
m

#>              [,1]       [,2]       [,3]        [,4]        [,5]        [,6]
#>  [1,]  0.00000000  0.0000000  0.0000000  0.00000000  0.00000000  0.00000000
#>  [2,] -0.23017749  0.3598138 -0.2179749 -0.29507148 -0.20791728 -0.02854676
#>  [3,]  1.55870831  0.4007715 -1.0260044  0.89512566 -1.26539635 -0.04287046
#>  [4,]  0.07050839  0.1106827 -0.7288912  0.87813349  2.16895597  1.36860228
#>  [5,]  0.12928774 -0.5558411 -0.6250393  0.82158108  1.20796200 -0.22577099
#>  [6,]  1.71506499         NA -1.6866933  0.68864025 -1.12310858  1.51647060
#>  [7,]  0.46091621  0.4978505  0.8377870  0.55391765 -0.40288484 -1.54875280
#>  [8,] -1.26506123 -1.9666172  0.1533731 -0.06191171 -0.46665535  0.58461375
#>  [9,] -0.68685285  0.7013559 -1.1381369 -0.30596266  0.77996512  0.12385424
#> [10,] -0.44566197 -0.4727914         NA -0.38047100 -0.08336907  0.21594157
#>              [,7]       [,8]       [,9]      [,10]
#>  [1,]  0.00000000  0.0000000  0.0000000  0.0000000
#>  [2,] -0.50232345 -2.3091689  0.3852804  0.5483970
#>  [3,] -0.33320738  1.0057385 -0.3706600  0.2387317
#>  [4,] -1.01857538 -0.7092008  0.6443765         NA
#>  [5,] -1.07179123 -0.6880086 -0.2204866  1.3606524
#>  [6,]  0.30352864  1.0255714  0.3317820 -0.6002596
#>  [7,]  0.44820978 -0.2847730  1.0968390  2.1873330
#>  [8,]  0.05300423 -1.2207177         NA  1.5326106
#>  [9,]  0.92226747  0.1813035         NA -0.2357004
#> [10,]  2.05008469 -0.1388914  1.1488076 -1.0264209

m2 = adjust_matrix(m)

#> There are NA values in the data, now impute missing data.

#> 1 rows have been removed with zero variance.

#> 1 rows have been removed with too low variance (sd < 0.05 quantile)

m2

#>             [,1]       [,2]       [,3]        [,4]        [,5]        [,6]
#> [1,]  1.30987191  0.4007715 -1.0260044  0.89512566 -1.15767000 -0.04287046
#> [2,]  0.07050839  0.1106827 -0.7288912  0.87813349  1.80879681  1.36860228
#> [3,]  0.12928774 -0.5558411 -0.6250393  0.82158108  1.20796200 -0.22577099
#> [4,]  1.62569751 -0.1370437 -1.4330802  0.68864025 -1.12310858  1.51647060
#> [5,]  0.46091621  0.4978505  0.8377870  0.55391765 -0.40288484 -1.03311222
#> [6,] -1.26506123 -1.6509170  0.1533731 -0.06191171 -0.46665535  0.58461375
#> [7,] -0.68685285  0.7013559 -0.9350591 -0.30596266  0.77996512  0.12385424
#> [8,] -0.44566197 -0.4727914 -0.4904371 -0.38047100 -0.08336907  0.21594157
#>             [,7]       [,8]       [,9]      [,10]
#> [1,] -0.33320738  1.0057385 -0.3706600  0.2387317
#> [2,] -0.88821751 -0.7092008  0.6443765  0.5154152
#> [3,] -0.89908905 -0.6880086 -0.2204866  1.2919417
#> [4,]  0.30352864  1.0255714  0.3317820 -0.6002596
#> [5,]  0.44820978 -0.2847730  1.0968390  1.6966107
#> [6,]  0.05300423 -1.2207177  0.3196720  1.1060120
#> [7,]  0.85823141  0.1813035  0.3196720 -0.2357004
#> [8,]  1.64451001 -0.1388914  1.1488076 -0.7852282