Make HTML report from the ConsensusPartitionList object

Make HTML report from the ConsensusPartitionList object

# S4 method for ConsensusPartitionList
cola_report(object, output_dir = getwd(), mc.cores = 1, cores = mc.cores,
    title = "cola Report for Consensus Partitioning", env = parent.frame())

Arguments

object
A ConsensusPartitionList-class object.

output_dir
The output directory where the report is saved.

mc.cores
Multiple cores to use. This argument will be removed in future versions.

cores
Number of cores, or a cluster object returned by makeCluster.

title
Title of the report.

env
Where the objects in the report are found, internally used.

Details

The ConsensusPartitionList-class object contains results for all combinations of top-value methods and partitioning methods. This function generates a HTML report which contains all plots and tables for every combination of method.

The report generation may take a while because it generates A LOT of heatmaps.

Examples of reports can be found at https://jokergoo.github.io/cola_collection/ .

Value

No value is returned.

Author

Zuguang Gu <z.gu@dkfz.de>

Examples

if(FALSE) {
# the following code is runnable
data(golub_cola)
cola_report(golub_cola[c("SD", "MAD"), c("hclust", "skmeans")], output_dir = "~/test_cola_cl_report")
}