Perform DAVID enrichment analysis

Perform DAVID enrichment analysis

david_enrichment(genes, email,
    catalog = c("GOTERM_CC_FAT", "GOTERM_BP_FAT", "GOTERM_MF_FAT", "KEGG_PATHWAY"),
    idtype = "ENSEMBL_GENE_ID", species = "Homo sapiens")

Arguments

genes

A vector of gene identifiers.

email

The email that user registered on DAVID web service (https://david.ncifcrf.gov/content.jsp?file=WS.html ).

catalog

A vector of function catalogs. Valid values should be in cola:::DAVID_ALL_CATALOGS.

idtype

ID types for the input gene list. Valid values should be in cola:::DAVID_ALL_ID_TYPES.

species

Full species name if the ID type is not uniquely mapped to one single species.

Details

This function directly sends the HTTP request to DAVID web service (https://david.ncifcrf.gov/content.jsp?file=WS.html ) and parses the returned XML. The reason of writing this function is I have problems with other R packages doing DAVID analysis (e.g. RDAVIDWebService, https://bioconductor.org/packages/devel/bioc/html/RDAVIDWebService.html ) because the rJava package RDAVIDWebService depends on can not be installed on my machine.

Users are encouraged to use more advanced gene set enrichment tools such as clusterProfiler (http://www.bioconductor.org/packages/release/bioc/html/clusterProfiler.html ), or fgsea (http://www.bioconductor.org/packages/release/bioc/html/fgsea.html ).

If you want to run this function multiple times, please set time intervals between runs.

Value

A data frame with functional enrichment results.

See also

Author

Zuguang Gu <z.gu@dkfz.de>

Examples

# There is no example
NULL
#> NULL