Perform functional enrichment on signature genes

Perform functional enrichment on signature genes

# S4 method for ANY
functional_enrichment(object,
    id_mapping = guess_id_mapping(object, org_db, verbose),
    org_db = "org.Hs.eg.db", ontology = "BP",
    min_set_size = 10, max_set_size = 1000,
    verbose = TRUE, prefix = "", ...)

Arguments

object
A vector of gene IDs.

id_mapping
If the gene IDs are not Entrez IDs, a named vector should be provided where the names are the gene IDs and values are the correspoinding Entrez IDs. The value can also be a function that converts gene IDs.

org_db
Annotation database.

ontology
Following ontologies are allowed: BP, CC, MF, KEGG, Reactome. MSigDb with the gmt file set by gmt_file argument, or gmt for general gmt gene sets.

min_set_size
The minimal size of the gene sets.

max_set_size
The maximal size of the gene sets.

verbose
Whether to print messages.

prefix
Used internally.

...
Pass to enrichGO, enrichKEGG, enricher, enrichDO or enrichPathway.

Details

The function enrichment is applied by clusterProfiler, DOSE or ReactomePA packages.

See also

Value

A data frame.

Examples

# There is no example
NULL

#> NULL