Perform functional enrichment on signature genes

Perform functional enrichment on signature genes

# S4 method for HierarchicalPartition
functional_enrichment(object, merge_node = merge_node_param(),
    gene_fdr_cutoff = cola_opt$fdr_cutoff,
    row_km = NULL, id_mapping = guess_id_mapping(rownames(object), org_db, verbose),
    org_db = "org.Hs.eg.db", ontology = "BP",
    min_set_size = 10, max_set_size = 1000,
    verbose = TRUE, ...)

Arguments

object

a HierarchicalPartition-class object from hierarchical_partition.

merge_node

Parameters to merge sub-dendrograms, see merge_node_param.

gene_fdr_cutoff

Cutoff of FDR to define significant signature genes.

row_km

Number of row clusterings by k-means to separate the matrix that only contains signatures.

id_mapping

If the gene IDs which are row names of the original matrix are not Entrez IDs, a named vector should be provided where the names are the gene IDs in the matrix and values are correspoinding Entrez IDs. The value can also be a function that converts gene IDs.

org_db

Annotation database.

ontology

See corresponding argumnet in functional_enrichment,ANY-method.

min_set_size

The minimal size of the gene sets.

max_set_size

The maximal size of the gene sets.

verbose

Whether to print messages.

...

Pass to functional_enrichment,ANY-method.

Details

For how to control the parameters of functional enrichment, see help page of functional_enrichment,ANY-method.

Value

A list of data frames which correspond to results for the functional ontologies:

Examples

# There is no example
NULL
#> NULL