Get subgroup labels

Get subgroup labels

# S4 method for ConsensusPartitionList
get_classes(object, k)

Arguments

object

A ConsensusPartitionList-class object.

k

Number of subgroups.

Details

The subgroup labels are inferred by merging partitions from all methods by weighting the mean silhouette scores in each method.

Value

A data frame with subgroup labels and other columns which are entropy of the percent membership matrix and the silhouette scores which measure the stability of a sample to stay in its group.

Author

Zuguang Gu <z.gu@dkfz.de>

Examples

data(golub_cola)
get_classes(golub_cola, k = 2)
#>           class    entropy silhouette
#> sample_39     1 0.83997042  0.7818134
#> sample_40     1 0.85493594  0.7818134
#> sample_42     1 0.38461700  0.8765723
#> sample_47     1 0.07530554  0.8702167
#> sample_48     1 0.00000000  0.8702167
#> sample_49     1 0.86529683  0.7818134
#> sample_41     1 0.00000000  0.8702167
#> sample_43     1 0.33400971  0.8765723
#> sample_44     1 0.00000000  0.8702167
#> sample_45     1 0.04943860  0.8702167
#> sample_46     1 0.14205866  0.8702167
#> sample_70     1 0.33918131  0.8765723
#> sample_71     1 0.44594517  0.8765723
#> sample_72     1 0.39220231  0.8765723
#> sample_68     1 0.00000000  0.8702167
#> sample_69     1 0.03477109  0.8702167
#> sample_67     2 0.87795666  0.5231260
#> sample_55     1 0.90478539  0.7266124
#> sample_56     1 0.83769035  0.7818134
#> sample_59     1 0.43714985  0.8765723
#> sample_52     2 0.08491589  0.9352635
#> sample_53     2 0.02840360  0.9352635
#> sample_51     2 0.02296957  0.9352635
#> sample_50     2 0.02296957  0.9352635
#> sample_54     2 0.75192484  0.7459714
#> sample_57     2 0.09072873  0.9352635
#> sample_58     2 0.05973264  0.9352635
#> sample_60     2 0.69448102  0.7459714
#> sample_61     2 0.07294518  0.9352635
#> sample_65     2 0.06861367  0.9352635
#> sample_66     1 0.97337105  0.3507463
#> sample_63     2 0.08491589  0.9352635
#> sample_64     2 0.31548609  0.9352635
#> sample_62     2 0.08491589  0.9352635
#> sample_1      1 0.75874454  0.8171256
#> sample_2      2 0.94857640  0.5231260
#> sample_3      1 0.87964906  0.7219835
#> sample_4      1 0.74417636  0.8171256
#> sample_5      1 0.00000000  0.8702167
#> sample_6      1 0.89516448  0.7219835
#> sample_7      1 0.87277257  0.7266124
#> sample_8      1 0.83819027  0.7818134
#> sample_9      1 0.36386849  0.8765723
#> sample_10     1 0.75911495  0.8025851
#> sample_11     1 0.37125434  0.8765723
#> sample_12     2 0.50961344  0.7600291
#> sample_13     1 0.00000000  0.8702167
#> sample_14     1 0.32027306  0.8702167
#> sample_15     1 0.00000000  0.8702167
#> sample_16     1 0.30705513  0.8765723
#> sample_17     1 0.79470698  0.7466384
#> sample_18     1 0.57239199  0.8638746
#> sample_19     1 0.30705513  0.8765723
#> sample_20     1 0.00000000  0.8702167
#> sample_21     1 0.00000000  0.8702167
#> sample_22     1 0.84467027  0.7818134
#> sample_23     1 0.83786840  0.7818134
#> sample_24     1 0.01857244  0.8702167
#> sample_25     1 0.80732557  0.8171256
#> sample_26     1 0.42671215  0.8765723
#> sample_27     1 0.87926203  0.7266124
#> sample_34     2 0.12155294  0.9352635
#> sample_35     2 0.24471685  0.9352635
#> sample_36     2 0.03657044  0.9352635
#> sample_37     2 0.02296957  0.9352635
#> sample_38     2 0.13065661  0.9352635
#> sample_28     2 0.13592062  0.9352635
#> sample_29     2 0.75163297  0.7459714
#> sample_30     2 0.05185848  0.9352635
#> sample_31     2 0.12484085  0.9352635
#> sample_32     2 0.12435718  0.9352635
#> sample_33     2 0.02296957  0.9352635