Get consensus matrix

Get consensus matrix

# S4 method for ConsensusPartition
get_consensus(object, k)

Arguments

object
A ConsensusPartition-class object.

k
Number of partitions.

Details

For row i and column j in the consensus matrix, the value of corresponding x_ij is the probability of sample i and sample j being in the same group from all partitions.

Value

A consensus matrix corresponding to the current k.

Examples

data(cola_rl)
obj = cola_rl["sd", "kmeans"]
get_consensus(obj, k = 2)

#>            [,1]      [,2]      [,3]       [,4]      [,5]      [,6]       [,7]
#>  [1,] 1.0000000 1.0000000 1.0000000 0.92000002 0.9933333 0.9800000 0.93333334
#>  [2,] 1.0000000 1.0000000 1.0000000 0.92000002 0.9933333 0.9800000 0.93333334
#>  [3,] 1.0000000 1.0000000 1.0000000 0.92000002 0.9933333 0.9800000 0.93333334
#>  [4,] 0.9200000 0.9200000 0.9200000 1.00000000 0.9133334 0.9000000 0.98666668
#>  [5,] 0.9933333 0.9933333 0.9933333 0.91333336 1.0000000 0.9733334 0.92666668
#>  [6,] 0.9800000 0.9800000 0.9800000 0.89999998 0.9733334 1.0000000 0.91333336
#>  [7,] 0.9333333 0.9333333 0.9333333 0.98666668 0.9266667 0.9133334 1.00000000
#>  [8,] 1.0000000 1.0000000 1.0000000 0.92000002 0.9933333 0.9800000 0.93333334
#>  [9,] 1.0000000 1.0000000 1.0000000 0.92000002 0.9933333 0.9800000 0.93333334
#> [10,] 1.0000000 1.0000000 1.0000000 0.92000002 0.9933333 0.9800000 0.93333334
#> [11,] 0.9200000 0.9200000 0.9200000 1.00000000 0.9133334 0.9000000 0.98666668
#> [12,] 0.9733334 0.9733334 0.9733334 0.93333334 0.9666666 0.9533333 0.94666666
#> [13,] 1.0000000 1.0000000 1.0000000 0.92000002 0.9933333 0.9800000 0.93333334
#> [14,] 1.0000000 1.0000000 1.0000000 0.92000002 0.9933333 0.9800000 0.93333334
#> [15,] 1.0000000 1.0000000 1.0000000 0.92000002 0.9933333 0.9800000 0.93333334
#> [16,] 1.0000000 1.0000000 1.0000000 0.92000002 0.9933333 0.9800000 0.93333334
#> [17,] 0.9933333 0.9933333 0.9933333 0.92666668 0.9866667 0.9733334 0.94000000
#> [18,] 0.9866667 0.9866667 0.9866667 0.93333334 0.9800000 0.9666666 0.94666666
#> [19,] 1.0000000 1.0000000 1.0000000 0.92000002 0.9933333 0.9800000 0.93333334
#> [20,] 1.0000000 1.0000000 1.0000000 0.92000002 0.9933333 0.9800000 0.93333334
#> [21,] 0.1666667 0.1666667 0.1666667 0.08666667 0.1733333 0.1466667 0.10000000
#> [22,] 0.1666667 0.1666667 0.1666667 0.08666667 0.1733333 0.1466667 0.10000000
#> [23,] 0.1666667 0.1666667 0.1666667 0.10000000 0.1733333 0.1466667 0.10000000
#> [24,] 0.1666667 0.1666667 0.1666667 0.08666667 0.1733333 0.1466667 0.10000000
#> [25,] 0.1600000 0.1600000 0.1600000 0.09333333 0.1666667 0.1400000 0.09333333
#> [26,] 0.1600000 0.1600000 0.1600000 0.09333333 0.1666667 0.1400000 0.09333333
#> [27,] 0.1733333 0.1733333 0.1733333 0.09333333 0.1800000 0.1533333 0.10666667
#> [28,] 0.1600000 0.1600000 0.1600000 0.09333333 0.1666667 0.1400000 0.09333333
#> [29,] 0.1666667 0.1666667 0.1666667 0.08666667 0.1733333 0.1466667 0.10000000
#> [30,] 0.1666667 0.1666667 0.1666667 0.08666667 0.1733333 0.1466667 0.10000000
#> [31,] 0.1066667 0.1066667 0.1066667 0.04000000 0.1133333 0.1133333 0.04000000
#> [32,] 0.1533333 0.1533333 0.1533333 0.08666667 0.1600000 0.1333333 0.08666667
#> [33,] 0.1666667 0.1666667 0.1666667 0.08666667 0.1733333 0.1466667 0.10000000
#> [34,] 0.1600000 0.1600000 0.1600000 0.09333333 0.1666667 0.1400000 0.09333333
#> [35,] 0.1600000 0.1600000 0.1600000 0.09333333 0.1666667 0.1400000 0.09333333
#> [36,] 0.1533333 0.1533333 0.1533333 0.08666667 0.1600000 0.1333333 0.08666667
#> [37,] 0.1466667 0.1466667 0.1466667 0.08000000 0.1533333 0.1400000 0.08000000
#> [38,] 0.1666667 0.1666667 0.1666667 0.08666667 0.1733333 0.1466667 0.10000000
#> [39,] 0.1533333 0.1533333 0.1533333 0.08666667 0.1600000 0.1333333 0.08666667
#> [40,] 0.1533333 0.1533333 0.1533333 0.08666667 0.1600000 0.1333333 0.08666667
#> [41,] 0.6133333 0.6133333 0.6133333 0.69333333 0.6066667 0.6333333 0.68000001
#> [42,] 0.6066667 0.6066667 0.6066667 0.68666667 0.6000000 0.6266667 0.67333335
#> [43,] 0.6066667 0.6066667 0.6066667 0.68666667 0.6000000 0.6266667 0.67333335
#> [44,] 0.6266667 0.6266667 0.6266667 0.70666665 0.6200000 0.6466666 0.69333333
#> [45,] 0.6200000 0.6200000 0.6200000 0.69999999 0.6133333 0.6400000 0.68666667
#> [46,] 0.6266667 0.6266667 0.6266667 0.70666665 0.6200000 0.6466666 0.69333333
#> [47,] 0.6066667 0.6066667 0.6066667 0.68666667 0.6000000 0.6266667 0.67333335
#> [48,] 0.5733333 0.5733333 0.5733333 0.65333331 0.5666667 0.5933333 0.63999999
#> [49,] 0.6066667 0.6066667 0.6066667 0.68666667 0.6000000 0.6266667 0.67333335
#> [50,] 0.6133333 0.6133333 0.6133333 0.69333333 0.6066667 0.6333333 0.68000001
#> [51,] 0.6066667 0.6066667 0.6066667 0.68666667 0.6000000 0.6266667 0.67333335
#> [52,] 0.7000000 0.7000000 0.7000000 0.77999997 0.6933333 0.7200000 0.76666665
#> [53,] 0.6333333 0.6333333 0.6333333 0.71333331 0.6266667 0.6533333 0.69999999
#> [54,] 0.4133333 0.4133333 0.4133333 0.49333334 0.4200000 0.4333333 0.47999999
#> [55,] 0.4866667 0.4866667 0.4866667 0.56666666 0.4800000 0.5066667 0.55333334
#> [56,] 0.6200000 0.6200000 0.6200000 0.69999999 0.6133333 0.6400000 0.68666667
#> [57,] 0.6200000 0.6200000 0.6200000 0.69999999 0.6133333 0.6400000 0.68666667
#> [58,] 0.6200000 0.6200000 0.6200000 0.69999999 0.6133333 0.6400000 0.68666667
#> [59,] 0.6133333 0.6133333 0.6133333 0.69333333 0.6066667 0.6333333 0.68000001
#> [60,] 0.6066667 0.6066667 0.6066667 0.68666667 0.6000000 0.6266667 0.67333335
#>            [,8]      [,9]     [,10]      [,11]      [,12]     [,13]     [,14]
#>  [1,] 1.0000000 1.0000000 1.0000000 0.92000002 0.97333336 1.0000000 1.0000000
#>  [2,] 1.0000000 1.0000000 1.0000000 0.92000002 0.97333336 1.0000000 1.0000000
#>  [3,] 1.0000000 1.0000000 1.0000000 0.92000002 0.97333336 1.0000000 1.0000000
#>  [4,] 0.9200000 0.9200000 0.9200000 1.00000000 0.93333334 0.9200000 0.9200000
#>  [5,] 0.9933333 0.9933333 0.9933333 0.91333336 0.96666664 0.9933333 0.9933333
#>  [6,] 0.9800000 0.9800000 0.9800000 0.89999998 0.95333332 0.9800000 0.9800000
#>  [7,] 0.9333333 0.9333333 0.9333333 0.98666668 0.94666666 0.9333333 0.9333333
#>  [8,] 1.0000000 1.0000000 1.0000000 0.92000002 0.97333336 1.0000000 1.0000000
#>  [9,] 1.0000000 1.0000000 1.0000000 0.92000002 0.97333336 1.0000000 1.0000000
#> [10,] 1.0000000 1.0000000 1.0000000 0.92000002 0.97333336 1.0000000 1.0000000
#> [11,] 0.9200000 0.9200000 0.9200000 1.00000000 0.93333334 0.9200000 0.9200000
#> [12,] 0.9733334 0.9733334 0.9733334 0.93333334 1.00000000 0.9733334 0.9733334
#> [13,] 1.0000000 1.0000000 1.0000000 0.92000002 0.97333336 1.0000000 1.0000000
#> [14,] 1.0000000 1.0000000 1.0000000 0.92000002 0.97333336 1.0000000 1.0000000
#> [15,] 1.0000000 1.0000000 1.0000000 0.92000002 0.97333336 1.0000000 1.0000000
#> [16,] 1.0000000 1.0000000 1.0000000 0.92000002 0.97333336 1.0000000 1.0000000
#> [17,] 0.9933333 0.9933333 0.9933333 0.92666668 0.96666664 0.9933333 0.9933333
#> [18,] 0.9866667 0.9866667 0.9866667 0.93333334 0.98666668 0.9866667 0.9866667
#> [19,] 1.0000000 1.0000000 1.0000000 0.92000002 0.97333336 1.0000000 1.0000000
#> [20,] 1.0000000 1.0000000 1.0000000 0.92000002 0.97333336 1.0000000 1.0000000
#> [21,] 0.1666667 0.1666667 0.1666667 0.08666667 0.14000000 0.1666667 0.1666667
#> [22,] 0.1666667 0.1666667 0.1666667 0.08666667 0.14000000 0.1666667 0.1666667
#> [23,] 0.1666667 0.1666667 0.1666667 0.10000000 0.14000000 0.1666667 0.1666667
#> [24,] 0.1666667 0.1666667 0.1666667 0.08666667 0.14000000 0.1666667 0.1666667
#> [25,] 0.1600000 0.1600000 0.1600000 0.09333333 0.13333334 0.1600000 0.1600000
#> [26,] 0.1600000 0.1600000 0.1600000 0.09333333 0.13333334 0.1600000 0.1600000
#> [27,] 0.1733333 0.1733333 0.1733333 0.09333333 0.14666666 0.1733333 0.1733333
#> [28,] 0.1600000 0.1600000 0.1600000 0.09333333 0.13333334 0.1600000 0.1600000
#> [29,] 0.1666667 0.1666667 0.1666667 0.08666667 0.14000000 0.1666667 0.1666667
#> [30,] 0.1666667 0.1666667 0.1666667 0.08666667 0.14000000 0.1666667 0.1666667
#> [31,] 0.1066667 0.1066667 0.1066667 0.04000000 0.09333333 0.1066667 0.1066667
#> [32,] 0.1533333 0.1533333 0.1533333 0.08666667 0.12666667 0.1533333 0.1533333
#> [33,] 0.1666667 0.1666667 0.1666667 0.08666667 0.14000000 0.1666667 0.1666667
#> [34,] 0.1600000 0.1600000 0.1600000 0.09333333 0.13333334 0.1600000 0.1600000
#> [35,] 0.1600000 0.1600000 0.1600000 0.09333333 0.13333334 0.1600000 0.1600000
#> [36,] 0.1533333 0.1533333 0.1533333 0.08666667 0.12666667 0.1533333 0.1533333
#> [37,] 0.1466667 0.1466667 0.1466667 0.08000000 0.12000000 0.1466667 0.1466667
#> [38,] 0.1666667 0.1666667 0.1666667 0.08666667 0.14000000 0.1666667 0.1666667
#> [39,] 0.1533333 0.1533333 0.1533333 0.08666667 0.12666667 0.1533333 0.1533333
#> [40,] 0.1533333 0.1533333 0.1533333 0.08666667 0.12666667 0.1533333 0.1533333
#> [41,] 0.6133333 0.6133333 0.6133333 0.69333333 0.63999999 0.6133333 0.6133333
#> [42,] 0.6066667 0.6066667 0.6066667 0.68666667 0.63333333 0.6066667 0.6066667
#> [43,] 0.6066667 0.6066667 0.6066667 0.68666667 0.63333333 0.6066667 0.6066667
#> [44,] 0.6266667 0.6266667 0.6266667 0.70666665 0.65333331 0.6266667 0.6266667
#> [45,] 0.6200000 0.6200000 0.6200000 0.69999999 0.64666665 0.6200000 0.6200000
#> [46,] 0.6266667 0.6266667 0.6266667 0.70666665 0.65333331 0.6266667 0.6266667
#> [47,] 0.6066667 0.6066667 0.6066667 0.68666667 0.63333333 0.6066667 0.6066667
#> [48,] 0.5733333 0.5733333 0.5733333 0.65333331 0.60000002 0.5733333 0.5733333
#> [49,] 0.6066667 0.6066667 0.6066667 0.68666667 0.63333333 0.6066667 0.6066667
#> [50,] 0.6133333 0.6133333 0.6133333 0.69333333 0.63999999 0.6133333 0.6133333
#> [51,] 0.6066667 0.6066667 0.6066667 0.68666667 0.63333333 0.6066667 0.6066667
#> [52,] 0.7000000 0.7000000 0.7000000 0.77999997 0.72666669 0.7000000 0.7000000
#> [53,] 0.6333333 0.6333333 0.6333333 0.71333331 0.66000003 0.6333333 0.6333333
#> [54,] 0.4133333 0.4133333 0.4133333 0.49333334 0.44000000 0.4133333 0.4133333
#> [55,] 0.4866667 0.4866667 0.4866667 0.56666666 0.51333332 0.4866667 0.4866667
#> [56,] 0.6200000 0.6200000 0.6200000 0.69999999 0.64666665 0.6200000 0.6200000
#> [57,] 0.6200000 0.6200000 0.6200000 0.69999999 0.64666665 0.6200000 0.6200000
#> [58,] 0.6200000 0.6200000 0.6200000 0.69999999 0.64666665 0.6200000 0.6200000
#> [59,] 0.6133333 0.6133333 0.6133333 0.69333333 0.63999999 0.6133333 0.6133333
#> [60,] 0.6066667 0.6066667 0.6066667 0.68666667 0.63333333 0.6066667 0.6066667
#>           [,15]     [,16]     [,17]      [,18]     [,19]     [,20]      [,21]
#>  [1,] 1.0000000 1.0000000 0.9933333 0.98666668 1.0000000 1.0000000 0.16666667
#>  [2,] 1.0000000 1.0000000 0.9933333 0.98666668 1.0000000 1.0000000 0.16666667
#>  [3,] 1.0000000 1.0000000 0.9933333 0.98666668 1.0000000 1.0000000 0.16666667
#>  [4,] 0.9200000 0.9200000 0.9266667 0.93333334 0.9200000 0.9200000 0.08666667
#>  [5,] 0.9933333 0.9933333 0.9866667 0.98000002 0.9933333 0.9933333 0.17333333
#>  [6,] 0.9800000 0.9800000 0.9733334 0.96666664 0.9800000 0.9800000 0.14666666
#>  [7,] 0.9333333 0.9333333 0.9400000 0.94666666 0.9333333 0.9333333 0.10000000
#>  [8,] 1.0000000 1.0000000 0.9933333 0.98666668 1.0000000 1.0000000 0.16666667
#>  [9,] 1.0000000 1.0000000 0.9933333 0.98666668 1.0000000 1.0000000 0.16666667
#> [10,] 1.0000000 1.0000000 0.9933333 0.98666668 1.0000000 1.0000000 0.16666667
#> [11,] 0.9200000 0.9200000 0.9266667 0.93333334 0.9200000 0.9200000 0.08666667
#> [12,] 0.9733334 0.9733334 0.9666666 0.98666668 0.9733334 0.9733334 0.14000000
#> [13,] 1.0000000 1.0000000 0.9933333 0.98666668 1.0000000 1.0000000 0.16666667
#> [14,] 1.0000000 1.0000000 0.9933333 0.98666668 1.0000000 1.0000000 0.16666667
#> [15,] 1.0000000 1.0000000 0.9933333 0.98666668 1.0000000 1.0000000 0.16666667
#> [16,] 1.0000000 1.0000000 0.9933333 0.98666668 1.0000000 1.0000000 0.16666667
#> [17,] 0.9933333 0.9933333 1.0000000 0.98000002 0.9933333 0.9933333 0.16000000
#> [18,] 0.9866667 0.9866667 0.9800000 1.00000000 0.9866667 0.9866667 0.15333334
#> [19,] 1.0000000 1.0000000 0.9933333 0.98666668 1.0000000 1.0000000 0.16666667
#> [20,] 1.0000000 1.0000000 0.9933333 0.98666668 1.0000000 1.0000000 0.16666667
#> [21,] 0.1666667 0.1666667 0.1600000 0.15333334 0.1666667 0.1666667 1.00000000
#> [22,] 0.1666667 0.1666667 0.1600000 0.15333334 0.1666667 0.1666667 1.00000000
#> [23,] 0.1666667 0.1666667 0.1600000 0.15333334 0.1666667 0.1666667 0.98666668
#> [24,] 0.1666667 0.1666667 0.1600000 0.15333334 0.1666667 0.1666667 1.00000000
#> [25,] 0.1600000 0.1600000 0.1533333 0.14666666 0.1600000 0.1600000 0.99333334
#> [26,] 0.1600000 0.1600000 0.1533333 0.14666666 0.1600000 0.1600000 0.99333334
#> [27,] 0.1733333 0.1733333 0.1666667 0.16000000 0.1733333 0.1733333 0.99333334
#> [28,] 0.1600000 0.1600000 0.1533333 0.14666666 0.1600000 0.1600000 0.99333334
#> [29,] 0.1666667 0.1666667 0.1600000 0.15333334 0.1666667 0.1666667 1.00000000
#> [30,] 0.1666667 0.1666667 0.1600000 0.15333334 0.1666667 0.1666667 1.00000000
#> [31,] 0.1066667 0.1066667 0.1000000 0.09333333 0.1066667 0.1066667 0.94000000
#> [32,] 0.1533333 0.1533333 0.1466667 0.14000000 0.1533333 0.1533333 0.98666668
#> [33,] 0.1666667 0.1666667 0.1600000 0.15333334 0.1666667 0.1666667 1.00000000
#> [34,] 0.1600000 0.1600000 0.1533333 0.14666666 0.1600000 0.1600000 0.99333334
#> [35,] 0.1600000 0.1600000 0.1533333 0.14666666 0.1600000 0.1600000 0.99333334
#> [36,] 0.1533333 0.1533333 0.1466667 0.14000000 0.1533333 0.1533333 0.98666668
#> [37,] 0.1466667 0.1466667 0.1400000 0.13333334 0.1466667 0.1466667 0.98000002
#> [38,] 0.1666667 0.1666667 0.1600000 0.15333334 0.1666667 0.1666667 1.00000000
#> [39,] 0.1533333 0.1533333 0.1466667 0.14000000 0.1533333 0.1533333 0.98666668
#> [40,] 0.1533333 0.1533333 0.1466667 0.14000000 0.1533333 0.1533333 0.98666668
#> [41,] 0.6133333 0.6133333 0.6200000 0.62666667 0.6133333 0.6133333 0.22000000
#> [42,] 0.6066667 0.6066667 0.6133333 0.62000000 0.6066667 0.6066667 0.22666667
#> [43,] 0.6066667 0.6066667 0.6133333 0.62000000 0.6066667 0.6066667 0.22666667
#> [44,] 0.6266667 0.6266667 0.6333333 0.63999999 0.6266667 0.6266667 0.20666666
#> [45,] 0.6200000 0.6200000 0.6266667 0.63333333 0.6200000 0.6200000 0.21333334
#> [46,] 0.6266667 0.6266667 0.6333333 0.63999999 0.6266667 0.6266667 0.20666666
#> [47,] 0.6066667 0.6066667 0.6133333 0.62000000 0.6066667 0.6066667 0.22666667
#> [48,] 0.5733333 0.5733333 0.5800000 0.58666664 0.5733333 0.5733333 0.25999999
#> [49,] 0.6066667 0.6066667 0.6133333 0.62000000 0.6066667 0.6066667 0.22666667
#> [50,] 0.6133333 0.6133333 0.6200000 0.62666667 0.6133333 0.6133333 0.22000000
#> [51,] 0.6066667 0.6066667 0.6133333 0.62000000 0.6066667 0.6066667 0.22666667
#> [52,] 0.7000000 0.7000000 0.7066666 0.71333331 0.7000000 0.7000000 0.13333334
#> [53,] 0.6333333 0.6333333 0.6400000 0.64666665 0.6333333 0.6333333 0.20000000
#> [54,] 0.4133333 0.4133333 0.4200000 0.42666668 0.4133333 0.4133333 0.41999999
#> [55,] 0.4866667 0.4866667 0.4933333 0.50000000 0.4866667 0.4866667 0.34666666
#> [56,] 0.6200000 0.6200000 0.6266667 0.63333333 0.6200000 0.6200000 0.21333334
#> [57,] 0.6200000 0.6200000 0.6266667 0.63333333 0.6200000 0.6200000 0.21333334
#> [58,] 0.6200000 0.6200000 0.6266667 0.63333333 0.6200000 0.6200000 0.21333334
#> [59,] 0.6133333 0.6133333 0.6200000 0.62666667 0.6133333 0.6133333 0.22000000
#> [60,] 0.6066667 0.6066667 0.6133333 0.62000000 0.6066667 0.6066667 0.22666667
#>            [,22]     [,23]      [,24]      [,25]      [,26]      [,27]
#>  [1,] 0.16666667 0.1666667 0.16666667 0.16000000 0.16000000 0.17333333
#>  [2,] 0.16666667 0.1666667 0.16666667 0.16000000 0.16000000 0.17333333
#>  [3,] 0.16666667 0.1666667 0.16666667 0.16000000 0.16000000 0.17333333
#>  [4,] 0.08666667 0.1000000 0.08666667 0.09333333 0.09333333 0.09333333
#>  [5,] 0.17333333 0.1733333 0.17333333 0.16666667 0.16666667 0.18000001
#>  [6,] 0.14666666 0.1466667 0.14666666 0.14000000 0.14000000 0.15333334
#>  [7,] 0.10000000 0.1000000 0.10000000 0.09333333 0.09333333 0.10666667
#>  [8,] 0.16666667 0.1666667 0.16666667 0.16000000 0.16000000 0.17333333
#>  [9,] 0.16666667 0.1666667 0.16666667 0.16000000 0.16000000 0.17333333
#> [10,] 0.16666667 0.1666667 0.16666667 0.16000000 0.16000000 0.17333333
#> [11,] 0.08666667 0.1000000 0.08666667 0.09333333 0.09333333 0.09333333
#> [12,] 0.14000000 0.1400000 0.14000000 0.13333334 0.13333334 0.14666666
#> [13,] 0.16666667 0.1666667 0.16666667 0.16000000 0.16000000 0.17333333
#> [14,] 0.16666667 0.1666667 0.16666667 0.16000000 0.16000000 0.17333333
#> [15,] 0.16666667 0.1666667 0.16666667 0.16000000 0.16000000 0.17333333
#> [16,] 0.16666667 0.1666667 0.16666667 0.16000000 0.16000000 0.17333333
#> [17,] 0.16000000 0.1600000 0.16000000 0.15333334 0.15333334 0.16666667
#> [18,] 0.15333334 0.1533333 0.15333334 0.14666666 0.14666666 0.16000000
#> [19,] 0.16666667 0.1666667 0.16666667 0.16000000 0.16000000 0.17333333
#> [20,] 0.16666667 0.1666667 0.16666667 0.16000000 0.16000000 0.17333333
#> [21,] 1.00000000 0.9866667 1.00000000 0.99333334 0.99333334 0.99333334
#> [22,] 1.00000000 0.9866667 1.00000000 0.99333334 0.99333334 0.99333334
#> [23,] 0.98666668 1.0000000 0.98666668 0.99333334 0.99333334 0.98000002
#> [24,] 1.00000000 0.9866667 1.00000000 0.99333334 0.99333334 0.99333334
#> [25,] 0.99333334 0.9933333 0.99333334 1.00000000 1.00000000 0.98666668
#> [26,] 0.99333334 0.9933333 0.99333334 1.00000000 1.00000000 0.98666668
#> [27,] 0.99333334 0.9800000 0.99333334 0.98666668 0.98666668 1.00000000
#> [28,] 0.99333334 0.9933333 0.99333334 1.00000000 1.00000000 0.98666668
#> [29,] 1.00000000 0.9866667 1.00000000 0.99333334 0.99333334 0.99333334
#> [30,] 1.00000000 0.9866667 1.00000000 0.99333334 0.99333334 0.99333334
#> [31,] 0.94000000 0.9400000 0.94000000 0.94666666 0.94666666 0.93333334
#> [32,] 0.98666668 0.9866667 0.98666668 0.99333334 0.99333334 0.98000002
#> [33,] 1.00000000 0.9866667 1.00000000 0.99333334 0.99333334 0.99333334
#> [34,] 0.99333334 0.9933333 0.99333334 1.00000000 1.00000000 0.98666668
#> [35,] 0.99333334 0.9933333 0.99333334 1.00000000 1.00000000 0.98666668
#> [36,] 0.98666668 0.9866667 0.98666668 0.99333334 0.99333334 0.98000002
#> [37,] 0.98000002 0.9800000 0.98000002 0.98666668 0.98666668 0.97333336
#> [38,] 1.00000000 0.9866667 1.00000000 0.99333334 0.99333334 0.99333334
#> [39,] 0.98666668 0.9866667 0.98666668 0.99333334 0.99333334 0.98000002
#> [40,] 0.98666668 0.9866667 0.98666668 0.99333334 0.99333334 0.98000002
#> [41,] 0.22000000 0.2333333 0.22000000 0.22666667 0.22666667 0.21333334
#> [42,] 0.22666667 0.2400000 0.22666667 0.23333333 0.23333333 0.22000000
#> [43,] 0.22666667 0.2400000 0.22666667 0.23333333 0.23333333 0.22000000
#> [44,] 0.20666666 0.2200000 0.20666666 0.21333334 0.21333334 0.20000000
#> [45,] 0.21333334 0.2266667 0.21333334 0.22000000 0.22000000 0.20666666
#> [46,] 0.20666666 0.2200000 0.20666666 0.21333334 0.21333334 0.20000000
#> [47,] 0.22666667 0.2400000 0.22666667 0.23333333 0.23333333 0.22000000
#> [48,] 0.25999999 0.2733333 0.25999999 0.26666668 0.26666668 0.25333333
#> [49,] 0.22666667 0.2400000 0.22666667 0.23333333 0.23333333 0.22000000
#> [50,] 0.22000000 0.2333333 0.22000000 0.22666667 0.22666667 0.21333334
#> [51,] 0.22666667 0.2400000 0.22666667 0.23333333 0.23333333 0.22000000
#> [52,] 0.13333334 0.1466667 0.13333334 0.14000000 0.14000000 0.12666667
#> [53,] 0.20000000 0.2133333 0.20000000 0.20666666 0.20666666 0.19333333
#> [54,] 0.41999999 0.4333333 0.41999999 0.42666668 0.42666668 0.41333333
#> [55,] 0.34666666 0.3600000 0.34666666 0.35333332 0.35333332 0.34000000
#> [56,] 0.21333334 0.2266667 0.21333334 0.22000000 0.22000000 0.20666666
#> [57,] 0.21333334 0.2266667 0.21333334 0.22000000 0.22000000 0.20666666
#> [58,] 0.21333334 0.2266667 0.21333334 0.22000000 0.22000000 0.20666666
#> [59,] 0.22000000 0.2333333 0.22000000 0.22666667 0.22666667 0.21333334
#> [60,] 0.22666667 0.2400000 0.22666667 0.23333333 0.23333333 0.22000000
#>            [,28]      [,29]      [,30]      [,31]      [,32]      [,33]
#>  [1,] 0.16000000 0.16666667 0.16666667 0.10666667 0.15333334 0.16666667
#>  [2,] 0.16000000 0.16666667 0.16666667 0.10666667 0.15333334 0.16666667
#>  [3,] 0.16000000 0.16666667 0.16666667 0.10666667 0.15333334 0.16666667
#>  [4,] 0.09333333 0.08666667 0.08666667 0.04000000 0.08666667 0.08666667
#>  [5,] 0.16666667 0.17333333 0.17333333 0.11333334 0.16000000 0.17333333
#>  [6,] 0.14000000 0.14666666 0.14666666 0.11333334 0.13333334 0.14666666
#>  [7,] 0.09333333 0.10000000 0.10000000 0.04000000 0.08666667 0.10000000
#>  [8,] 0.16000000 0.16666667 0.16666667 0.10666667 0.15333334 0.16666667
#>  [9,] 0.16000000 0.16666667 0.16666667 0.10666667 0.15333334 0.16666667
#> [10,] 0.16000000 0.16666667 0.16666667 0.10666667 0.15333334 0.16666667
#> [11,] 0.09333333 0.08666667 0.08666667 0.04000000 0.08666667 0.08666667
#> [12,] 0.13333334 0.14000000 0.14000000 0.09333333 0.12666667 0.14000000
#> [13,] 0.16000000 0.16666667 0.16666667 0.10666667 0.15333334 0.16666667
#> [14,] 0.16000000 0.16666667 0.16666667 0.10666667 0.15333334 0.16666667
#> [15,] 0.16000000 0.16666667 0.16666667 0.10666667 0.15333334 0.16666667
#> [16,] 0.16000000 0.16666667 0.16666667 0.10666667 0.15333334 0.16666667
#> [17,] 0.15333334 0.16000000 0.16000000 0.10000000 0.14666666 0.16000000
#> [18,] 0.14666666 0.15333334 0.15333334 0.09333333 0.14000000 0.15333334
#> [19,] 0.16000000 0.16666667 0.16666667 0.10666667 0.15333334 0.16666667
#> [20,] 0.16000000 0.16666667 0.16666667 0.10666667 0.15333334 0.16666667
#> [21,] 0.99333334 1.00000000 1.00000000 0.94000000 0.98666668 1.00000000
#> [22,] 0.99333334 1.00000000 1.00000000 0.94000000 0.98666668 1.00000000
#> [23,] 0.99333334 0.98666668 0.98666668 0.94000000 0.98666668 0.98666668
#> [24,] 0.99333334 1.00000000 1.00000000 0.94000000 0.98666668 1.00000000
#> [25,] 1.00000000 0.99333334 0.99333334 0.94666666 0.99333334 0.99333334
#> [26,] 1.00000000 0.99333334 0.99333334 0.94666666 0.99333334 0.99333334
#> [27,] 0.98666668 0.99333334 0.99333334 0.93333334 0.98000002 0.99333334
#> [28,] 1.00000000 0.99333334 0.99333334 0.94666666 0.99333334 0.99333334
#> [29,] 0.99333334 1.00000000 1.00000000 0.94000000 0.98666668 1.00000000
#> [30,] 0.99333334 1.00000000 1.00000000 0.94000000 0.98666668 1.00000000
#> [31,] 0.94666666 0.94000000 0.94000000 1.00000000 0.95333332 0.94000000
#> [32,] 0.99333334 0.98666668 0.98666668 0.95333332 1.00000000 0.98666668
#> [33,] 0.99333334 1.00000000 1.00000000 0.94000000 0.98666668 1.00000000
#> [34,] 1.00000000 0.99333334 0.99333334 0.94666666 0.99333334 0.99333334
#> [35,] 1.00000000 0.99333334 0.99333334 0.94666666 0.99333334 0.99333334
#> [36,] 0.99333334 0.98666668 0.98666668 0.95333332 1.00000000 0.98666668
#> [37,] 0.98666668 0.98000002 0.98000002 0.95999998 0.99333334 0.98000002
#> [38,] 0.99333334 1.00000000 1.00000000 0.94000000 0.98666668 1.00000000
#> [39,] 0.99333334 0.98666668 0.98666668 0.95333332 1.00000000 0.98666668
#> [40,] 0.99333334 0.98666668 0.98666668 0.95333332 1.00000000 0.98666668
#> [41,] 0.22666667 0.22000000 0.22000000 0.28000000 0.23333333 0.22000000
#> [42,] 0.23333333 0.22666667 0.22666667 0.28666666 0.23999999 0.22666667
#> [43,] 0.23333333 0.22666667 0.22666667 0.28666666 0.23999999 0.22666667
#> [44,] 0.21333334 0.20666666 0.20666666 0.26666668 0.22000000 0.20666666
#> [45,] 0.22000000 0.21333334 0.21333334 0.27333334 0.22666667 0.21333334
#> [46,] 0.21333334 0.20666666 0.20666666 0.26666668 0.22000000 0.20666666
#> [47,] 0.23333333 0.22666667 0.22666667 0.28666666 0.23999999 0.22666667
#> [48,] 0.26666668 0.25999999 0.25999999 0.31999999 0.27333334 0.25999999
#> [49,] 0.23333333 0.22666667 0.22666667 0.28666666 0.23999999 0.22666667
#> [50,] 0.22666667 0.22000000 0.22000000 0.28000000 0.23333333 0.22000000
#> [51,] 0.23333333 0.22666667 0.22666667 0.28666666 0.23999999 0.22666667
#> [52,] 0.14000000 0.13333334 0.13333334 0.19333333 0.14666666 0.13333334
#> [53,] 0.20666666 0.20000000 0.20000000 0.25999999 0.21333334 0.20000000
#> [54,] 0.42666668 0.41999999 0.41999999 0.47999999 0.43333334 0.41999999
#> [55,] 0.35333332 0.34666666 0.34666666 0.40666667 0.36000001 0.34666666
#> [56,] 0.22000000 0.21333334 0.21333334 0.27333334 0.22666667 0.21333334
#> [57,] 0.22000000 0.21333334 0.21333334 0.27333334 0.22666667 0.21333334
#> [58,] 0.22000000 0.21333334 0.21333334 0.27333334 0.22666667 0.21333334
#> [59,] 0.22666667 0.22000000 0.22000000 0.28000000 0.23333333 0.22000000
#> [60,] 0.23333333 0.22666667 0.22666667 0.28666666 0.23999999 0.22666667
#>            [,34]      [,35]      [,36]     [,37]      [,38]      [,39]
#>  [1,] 0.16000000 0.16000000 0.15333334 0.1466667 0.16666667 0.15333334
#>  [2,] 0.16000000 0.16000000 0.15333334 0.1466667 0.16666667 0.15333334
#>  [3,] 0.16000000 0.16000000 0.15333334 0.1466667 0.16666667 0.15333334
#>  [4,] 0.09333333 0.09333333 0.08666667 0.0800000 0.08666667 0.08666667
#>  [5,] 0.16666667 0.16666667 0.16000000 0.1533333 0.17333333 0.16000000
#>  [6,] 0.14000000 0.14000000 0.13333334 0.1400000 0.14666666 0.13333334
#>  [7,] 0.09333333 0.09333333 0.08666667 0.0800000 0.10000000 0.08666667
#>  [8,] 0.16000000 0.16000000 0.15333334 0.1466667 0.16666667 0.15333334
#>  [9,] 0.16000000 0.16000000 0.15333334 0.1466667 0.16666667 0.15333334
#> [10,] 0.16000000 0.16000000 0.15333334 0.1466667 0.16666667 0.15333334
#> [11,] 0.09333333 0.09333333 0.08666667 0.0800000 0.08666667 0.08666667
#> [12,] 0.13333334 0.13333334 0.12666667 0.1200000 0.14000000 0.12666667
#> [13,] 0.16000000 0.16000000 0.15333334 0.1466667 0.16666667 0.15333334
#> [14,] 0.16000000 0.16000000 0.15333334 0.1466667 0.16666667 0.15333334
#> [15,] 0.16000000 0.16000000 0.15333334 0.1466667 0.16666667 0.15333334
#> [16,] 0.16000000 0.16000000 0.15333334 0.1466667 0.16666667 0.15333334
#> [17,] 0.15333334 0.15333334 0.14666666 0.1400000 0.16000000 0.14666666
#> [18,] 0.14666666 0.14666666 0.14000000 0.1333333 0.15333334 0.14000000
#> [19,] 0.16000000 0.16000000 0.15333334 0.1466667 0.16666667 0.15333334
#> [20,] 0.16000000 0.16000000 0.15333334 0.1466667 0.16666667 0.15333334
#> [21,] 0.99333334 0.99333334 0.98666668 0.9800000 1.00000000 0.98666668
#> [22,] 0.99333334 0.99333334 0.98666668 0.9800000 1.00000000 0.98666668
#> [23,] 0.99333334 0.99333334 0.98666668 0.9800000 0.98666668 0.98666668
#> [24,] 0.99333334 0.99333334 0.98666668 0.9800000 1.00000000 0.98666668
#> [25,] 1.00000000 1.00000000 0.99333334 0.9866667 0.99333334 0.99333334
#> [26,] 1.00000000 1.00000000 0.99333334 0.9866667 0.99333334 0.99333334
#> [27,] 0.98666668 0.98666668 0.98000002 0.9733334 0.99333334 0.98000002
#> [28,] 1.00000000 1.00000000 0.99333334 0.9866667 0.99333334 0.99333334
#> [29,] 0.99333334 0.99333334 0.98666668 0.9800000 1.00000000 0.98666668
#> [30,] 0.99333334 0.99333334 0.98666668 0.9800000 1.00000000 0.98666668
#> [31,] 0.94666666 0.94666666 0.95333332 0.9600000 0.94000000 0.95333332
#> [32,] 0.99333334 0.99333334 1.00000000 0.9933333 0.98666668 1.00000000
#> [33,] 0.99333334 0.99333334 0.98666668 0.9800000 1.00000000 0.98666668
#> [34,] 1.00000000 1.00000000 0.99333334 0.9866667 0.99333334 0.99333334
#> [35,] 1.00000000 1.00000000 0.99333334 0.9866667 0.99333334 0.99333334
#> [36,] 0.99333334 0.99333334 1.00000000 0.9933333 0.98666668 1.00000000
#> [37,] 0.98666668 0.98666668 0.99333334 1.0000000 0.98000002 0.99333334
#> [38,] 0.99333334 0.99333334 0.98666668 0.9800000 1.00000000 0.98666668
#> [39,] 0.99333334 0.99333334 1.00000000 0.9933333 0.98666668 1.00000000
#> [40,] 0.99333334 0.99333334 1.00000000 0.9933333 0.98666668 1.00000000
#> [41,] 0.22666667 0.22666667 0.23333333 0.2400000 0.22000000 0.23333333
#> [42,] 0.23333333 0.23333333 0.23999999 0.2466667 0.22666667 0.23999999
#> [43,] 0.23333333 0.23333333 0.23999999 0.2466667 0.22666667 0.23999999
#> [44,] 0.21333334 0.21333334 0.22000000 0.2266667 0.20666666 0.22000000
#> [45,] 0.22000000 0.22000000 0.22666667 0.2333333 0.21333334 0.22666667
#> [46,] 0.21333334 0.21333334 0.22000000 0.2266667 0.20666666 0.22000000
#> [47,] 0.23333333 0.23333333 0.23999999 0.2466667 0.22666667 0.23999999
#> [48,] 0.26666668 0.26666668 0.27333334 0.2800000 0.25999999 0.27333334
#> [49,] 0.23333333 0.23333333 0.23999999 0.2466667 0.22666667 0.23999999
#> [50,] 0.22666667 0.22666667 0.23333333 0.2400000 0.22000000 0.23333333
#> [51,] 0.23333333 0.23333333 0.23999999 0.2466667 0.22666667 0.23999999
#> [52,] 0.14000000 0.14000000 0.14666666 0.1533333 0.13333334 0.14666666
#> [53,] 0.20666666 0.20666666 0.21333334 0.2200000 0.20000000 0.21333334
#> [54,] 0.42666668 0.42666668 0.43333334 0.4400000 0.41999999 0.43333334
#> [55,] 0.35333332 0.35333332 0.36000001 0.3666667 0.34666666 0.36000001
#> [56,] 0.22000000 0.22000000 0.22666667 0.2333333 0.21333334 0.22666667
#> [57,] 0.22000000 0.22000000 0.22666667 0.2333333 0.21333334 0.22666667
#> [58,] 0.22000000 0.22000000 0.22666667 0.2333333 0.21333334 0.22666667
#> [59,] 0.22666667 0.22666667 0.23333333 0.2400000 0.22000000 0.23333333
#> [60,] 0.23333333 0.23333333 0.23999999 0.2466667 0.22666667 0.23999999
#>            [,40]     [,41]     [,42]     [,43]     [,44]     [,45]     [,46]
#>  [1,] 0.15333334 0.6133333 0.6066667 0.6066667 0.6266667 0.6200000 0.6266667
#>  [2,] 0.15333334 0.6133333 0.6066667 0.6066667 0.6266667 0.6200000 0.6266667
#>  [3,] 0.15333334 0.6133333 0.6066667 0.6066667 0.6266667 0.6200000 0.6266667
#>  [4,] 0.08666667 0.6933333 0.6866667 0.6866667 0.7066666 0.7000000 0.7066666
#>  [5,] 0.16000000 0.6066667 0.6000000 0.6000000 0.6200000 0.6133333 0.6200000
#>  [6,] 0.13333334 0.6333333 0.6266667 0.6266667 0.6466666 0.6400000 0.6466666
#>  [7,] 0.08666667 0.6800000 0.6733333 0.6733333 0.6933333 0.6866667 0.6933333
#>  [8,] 0.15333334 0.6133333 0.6066667 0.6066667 0.6266667 0.6200000 0.6266667
#>  [9,] 0.15333334 0.6133333 0.6066667 0.6066667 0.6266667 0.6200000 0.6266667
#> [10,] 0.15333334 0.6133333 0.6066667 0.6066667 0.6266667 0.6200000 0.6266667
#> [11,] 0.08666667 0.6933333 0.6866667 0.6866667 0.7066666 0.7000000 0.7066666
#> [12,] 0.12666667 0.6400000 0.6333333 0.6333333 0.6533333 0.6466666 0.6533333
#> [13,] 0.15333334 0.6133333 0.6066667 0.6066667 0.6266667 0.6200000 0.6266667
#> [14,] 0.15333334 0.6133333 0.6066667 0.6066667 0.6266667 0.6200000 0.6266667
#> [15,] 0.15333334 0.6133333 0.6066667 0.6066667 0.6266667 0.6200000 0.6266667
#> [16,] 0.15333334 0.6133333 0.6066667 0.6066667 0.6266667 0.6200000 0.6266667
#> [17,] 0.14666666 0.6200000 0.6133333 0.6133333 0.6333333 0.6266667 0.6333333
#> [18,] 0.14000000 0.6266667 0.6200000 0.6200000 0.6400000 0.6333333 0.6400000
#> [19,] 0.15333334 0.6133333 0.6066667 0.6066667 0.6266667 0.6200000 0.6266667
#> [20,] 0.15333334 0.6133333 0.6066667 0.6066667 0.6266667 0.6200000 0.6266667
#> [21,] 0.98666668 0.2200000 0.2266667 0.2266667 0.2066667 0.2133333 0.2066667
#> [22,] 0.98666668 0.2200000 0.2266667 0.2266667 0.2066667 0.2133333 0.2066667
#> [23,] 0.98666668 0.2333333 0.2400000 0.2400000 0.2200000 0.2266667 0.2200000
#> [24,] 0.98666668 0.2200000 0.2266667 0.2266667 0.2066667 0.2133333 0.2066667
#> [25,] 0.99333334 0.2266667 0.2333333 0.2333333 0.2133333 0.2200000 0.2133333
#> [26,] 0.99333334 0.2266667 0.2333333 0.2333333 0.2133333 0.2200000 0.2133333
#> [27,] 0.98000002 0.2133333 0.2200000 0.2200000 0.2000000 0.2066667 0.2000000
#> [28,] 0.99333334 0.2266667 0.2333333 0.2333333 0.2133333 0.2200000 0.2133333
#> [29,] 0.98666668 0.2200000 0.2266667 0.2266667 0.2066667 0.2133333 0.2066667
#> [30,] 0.98666668 0.2200000 0.2266667 0.2266667 0.2066667 0.2133333 0.2066667
#> [31,] 0.95333332 0.2800000 0.2866667 0.2866667 0.2666667 0.2733333 0.2666667
#> [32,] 1.00000000 0.2333333 0.2400000 0.2400000 0.2200000 0.2266667 0.2200000
#> [33,] 0.98666668 0.2200000 0.2266667 0.2266667 0.2066667 0.2133333 0.2066667
#> [34,] 0.99333334 0.2266667 0.2333333 0.2333333 0.2133333 0.2200000 0.2133333
#> [35,] 0.99333334 0.2266667 0.2333333 0.2333333 0.2133333 0.2200000 0.2133333
#> [36,] 1.00000000 0.2333333 0.2400000 0.2400000 0.2200000 0.2266667 0.2200000
#> [37,] 0.99333334 0.2400000 0.2466667 0.2466667 0.2266667 0.2333333 0.2266667
#> [38,] 0.98666668 0.2200000 0.2266667 0.2266667 0.2066667 0.2133333 0.2066667
#> [39,] 1.00000000 0.2333333 0.2400000 0.2400000 0.2200000 0.2266667 0.2200000
#> [40,] 1.00000000 0.2333333 0.2400000 0.2400000 0.2200000 0.2266667 0.2200000
#> [41,] 0.23333333 1.0000000 0.9933333 0.9933333 0.9866667 0.9933333 0.9866667
#> [42,] 0.23999999 0.9933333 1.0000000 1.0000000 0.9800000 0.9866667 0.9800000
#> [43,] 0.23999999 0.9933333 1.0000000 1.0000000 0.9800000 0.9866667 0.9800000
#> [44,] 0.22000000 0.9866667 0.9800000 0.9800000 1.0000000 0.9800000 1.0000000
#> [45,] 0.22666667 0.9933333 0.9866667 0.9866667 0.9800000 1.0000000 0.9800000
#> [46,] 0.22000000 0.9866667 0.9800000 0.9800000 1.0000000 0.9800000 1.0000000
#> [47,] 0.23999999 0.9933333 1.0000000 1.0000000 0.9800000 0.9866667 0.9800000
#> [48,] 0.27333334 0.9600000 0.9666666 0.9666666 0.9466667 0.9533333 0.9466667
#> [49,] 0.23999999 0.9933333 1.0000000 1.0000000 0.9800000 0.9866667 0.9800000
#> [50,] 0.23333333 1.0000000 0.9933333 0.9933333 0.9866667 0.9933333 0.9866667
#> [51,] 0.23999999 0.9933333 1.0000000 1.0000000 0.9800000 0.9866667 0.9800000
#> [52,] 0.14666666 0.9133334 0.9066667 0.9066667 0.9133334 0.9200000 0.9133334
#> [53,] 0.21333334 0.9800000 0.9733334 0.9733334 0.9800000 0.9733334 0.9800000
#> [54,] 0.43333334 0.8000000 0.8066667 0.8066667 0.7866667 0.7933334 0.7866667
#> [55,] 0.36000001 0.8733333 0.8800000 0.8800000 0.8600000 0.8666667 0.8600000
#> [56,] 0.22666667 0.9933333 0.9866667 0.9866667 0.9800000 0.9866667 0.9800000
#> [57,] 0.22666667 0.9933333 0.9866667 0.9866667 0.9933333 0.9866667 0.9933333
#> [58,] 0.22666667 0.9933333 0.9866667 0.9866667 0.9933333 0.9866667 0.9933333
#> [59,] 0.23333333 1.0000000 0.9933333 0.9933333 0.9866667 0.9933333 0.9866667
#> [60,] 0.23999999 0.9933333 1.0000000 1.0000000 0.9800000 0.9866667 0.9800000
#>           [,47]     [,48]     [,49]     [,50]     [,51]     [,52]     [,53]
#>  [1,] 0.6066667 0.5733333 0.6066667 0.6133333 0.6066667 0.7000000 0.6333333
#>  [2,] 0.6066667 0.5733333 0.6066667 0.6133333 0.6066667 0.7000000 0.6333333
#>  [3,] 0.6066667 0.5733333 0.6066667 0.6133333 0.6066667 0.7000000 0.6333333
#>  [4,] 0.6866667 0.6533333 0.6866667 0.6933333 0.6866667 0.7800000 0.7133333
#>  [5,] 0.6000000 0.5666667 0.6000000 0.6066667 0.6000000 0.6933333 0.6266667
#>  [6,] 0.6266667 0.5933333 0.6266667 0.6333333 0.6266667 0.7200000 0.6533333
#>  [7,] 0.6733333 0.6400000 0.6733333 0.6800000 0.6733333 0.7666667 0.7000000
#>  [8,] 0.6066667 0.5733333 0.6066667 0.6133333 0.6066667 0.7000000 0.6333333
#>  [9,] 0.6066667 0.5733333 0.6066667 0.6133333 0.6066667 0.7000000 0.6333333
#> [10,] 0.6066667 0.5733333 0.6066667 0.6133333 0.6066667 0.7000000 0.6333333
#> [11,] 0.6866667 0.6533333 0.6866667 0.6933333 0.6866667 0.7800000 0.7133333
#> [12,] 0.6333333 0.6000000 0.6333333 0.6400000 0.6333333 0.7266667 0.6600000
#> [13,] 0.6066667 0.5733333 0.6066667 0.6133333 0.6066667 0.7000000 0.6333333
#> [14,] 0.6066667 0.5733333 0.6066667 0.6133333 0.6066667 0.7000000 0.6333333
#> [15,] 0.6066667 0.5733333 0.6066667 0.6133333 0.6066667 0.7000000 0.6333333
#> [16,] 0.6066667 0.5733333 0.6066667 0.6133333 0.6066667 0.7000000 0.6333333
#> [17,] 0.6133333 0.5800000 0.6133333 0.6200000 0.6133333 0.7066666 0.6400000
#> [18,] 0.6200000 0.5866666 0.6200000 0.6266667 0.6200000 0.7133333 0.6466666
#> [19,] 0.6066667 0.5733333 0.6066667 0.6133333 0.6066667 0.7000000 0.6333333
#> [20,] 0.6066667 0.5733333 0.6066667 0.6133333 0.6066667 0.7000000 0.6333333
#> [21,] 0.2266667 0.2600000 0.2266667 0.2200000 0.2266667 0.1333333 0.2000000
#> [22,] 0.2266667 0.2600000 0.2266667 0.2200000 0.2266667 0.1333333 0.2000000
#> [23,] 0.2400000 0.2733333 0.2400000 0.2333333 0.2400000 0.1466667 0.2133333
#> [24,] 0.2266667 0.2600000 0.2266667 0.2200000 0.2266667 0.1333333 0.2000000
#> [25,] 0.2333333 0.2666667 0.2333333 0.2266667 0.2333333 0.1400000 0.2066667
#> [26,] 0.2333333 0.2666667 0.2333333 0.2266667 0.2333333 0.1400000 0.2066667
#> [27,] 0.2200000 0.2533333 0.2200000 0.2133333 0.2200000 0.1266667 0.1933333
#> [28,] 0.2333333 0.2666667 0.2333333 0.2266667 0.2333333 0.1400000 0.2066667
#> [29,] 0.2266667 0.2600000 0.2266667 0.2200000 0.2266667 0.1333333 0.2000000
#> [30,] 0.2266667 0.2600000 0.2266667 0.2200000 0.2266667 0.1333333 0.2000000
#> [31,] 0.2866667 0.3200000 0.2866667 0.2800000 0.2866667 0.1933333 0.2600000
#> [32,] 0.2400000 0.2733333 0.2400000 0.2333333 0.2400000 0.1466667 0.2133333
#> [33,] 0.2266667 0.2600000 0.2266667 0.2200000 0.2266667 0.1333333 0.2000000
#> [34,] 0.2333333 0.2666667 0.2333333 0.2266667 0.2333333 0.1400000 0.2066667
#> [35,] 0.2333333 0.2666667 0.2333333 0.2266667 0.2333333 0.1400000 0.2066667
#> [36,] 0.2400000 0.2733333 0.2400000 0.2333333 0.2400000 0.1466667 0.2133333
#> [37,] 0.2466667 0.2800000 0.2466667 0.2400000 0.2466667 0.1533333 0.2200000
#> [38,] 0.2266667 0.2600000 0.2266667 0.2200000 0.2266667 0.1333333 0.2000000
#> [39,] 0.2400000 0.2733333 0.2400000 0.2333333 0.2400000 0.1466667 0.2133333
#> [40,] 0.2400000 0.2733333 0.2400000 0.2333333 0.2400000 0.1466667 0.2133333
#> [41,] 0.9933333 0.9600000 0.9933333 1.0000000 0.9933333 0.9133334 0.9800000
#> [42,] 1.0000000 0.9666666 1.0000000 0.9933333 1.0000000 0.9066667 0.9733334
#> [43,] 1.0000000 0.9666666 1.0000000 0.9933333 1.0000000 0.9066667 0.9733334
#> [44,] 0.9800000 0.9466667 0.9800000 0.9866667 0.9800000 0.9133334 0.9800000
#> [45,] 0.9866667 0.9533333 0.9866667 0.9933333 0.9866667 0.9200000 0.9733334
#> [46,] 0.9800000 0.9466667 0.9800000 0.9866667 0.9800000 0.9133334 0.9800000
#> [47,] 1.0000000 0.9666666 1.0000000 0.9933333 1.0000000 0.9066667 0.9733334
#> [48,] 0.9666666 1.0000000 0.9666666 0.9600000 0.9666666 0.8733333 0.9400000
#> [49,] 1.0000000 0.9666666 1.0000000 0.9933333 1.0000000 0.9066667 0.9733334
#> [50,] 0.9933333 0.9600000 0.9933333 1.0000000 0.9933333 0.9133334 0.9800000
#> [51,] 1.0000000 0.9666666 1.0000000 0.9933333 1.0000000 0.9066667 0.9733334
#> [52,] 0.9066667 0.8733333 0.9066667 0.9133334 0.9066667 1.0000000 0.9333333
#> [53,] 0.9733334 0.9400000 0.9733334 0.9800000 0.9733334 0.9333333 1.0000000
#> [54,] 0.8066667 0.8400000 0.8066667 0.8000000 0.8066667 0.7133333 0.7800000
#> [55,] 0.8800000 0.9000000 0.8800000 0.8733333 0.8800000 0.7866667 0.8533334
#> [56,] 0.9866667 0.9533333 0.9866667 0.9933333 0.9866667 0.9200000 0.9733334
#> [57,] 0.9866667 0.9533333 0.9866667 0.9933333 0.9866667 0.9200000 0.9866667
#> [58,] 0.9866667 0.9533333 0.9866667 0.9933333 0.9866667 0.9200000 0.9866667
#> [59,] 0.9933333 0.9600000 0.9933333 1.0000000 0.9933333 0.9133334 0.9800000
#> [60,] 1.0000000 0.9666666 1.0000000 0.9933333 1.0000000 0.9066667 0.9733334
#>           [,54]     [,55]     [,56]     [,57]     [,58]     [,59]     [,60]
#>  [1,] 0.4133333 0.4866667 0.6200000 0.6200000 0.6200000 0.6133333 0.6066667
#>  [2,] 0.4133333 0.4866667 0.6200000 0.6200000 0.6200000 0.6133333 0.6066667
#>  [3,] 0.4133333 0.4866667 0.6200000 0.6200000 0.6200000 0.6133333 0.6066667
#>  [4,] 0.4933333 0.5666667 0.7000000 0.7000000 0.7000000 0.6933333 0.6866667
#>  [5,] 0.4200000 0.4800000 0.6133333 0.6133333 0.6133333 0.6066667 0.6000000
#>  [6,] 0.4333333 0.5066667 0.6400000 0.6400000 0.6400000 0.6333333 0.6266667
#>  [7,] 0.4800000 0.5533333 0.6866667 0.6866667 0.6866667 0.6800000 0.6733333
#>  [8,] 0.4133333 0.4866667 0.6200000 0.6200000 0.6200000 0.6133333 0.6066667
#>  [9,] 0.4133333 0.4866667 0.6200000 0.6200000 0.6200000 0.6133333 0.6066667
#> [10,] 0.4133333 0.4866667 0.6200000 0.6200000 0.6200000 0.6133333 0.6066667
#> [11,] 0.4933333 0.5666667 0.7000000 0.7000000 0.7000000 0.6933333 0.6866667
#> [12,] 0.4400000 0.5133333 0.6466666 0.6466666 0.6466666 0.6400000 0.6333333
#> [13,] 0.4133333 0.4866667 0.6200000 0.6200000 0.6200000 0.6133333 0.6066667
#> [14,] 0.4133333 0.4866667 0.6200000 0.6200000 0.6200000 0.6133333 0.6066667
#> [15,] 0.4133333 0.4866667 0.6200000 0.6200000 0.6200000 0.6133333 0.6066667
#> [16,] 0.4133333 0.4866667 0.6200000 0.6200000 0.6200000 0.6133333 0.6066667
#> [17,] 0.4200000 0.4933333 0.6266667 0.6266667 0.6266667 0.6200000 0.6133333
#> [18,] 0.4266667 0.5000000 0.6333333 0.6333333 0.6333333 0.6266667 0.6200000
#> [19,] 0.4133333 0.4866667 0.6200000 0.6200000 0.6200000 0.6133333 0.6066667
#> [20,] 0.4133333 0.4866667 0.6200000 0.6200000 0.6200000 0.6133333 0.6066667
#> [21,] 0.4200000 0.3466667 0.2133333 0.2133333 0.2133333 0.2200000 0.2266667
#> [22,] 0.4200000 0.3466667 0.2133333 0.2133333 0.2133333 0.2200000 0.2266667
#> [23,] 0.4333333 0.3600000 0.2266667 0.2266667 0.2266667 0.2333333 0.2400000
#> [24,] 0.4200000 0.3466667 0.2133333 0.2133333 0.2133333 0.2200000 0.2266667
#> [25,] 0.4266667 0.3533333 0.2200000 0.2200000 0.2200000 0.2266667 0.2333333
#> [26,] 0.4266667 0.3533333 0.2200000 0.2200000 0.2200000 0.2266667 0.2333333
#> [27,] 0.4133333 0.3400000 0.2066667 0.2066667 0.2066667 0.2133333 0.2200000
#> [28,] 0.4266667 0.3533333 0.2200000 0.2200000 0.2200000 0.2266667 0.2333333
#> [29,] 0.4200000 0.3466667 0.2133333 0.2133333 0.2133333 0.2200000 0.2266667
#> [30,] 0.4200000 0.3466667 0.2133333 0.2133333 0.2133333 0.2200000 0.2266667
#> [31,] 0.4800000 0.4066667 0.2733333 0.2733333 0.2733333 0.2800000 0.2866667
#> [32,] 0.4333333 0.3600000 0.2266667 0.2266667 0.2266667 0.2333333 0.2400000
#> [33,] 0.4200000 0.3466667 0.2133333 0.2133333 0.2133333 0.2200000 0.2266667
#> [34,] 0.4266667 0.3533333 0.2200000 0.2200000 0.2200000 0.2266667 0.2333333
#> [35,] 0.4266667 0.3533333 0.2200000 0.2200000 0.2200000 0.2266667 0.2333333
#> [36,] 0.4333333 0.3600000 0.2266667 0.2266667 0.2266667 0.2333333 0.2400000
#> [37,] 0.4400000 0.3666667 0.2333333 0.2333333 0.2333333 0.2400000 0.2466667
#> [38,] 0.4200000 0.3466667 0.2133333 0.2133333 0.2133333 0.2200000 0.2266667
#> [39,] 0.4333333 0.3600000 0.2266667 0.2266667 0.2266667 0.2333333 0.2400000
#> [40,] 0.4333333 0.3600000 0.2266667 0.2266667 0.2266667 0.2333333 0.2400000
#> [41,] 0.8000000 0.8733333 0.9933333 0.9933333 0.9933333 1.0000000 0.9933333
#> [42,] 0.8066667 0.8800000 0.9866667 0.9866667 0.9866667 0.9933333 1.0000000
#> [43,] 0.8066667 0.8800000 0.9866667 0.9866667 0.9866667 0.9933333 1.0000000
#> [44,] 0.7866667 0.8600000 0.9800000 0.9933333 0.9933333 0.9866667 0.9800000
#> [45,] 0.7933334 0.8666667 0.9866667 0.9866667 0.9866667 0.9933333 0.9866667
#> [46,] 0.7866667 0.8600000 0.9800000 0.9933333 0.9933333 0.9866667 0.9800000
#> [47,] 0.8066667 0.8800000 0.9866667 0.9866667 0.9866667 0.9933333 1.0000000
#> [48,] 0.8400000 0.9000000 0.9533333 0.9533333 0.9533333 0.9600000 0.9666666
#> [49,] 0.8066667 0.8800000 0.9866667 0.9866667 0.9866667 0.9933333 1.0000000
#> [50,] 0.8000000 0.8733333 0.9933333 0.9933333 0.9933333 1.0000000 0.9933333
#> [51,] 0.8066667 0.8800000 0.9866667 0.9866667 0.9866667 0.9933333 1.0000000
#> [52,] 0.7133333 0.7866667 0.9200000 0.9200000 0.9200000 0.9133334 0.9066667
#> [53,] 0.7800000 0.8533334 0.9733334 0.9866667 0.9866667 0.9800000 0.9733334
#> [54,] 1.0000000 0.7933334 0.7933334 0.7933334 0.7933334 0.8000000 0.8066667
#> [55,] 0.7933334 1.0000000 0.8666667 0.8666667 0.8666667 0.8733333 0.8800000
#> [56,] 0.7933334 0.8666667 1.0000000 0.9866667 0.9866667 0.9933333 0.9866667
#> [57,] 0.7933334 0.8666667 0.9866667 1.0000000 1.0000000 0.9933333 0.9866667
#> [58,] 0.7933334 0.8666667 0.9866667 1.0000000 1.0000000 0.9933333 0.9866667
#> [59,] 0.8000000 0.8733333 0.9933333 0.9933333 0.9933333 1.0000000 0.9933333
#> [60,] 0.8066667 0.8800000 0.9866667 0.9866667 0.9866667 0.9933333 1.0000000