Get membership matrix

Get membership matrix

# S4 method for ConsensusPartition
get_membership(object, k, each = FALSE)

Arguments

object
A ConsensusPartition-class object.

k
Number of subgroups.

each
Whether to return the percentage membership matrix which is summarized from all partitions or the individual membership in every single partition run.

Details

If each == FALSE, the value in the membership matrix is the probability to be in one subgroup, while if each == TRUE, the membership matrix contains the subgroup labels for every single partitions which are from randomly sampling from the original matrix.

The percent membership matrix is calculated by cl_consensus.

Value

If each == FALSE, it returns a membership matrix where rows correspond to the columns from the subgroups.

If each == TRUE, it returns a membership matrix where rows correspond to the columns from the original matrix.

See also

Author

Zuguang Gu <z.gu@dkfz.de>

Examples

data(golub_cola)
obj = golub_cola["ATC", "skmeans"]
get_membership(obj, k = 2)

#>             p1   p2
#> sample_39 1.00 0.00
#> sample_40 1.00 0.00
#> sample_42 0.00 1.00
#> sample_47 0.00 1.00
#> sample_48 0.00 1.00
#> sample_49 1.00 0.00
#> sample_41 0.00 1.00
#> sample_43 0.00 1.00
#> sample_44 0.00 1.00
#> sample_45 0.00 1.00
#> sample_46 0.00 1.00
#> sample_70 0.00 1.00
#> sample_71 0.00 1.00
#> sample_72 0.00 1.00
#> sample_68 0.00 1.00
#> sample_69 0.00 1.00
#> sample_67 0.00 1.00
#> sample_55 1.00 0.00
#> sample_56 1.00 0.00
#> sample_59 0.00 1.00
#> sample_52 1.00 0.00
#> sample_53 1.00 0.00
#> sample_51 1.00 0.00
#> sample_50 1.00 0.00
#> sample_54 0.00 1.00
#> sample_57 1.00 0.00
#> sample_58 1.00 0.00
#> sample_60 0.00 1.00
#> sample_61 1.00 0.00
#> sample_65 1.00 0.00
#> sample_66 0.00 1.00
#> sample_63 1.00 0.00
#> sample_64 1.00 0.00
#> sample_62 1.00 0.00
#> sample_1  1.00 0.00
#> sample_2  0.00 1.00
#> sample_3  1.00 0.00
#> sample_4  0.96 0.04
#> sample_5  0.00 1.00
#> sample_6  1.00 0.00
#> sample_7  1.00 0.00
#> sample_8  1.00 0.00
#> sample_9  0.00 1.00
#> sample_10 0.00 1.00
#> sample_11 0.00 1.00
#> sample_12 1.00 0.00
#> sample_13 0.00 1.00
#> sample_14 0.00 1.00
#> sample_15 0.00 1.00
#> sample_16 0.00 1.00
#> sample_17 0.00 1.00
#> sample_18 0.22 0.78
#> sample_19 0.00 1.00
#> sample_20 0.00 1.00
#> sample_21 0.00 1.00
#> sample_22 1.00 0.00
#> sample_23 1.00 0.00
#> sample_24 0.00 1.00
#> sample_25 1.00 0.00
#> sample_26 0.00 1.00
#> sample_27 1.00 0.00
#> sample_34 1.00 0.00
#> sample_35 1.00 0.00
#> sample_36 1.00 0.00
#> sample_37 1.00 0.00
#> sample_38 1.00 0.00
#> sample_28 1.00 0.00
#> sample_29 0.00 1.00
#> sample_30 1.00 0.00
#> sample_31 1.00 0.00
#> sample_32 1.00 0.00
#> sample_33 1.00 0.00

get_membership(obj, k = 2, each = TRUE)

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#>  [2,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#>  [3,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#>  [4,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#>  [5,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#>  [6,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#>  [7,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#>  [8,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#>  [9,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [10,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [11,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [12,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [13,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [14,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [15,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [16,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [17,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [18,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [19,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [20,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [21,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [22,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [23,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [24,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [25,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [26,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [27,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [28,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [29,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [30,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [31,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [32,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [33,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [34,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [35,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [36,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [37,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [38,]    1    1    1    1    1    2    1    1    1     1     1     1     1
#> [39,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [40,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [41,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [42,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [43,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [44,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [45,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [46,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [47,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [48,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [49,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [50,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [51,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [52,]    2    1    1    2    2    2    2    2    2     2     2     2     2
#> [53,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [54,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [55,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [56,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [57,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [58,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [59,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [60,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [61,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [62,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [63,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [64,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [65,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [66,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [67,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [68,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [69,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [70,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [71,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [72,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [2,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [3,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [4,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [5,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [6,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [7,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [8,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [9,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [10,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [11,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [12,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [13,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [14,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [15,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [16,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [17,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [18,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [19,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [20,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [21,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [22,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [23,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [24,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [25,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [26,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [27,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [28,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [29,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [30,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [31,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [32,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [33,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [34,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [35,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [36,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [37,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [38,]     1     2     1     1     1     1     1     1     1     1     1     1
#> [39,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [40,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [41,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [42,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [43,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [44,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [45,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [46,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [47,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [48,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [49,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [50,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [51,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [52,]     1     1     1     2     2     2     2     1     2     2     2     2
#> [53,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [54,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [55,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [56,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [57,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [58,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [59,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [60,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [61,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [62,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [63,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [64,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [65,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [66,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [67,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [68,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [69,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [70,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [71,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [72,]     1     1     1     1     1     1     1     1     1     1     1     1
#>       [,26] [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36] [,37]
#>  [1,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [2,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [3,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [4,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [5,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [6,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [7,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [8,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [9,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [10,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [11,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [12,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [13,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [14,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [15,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [16,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [17,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [18,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [19,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [20,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [21,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [22,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [23,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [24,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [25,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [26,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [27,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [28,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [29,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [30,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [31,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [32,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [33,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [34,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [35,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [36,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [37,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [38,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [39,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [40,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [41,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [42,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [43,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [44,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [45,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [46,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [47,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [48,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [49,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [50,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [51,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [52,]     2     2     2     1     2     2     2     2     1     2     2     2
#> [53,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [54,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [55,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [56,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [57,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [58,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [59,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [60,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [61,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [62,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [63,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [64,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [65,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [66,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [67,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [68,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [69,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [70,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [71,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [72,]     1     1     1     1     1     1     1     1     1     1     1     1
#>       [,38] [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46] [,47] [,48] [,49]
#>  [1,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [2,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [3,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [4,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [5,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [6,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [7,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [8,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [9,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [10,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [11,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [12,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [13,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [14,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [15,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [16,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [17,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [18,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [19,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [20,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [21,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [22,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [23,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [24,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [25,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [26,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [27,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [28,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [29,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [30,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [31,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [32,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [33,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [34,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [35,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [36,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [37,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [38,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [39,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [40,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [41,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [42,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [43,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [44,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [45,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [46,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [47,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [48,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [49,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [50,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [51,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [52,]     1     2     2     2     2     1     1     2     2     2     2     2
#> [53,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [54,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [55,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [56,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [57,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [58,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [59,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [60,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [61,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [62,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [63,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [64,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [65,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [66,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [67,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [68,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [69,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [70,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [71,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [72,]     1     1     1     1     1     1     1     1     1     1     1     1
#>       [,50]
#>  [1,]     1
#>  [2,]     1
#>  [3,]     2
#>  [4,]     2
#>  [5,]     2
#>  [6,]     1
#>  [7,]     2
#>  [8,]     2
#>  [9,]     2
#> [10,]     2
#> [11,]     2
#> [12,]     2
#> [13,]     2
#> [14,]     2
#> [15,]     2
#> [16,]     2
#> [17,]     2
#> [18,]     1
#> [19,]     1
#> [20,]     2
#> [21,]     1
#> [22,]     1
#> [23,]     1
#> [24,]     1
#> [25,]     2
#> [26,]     1
#> [27,]     1
#> [28,]     2
#> [29,]     1
#> [30,]     1
#> [31,]     2
#> [32,]     1
#> [33,]     1
#> [34,]     1
#> [35,]     1
#> [36,]     2
#> [37,]     1
#> [38,]     1
#> [39,]     2
#> [40,]     1
#> [41,]     1
#> [42,]     1
#> [43,]     2
#> [44,]     2
#> [45,]     2
#> [46,]     1
#> [47,]     2
#> [48,]     2
#> [49,]     2
#> [50,]     2
#> [51,]     2
#> [52,]     2
#> [53,]     2
#> [54,]     2
#> [55,]     2
#> [56,]     1
#> [57,]     1
#> [58,]     2
#> [59,]     1
#> [60,]     2
#> [61,]     1
#> [62,]     1
#> [63,]     1
#> [64,]     1
#> [65,]     1
#> [66,]     1
#> [67,]     1
#> [68,]     2
#> [69,]     1
#> [70,]     1
#> [71,]     1
#> [72,]     1