Get parameters

Get parameters

# S4 method for ConsensusPartition
get_param(object, k = object@k, unique = TRUE)

Arguments

object
A ConsensusPartition-class object.

k
Number of subgroups.

unique
Whether to apply unique to rows of the returned data frame.

Details

It is mainly used internally.

Value

A data frame of parameters corresponding to the current k. In the data frame, each row corresponds to a partition run.

Author

Zuguang Gu <z.gu@dkfz.de>

Examples

data(golub_cola)
obj = golub_cola["ATC", "skmeans"]
get_param(obj)

#>   top_n k n_row n_col
#> 1   412 2   330    72
#> 2   412 3   330    72
#> 3   412 4   330    72
#> 4   412 5   330    72
#> 5   412 6   330    72

get_param(obj, k = 2)

#>   top_n k n_row n_col
#> 1   412 2   330    72

get_param(obj, unique = FALSE)

#>     top_n k n_row n_col
#> 1     412 2   330    72
#> 2     412 2   330    72
#> 3     412 2   330    72
#> 4     412 2   330    72
#> 5     412 2   330    72
#> 6     412 2   330    72
#> 7     412 2   330    72
#> 8     412 2   330    72
#> 9     412 2   330    72
#> 10    412 2   330    72
#> 11    412 2   330    72
#> 12    412 2   330    72
#> 13    412 2   330    72
#> 14    412 2   330    72
#> 15    412 2   330    72
#> 16    412 2   330    72
#> 17    412 2   330    72
#> 18    412 2   330    72
#> 19    412 2   330    72
#> 20    412 2   330    72
#> 21    412 2   330    72
#> 22    412 2   330    72
#> 23    412 2   330    72
#> 24    412 2   330    72
#> 25    412 2   330    72
#> 26    412 2   330    72
#> 27    412 2   330    72
#> 28    412 2   330    72
#> 29    412 2   330    72
#> 30    412 2   330    72
#> 31    412 2   330    72
#> 32    412 2   330    72
#> 33    412 2   330    72
#> 34    412 2   330    72
#> 35    412 2   330    72
#> 36    412 2   330    72
#> 37    412 2   330    72
#> 38    412 2   330    72
#> 39    412 2   330    72
#> 40    412 2   330    72
#> 41    412 2   330    72
#> 42    412 2   330    72
#> 43    412 2   330    72
#> 44    412 2   330    72
#> 45    412 2   330    72
#> 46    412 2   330    72
#> 47    412 2   330    72
#> 48    412 2   330    72
#> 49    412 2   330    72
#> 50    412 2   330    72
#> 51    412 3   330    72
#> 52    412 3   330    72
#> 53    412 3   330    72
#> 54    412 3   330    72
#> 55    412 3   330    72
#> 56    412 3   330    72
#> 57    412 3   330    72
#> 58    412 3   330    72
#> 59    412 3   330    72
#> 60    412 3   330    72
#> 61    412 3   330    72
#> 62    412 3   330    72
#> 63    412 3   330    72
#> 64    412 3   330    72
#> 65    412 3   330    72
#> 66    412 3   330    72
#> 67    412 3   330    72
#> 68    412 3   330    72
#> 69    412 3   330    72
#> 70    412 3   330    72
#> 71    412 3   330    72
#> 72    412 3   330    72
#> 73    412 3   330    72
#> 74    412 3   330    72
#> 75    412 3   330    72
#> 76    412 3   330    72
#> 77    412 3   330    72
#> 78    412 3   330    72
#> 79    412 3   330    72
#> 80    412 3   330    72
#> 81    412 3   330    72
#> 82    412 3   330    72
#> 83    412 3   330    72
#> 84    412 3   330    72
#> 85    412 3   330    72
#> 86    412 3   330    72
#> 87    412 3   330    72
#> 88    412 3   330    72
#> 89    412 3   330    72
#> 90    412 3   330    72
#> 91    412 3   330    72
#> 92    412 3   330    72
#> 93    412 3   330    72
#> 94    412 3   330    72
#> 95    412 3   330    72
#> 96    412 3   330    72
#> 97    412 3   330    72
#> 98    412 3   330    72
#> 99    412 3   330    72
#> 100   412 3   330    72
#> 101   412 4   330    72
#> 102   412 4   330    72
#> 103   412 4   330    72
#> 104   412 4   330    72
#> 105   412 4   330    72
#> 106   412 4   330    72
#> 107   412 4   330    72
#> 108   412 4   330    72
#> 109   412 4   330    72
#> 110   412 4   330    72
#> 111   412 4   330    72
#> 112   412 4   330    72
#> 113   412 4   330    72
#> 114   412 4   330    72
#> 115   412 4   330    72
#> 116   412 4   330    72
#> 117   412 4   330    72
#> 118   412 4   330    72
#> 119   412 4   330    72
#> 120   412 4   330    72
#> 121   412 4   330    72
#> 122   412 4   330    72
#> 123   412 4   330    72
#> 124   412 4   330    72
#> 125   412 4   330    72
#> 126   412 4   330    72
#> 127   412 4   330    72
#> 128   412 4   330    72
#> 129   412 4   330    72
#> 130   412 4   330    72
#> 131   412 4   330    72
#> 132   412 4   330    72
#> 133   412 4   330    72
#> 134   412 4   330    72
#> 135   412 4   330    72
#> 136   412 4   330    72
#> 137   412 4   330    72
#> 138   412 4   330    72
#> 139   412 4   330    72
#> 140   412 4   330    72
#> 141   412 4   330    72
#> 142   412 4   330    72
#> 143   412 4   330    72
#> 144   412 4   330    72
#> 145   412 4   330    72
#> 146   412 4   330    72
#> 147   412 4   330    72
#> 148   412 4   330    72
#> 149   412 4   330    72
#> 150   412 4   330    72
#> 151   412 5   330    72
#> 152   412 5   330    72
#> 153   412 5   330    72
#> 154   412 5   330    72
#> 155   412 5   330    72
#> 156   412 5   330    72
#> 157   412 5   330    72
#> 158   412 5   330    72
#> 159   412 5   330    72
#> 160   412 5   330    72
#> 161   412 5   330    72
#> 162   412 5   330    72
#> 163   412 5   330    72
#> 164   412 5   330    72
#> 165   412 5   330    72
#> 166   412 5   330    72
#> 167   412 5   330    72
#> 168   412 5   330    72
#> 169   412 5   330    72
#> 170   412 5   330    72
#> 171   412 5   330    72
#> 172   412 5   330    72
#> 173   412 5   330    72
#> 174   412 5   330    72
#> 175   412 5   330    72
#> 176   412 5   330    72
#> 177   412 5   330    72
#> 178   412 5   330    72
#> 179   412 5   330    72
#> 180   412 5   330    72
#> 181   412 5   330    72
#> 182   412 5   330    72
#> 183   412 5   330    72
#> 184   412 5   330    72
#> 185   412 5   330    72
#> 186   412 5   330    72
#> 187   412 5   330    72
#> 188   412 5   330    72
#> 189   412 5   330    72
#> 190   412 5   330    72
#> 191   412 5   330    72
#> 192   412 5   330    72
#> 193   412 5   330    72
#> 194   412 5   330    72
#> 195   412 5   330    72
#> 196   412 5   330    72
#> 197   412 5   330    72
#> 198   412 5   330    72
#> 199   412 5   330    72
#> 200   412 5   330    72
#> 201   412 6   330    72
#> 202   412 6   330    72
#> 203   412 6   330    72
#> 204   412 6   330    72
#> 205   412 6   330    72
#> 206   412 6   330    72
#> 207   412 6   330    72
#> 208   412 6   330    72
#> 209   412 6   330    72
#> 210   412 6   330    72
#> 211   412 6   330    72
#> 212   412 6   330    72
#> 213   412 6   330    72
#> 214   412 6   330    72
#> 215   412 6   330    72
#> 216   412 6   330    72
#> 217   412 6   330    72
#> 218   412 6   330    72
#> 219   412 6   330    72
#> 220   412 6   330    72
#> 221   412 6   330    72
#> 222   412 6   330    72
#> 223   412 6   330    72
#> 224   412 6   330    72
#> 225   412 6   330    72
#> 226   412 6   330    72
#> 227   412 6   330    72
#> 228   412 6   330    72
#> 229   412 6   330    72
#> 230   412 6   330    72
#> 231   412 6   330    72
#> 232   412 6   330    72
#> 233   412 6   330    72
#> 234   412 6   330    72
#> 235   412 6   330    72
#> 236   412 6   330    72
#> 237   412 6   330    72
#> 238   412 6   330    72
#> 239   412 6   330    72
#> 240   412 6   330    72
#> 241   412 6   330    72
#> 242   412 6   330    72
#> 243   412 6   330    72
#> 244   412 6   330    72
#> 245   412 6   330    72
#> 246   412 6   330    72
#> 247   412 6   330    72
#> 248   412 6   330    72
#> 249   412 6   330    72
#> 250   412 6   330    72