Get statistics for the consensus partition

Get statistics for the consensus partition

# S4 method for ConsensusPartition
get_stats(object, k = object@k, all_stats = FALSE)

Arguments

object
A ConsensusPartition-class object.

k
Number of partitions. The value can be a vector.

all_stats
Whether to show all statistics that were calculated. Used internally.

Details

The statistics are:

PAC

proportion of ambiguous clustering, calculated by PAC.

mean_silhouette

the mean silhouette score. See https://en.wikipedia.org/wiki/Silhouette_(clustering) .

concordance

the mean probability that each partition fits the consensus partition, calculated by concordance.

area_increased

the increased area under ECDF (the empirical cumulative distribution function curve) to the previous k.

Rand

the Rand index which is the percent of pairs of samples that are both in a same cluster or both are not in a same cluster in the partition of k and k-1. See https://en.wikipedia.org/wiki/Rand_index .

Jaccard

the ratio of pairs of samples are both in a same cluster in the partition of k and k-1 and the pairs of samples are both in a same cluster in the partition k or k-1.

Value

A matrix of partition statistics.

Examples

data(cola_rl)
obj = cola_rl["sd", "kmeans"]
get_stats(obj)

#>   k     1-PAC mean_silhouette concordance area_increased      Rand   Jaccard
#> 2 2 0.3540100       0.8275595   0.8686667     0.46184150 0.5480226 0.5480226
#> 3 3 1.0000000       0.9842509   0.9786667     0.45750402 0.7740113 0.5876289
#> 4 4 0.7343358       0.9081489   0.8627778     0.08646858 1.0000000 1.0000000
#> 5 5 0.6791980       0.6387571   0.7048889     0.06287065 0.9638418 0.8877193
#> 6 6 0.6691729       0.3406871   0.6201111     0.04321192 0.9146893 0.7084942

get_stats(obj, k = 2)

#>   k   1-PAC mean_silhouette concordance area_increased      Rand   Jaccard
#> 2 2 0.35401       0.8275595   0.8686667      0.4618415 0.5480226 0.5480226