Get statistics

Get statistics

# S4 method for ConsensusPartitionList
get_stats(object, k, all_stats = FALSE)

Arguments

object
A ConsensusPartitionList-class object.

k
Number of subgroups. The value can only be a single value.

all_stats
Whether to show all statistics that were calculated. Used internally.

Value

A matrix of partition statistics for a selected k. Rows in the matrix correspond to combinations of top-value methods and partitioning methods.

Author

Zuguang Gu <z.gu@dkfz.de>

Examples

data(golub_cola)
get_stats(golub_cola, k = 2)

#>             k     1-PAC mean_silhouette concordance area_increased      Rand
#> SD:skmeans  2 1.0000000       0.9953730   0.9977778      0.4880976 0.5121283
#> CV:skmeans  2 1.0000000       0.9916216   0.9961111      0.4839104 0.5179969
#> MAD:skmeans 2 1.0000000       0.9720555   0.9886111      0.4869463 0.5179969
#> ATC:skmeans 2 1.0000000       0.9918746   0.9963889      0.5069181 0.4933490
#> SD:mclust   2 0.7946170       0.9258006   0.9669444      0.4928132 0.5121283
#> CV:mclust   2 0.9126427       0.9401687   0.9544444      0.4645412 0.5179969
#> MAD:mclust  2 0.9130817       0.9169410   0.9686111      0.4662366 0.5402973
#> ATC:mclust  2 0.9881475       0.9485706   0.9694444      0.3386532 0.6494523
#> SD:kmeans   2 1.0000000       0.9842643   0.9838889      0.4752546 0.5179969
#> CV:kmeans   2 1.0000000       0.9745355   0.9905556      0.4715537 0.5320814
#> MAD:kmeans  2 1.0000000       0.9994386   0.9997222      0.4760038 0.5246479
#> ATC:kmeans  2 1.0000000       0.9682558   0.9869444      0.5051609 0.4945227
#> SD:pam      2 0.8274802       0.9258975   0.9652778      0.4807613 0.5246479
#> CV:pam      2 0.7251975       0.9059746   0.9513889      0.5016730 0.4945227
#> MAD:pam     2 0.8257243       0.9257311   0.9638889      0.4842816 0.5179969
#> ATC:pam     2 0.9991220       0.9549769   0.9816667      0.4979682 0.5027387
#> SD:hclust   2 0.9556629       0.9385404   0.9716667      0.4628792 0.5402973
#> CV:hclust   2 0.4543459       0.8352513   0.9111111      0.4581123 0.5402973
#> MAD:hclust  2 0.7458297       0.9200550   0.9538889      0.4697769 0.5179969
#> ATC:hclust  2 0.6865672       0.8549144   0.9361111      0.4933646 0.4945227
#>               Jaccard
#> SD:skmeans  0.5121283
#> CV:skmeans  0.5179969
#> MAD:skmeans 0.5179969
#> ATC:skmeans 0.4933490
#> SD:mclust   0.5121283
#> CV:mclust   0.5179969
#> MAD:mclust  0.5402973
#> ATC:mclust  0.6494523
#> SD:kmeans   0.5179969
#> CV:kmeans   0.5320814
#> MAD:kmeans  0.5246479
#> ATC:kmeans  0.4945227
#> SD:pam      0.5246479
#> CV:pam      0.4945227
#> MAD:pam     0.5179969
#> ATC:pam     0.5027387
#> SD:hclust   0.5402973
#> CV:hclust   0.5402973
#> MAD:hclust  0.5179969
#> ATC:hclust  0.4945227