Hierarchical partition

Hierarchical partition

hierarchical_partition(data, top_value_method = "MAD", partition_method = "kmeans",
    PAC_cutoff = 0.1, silhouette_cutoff = 0.5,
    min_samples = 6, min_signatures = c(50, 0.05), max_k = 4, verbose = TRUE,
    mc.cores = 1, ...)

Arguments

data
A numeric matrix where subgroups are found by columns.

top_value_method
A single top-value method. Available methods are in all_top_value_methods.

partition_method
A single partition method. Available methods are in all_partition_methods.

PAC_cutoff
The cutoff of PAC scores to determine whether to continue looking for subgroups.

silhouette_cutoff
Cutoff for silhouette scores.

min_samples
The cutoff of number of samples to determine whether to continue looking for subgroups.

min_signatures
Minimal number of signatures to determine whether to continue looking for subgroups. The value should be a vector of length two. The first value is the absolute number of signatures and the second value is the proportion of signatures in the matrix. The hierarchical partitioning stops when both criterions are met.

max_k
Maximal number of partitions to try. The function will try 2:max_k partitions. Note this is the number of partitions that will be tried out on each node of the hierarchical partition. Since more subgroups will be found in the whole partition hierarchy, on each node, max_k should not be set to a large value.

verbose
Whether print message.

mc.cores
Multiple cores to use.

...
Pass to consensus_partition

Details

The function looks for subgroups in a hierarchical way.

There is a special way to encode the node in the hierarchy. The length of the node name is the depth of the node in the hierarchy and the substring excluding the last digit is the name node of the parent node. E.g. for the node 0011, the depth is 4 and the parent node is 001.

Value

A HierarchicalPartition-class object. Simply type object in the interactive R session to see which functions can be applied on it.

Examples

if (FALSE) {
set.seed(123)
m = cbind(rbind(matrix(rnorm(20*20, mean = 2, sd = 0.3), nr = 20),
                matrix(rnorm(20*20, mean = 0, sd = 0.3), nr = 20),
                matrix(rnorm(20*20, mean = 0, sd = 0.3), nr = 20)),
          rbind(matrix(rnorm(20*20, mean = 0, sd = 0.3), nr = 20),
                matrix(rnorm(20*20, mean = 1, sd = 0.3), nr = 20),
                matrix(rnorm(20*20, mean = 0, sd = 0.3), nr = 20)),
          rbind(matrix(rnorm(20*20, mean = 0, sd = 0.3), nr = 20),
                matrix(rnorm(20*20, mean = 0, sd = 0.3), nr = 20),
                matrix(rnorm(20*20, mean = 1, sd = 0.3), nr = 20))
         ) + matrix(rnorm(60*60, sd = 0.5), nr = 60)
cola_rh = hierarchical_partition(m, top_n = c(20, 30, 40), PAC_cutoff = 0.3)
}
data(cola_rh)
cola_rh

#> A 'HierarchicalPartition' object with 'MAD:kmeans' method.
#>   On a matrix with 60 rows and 60 columns.
#>   Performed in total 2250 partitions by row resampling.
#>   There are 3 groups.
#> 
#> Hierarchy of the partition:
#>   0, 60 cols
#>   |-- 01, 20 cols
#>   `-- 00, 40 cols
#>       |-- 001, 20 cols
#>       `-- 000, 20 cols
#> 
#> Following methods can be applied to this 'HierarchicalPartition' object:
#>  [1] "all_leaves"            "all_nodes"             "cola_report"          
#>  [4] "collect_classes"       "collect_plots"         "colnames"             
#>  [7] "dimension_reduction"   "get_anno_col"          "get_anno"             
#> [10] "get_classes"           "get_matrix"            "get_signatures"       
#> [13] "GO_enrichment"         "max_depth"             "ncol"                 
#> [16] "nrow"                  "rownames"              "show"                 
#> [19] "suggest_best_k"        "test_to_known_factors"
#> 
#> You can get result for a single node by e.g. object["01"]