Register user-defined top-value methods

Register user-defined top-value methods

register_top_value_methods(..., validate = TRUE)

Arguments

...
A named list of functions.

validate
Whether validate the functions.

Details

The user-defined function should accept one argument which is the data matrix where the scores are calculated by rows. Rows with top scores are treated as "top rows" in cola analysis. Following is how we register "SD" (standard deviation) top-value method:

  register_top_value_methods(SD = function(mat) apply(mat, 1, sd))  

Of course, you can use rowSds to give a faster calculation of row SD:

  register_top_value_methods(SD = rowSds)  

The registered top-value method will be used as defaults in run_all_consensus_partition_methods.

To remove a top-value method, use remove_top_value_methods.

There are four default top-value methods:

"SD"

standard deviation, by rowSds.

"CV"

coefficient variance, calculated as sd/(mean+s) where s is the 10^th percentile of all row means.

"MAD"

median absolute deviation, by rowMads.

"ATC"

the ATC method.

Value

No value is returned.

See also

Author

Zuguang Gu <z.gu@dkfz.de>

Examples

all_top_value_methods()

#> [1] "SD"  "CV"  "MAD" "ATC"

register_top_value_methods(
    ATC_spearman = function(mat) ATC(mat, method = "spearman")
)
all_top_value_methods()

#> [1] "SD"           "CV"           "MAD"          "ATC"          "ATC_spearman"

remove_top_value_methods("ATC_spearman")