Suggest the best number of subgroups

Suggest the best number of subgroups

# S4 method for ConsensusPartitionList
suggest_best_k(object, jaccard_index_cutoff = select_jaccard_cutoff(ncol(object)))

Arguments

object

A ConsensusPartitionList-class object.

jaccard_index_cutoff

The cutoff for Jaccard index for comparing to previous k.

Details

It basically gives the best k for each combination of top-value method and partitioning method by calling suggest_best_k,ConsensusPartition-method.

1-PAC score higher than 0.95 is treated as very stable partition (marked by **) and higher than 0.9 is treated as stable partition (marked by *).

Value

A data frame with the best k and other statistics for each combination of methods.

Author

Zuguang Gu <z.gu@dkfz.de>

Examples

data(golub_cola)
suggest_best_k(golub_cola)
#>             best_k     1-PAC mean_silhouette concordance    optional_k
#> SD:kmeans        2 1.0000000       0.9748797   0.9791667 **           
#> SD:skmeans       2 1.0000000       0.9878910   0.9950000 **           
#> CV:kmeans        2 1.0000000       0.9783007   0.9902778 **           
#> CV:skmeans       2 1.0000000       0.9754655   0.9894444 **           
#> CV:mclust        2 1.0000000       0.9706630   0.9875000 **           
#> MAD:kmeans       2 1.0000000       0.9954798   0.9900000 **           
#> MAD:skmeans      2 1.0000000       0.9772774   0.9902778 **           
#> ATC:skmeans      3 1.0000000       0.9788419   0.9900000 **          2
#> ATC:mclust       2 1.0000000       0.9684868   0.9819444 **           
#> SD:mclust        3 0.9957374       0.9730475   0.9869444 **          2
#> ATC:pam          3 0.9503951       0.9308092   0.9738889 **           
#> ATC:kmeans       2 0.9420544       0.9542907   0.9802778  *           
#> MAD:mclust       2 0.9139596       0.9429971   0.9766667  *           
#> ATC:hclust       2 0.8867428       0.8665934   0.9472222              
#> SD:hclust        2 0.8125549       0.9108634   0.9577778              
#> MAD:pam          2 0.8002634       0.9045510   0.9588889              
#> SD:pam           2 0.7695347       0.9214171   0.9597222              
#> CV:pam           2 0.6931519       0.8694426   0.9394444              
#> CV:hclust        2 0.5627744       0.8160426   0.9097222              
#> MAD:hclust       2 0.5610184       0.8694379   0.9308333