Test correspondance between predicted classes and known factors

Test correspondance between predicted classes and known factors

# S4 method for ConsensusPartitionList
test_to_known_factors(object, k, known = get_anno(object),
    silhouette_cutoff = 0.5, verbose = FALSE)

Arguments

object
A ConsensusPartitionList-class object.

k
Number of subgroups. It uses all k if it is not set.

known
A vector or a data frame with known factors. By default it is the annotation table set in consensus_partition or run_all_consensus_partition_methods.

silhouette_cutoff
Cutoff for sihouette scores. Samples with value less than this are omit.

verbose
Whether to print messages.

Details

The function basically sends each ConsensusPartition-class object to test_to_known_factors,ConsensusPartition-method and merges results afterwards.

Value

A data frame with the following columns:

number of samples used to test after filtered by silhouette_cutoff,

p-values from the tests,

number of subgroups.

If there are NA values, basically it means there are no efficient data points to perform the test.

See also

Author

Zuguang Gu <z.gu@dkfz.de>

Examples

data(golub_cola)
test_to_known_factors(golub_cola)

#>             n_sample ALL.AML(p-value) k
#> SD:skmeans        72     3.256935e-13 2
#> SD:skmeans        63     1.046260e-12 3
#> SD:skmeans        69     2.429664e-12 4
#> SD:skmeans        45     4.302268e-08 5
#> SD:skmeans        64     6.720007e-11 6
#> CV:skmeans        72     6.197475e-14 2
#> CV:skmeans        69     3.422182e-13 3
#> CV:skmeans        61     1.603917e-11 4
#> CV:skmeans        49     3.410657e-09 5
#> CV:skmeans        49     1.413531e-08 6
#> MAD:skmeans       71     9.787012e-14 2
#> MAD:skmeans       68     5.443398e-13 3
#> MAD:skmeans       59     2.285074e-10 4
#> MAD:skmeans       49     5.564718e-08 5
#> MAD:skmeans       53     1.607103e-07 6
#> ATC:skmeans       72     1.505160e-04 2
#> ATC:skmeans       72     1.826171e-10 3
#> ATC:skmeans       59     3.253261e-10 4
#> ATC:skmeans       68     5.428266e-10 5
#> ATC:skmeans       51     1.578559e-07 6
#> SD:mclust         70     1.547171e-13 2
#> SD:mclust         72     8.542092e-14 3
#> SD:mclust         63     3.289037e-11 4
#> SD:mclust         50     3.610865e-10 5
#> SD:mclust         59     1.955688e-11 6
#> CV:mclust         71     1.101449e-13 2
#> CV:mclust         64     6.501726e-13 3
#> CV:mclust         59     4.312854e-12 4
#> CV:mclust         33     3.220673e-07 5
#> CV:mclust         67     2.572640e-12 6
#> MAD:mclust        69     3.573622e-13 2
#> MAD:mclust        67     9.331052e-13 3
#> MAD:mclust        70     1.334254e-12 4
#> MAD:mclust        64     1.119707e-10 5
#> MAD:mclust        58     5.517492e-09 6
#> ATC:mclust        71     2.134128e-02 2
#> ATC:mclust        55     5.094604e-03 3
#> ATC:mclust        60     5.227652e-06 4
#> ATC:mclust        70     4.001814e-10 5
#> ATC:mclust        62     1.090872e-08 6
#> SD:kmeans         72     6.197475e-14 2
#> SD:kmeans         65     4.043137e-13 3
#> SD:kmeans         57     1.031365e-10 4
#> SD:kmeans         57     3.801812e-10 5
#> SD:kmeans         46     2.462851e-09 6
#> CV:kmeans         71     1.376488e-13 2
#> CV:kmeans         67     1.566450e-13 3
#> CV:kmeans         59     3.679317e-12 4
#> CV:kmeans         59     1.667791e-10 5
#> CV:kmeans         62     4.689400e-12 6
#> MAD:kmeans        72     5.108677e-13 2
#> MAD:kmeans        69     2.536450e-12 3
#> MAD:kmeans        50     1.388794e-11 4
#> MAD:kmeans        52     1.416312e-08 5
#> MAD:kmeans        41     6.603666e-07 6
#> ATC:kmeans        71     2.020619e-04 2
#> ATC:kmeans        70     2.687880e-06 3
#> ATC:kmeans        48     4.391333e-07 4
#> ATC:kmeans        67     5.236675e-07 5
#> ATC:kmeans        67     3.654693e-07 6
#> SD:pam            71     1.376488e-13 2
#> SD:pam            46     4.068465e-09 3
#> SD:pam            66     1.654487e-12 4
#> SD:pam            54     5.765138e-10 5
#> SD:pam            46     2.462851e-09 6
#> CV:pam            69     6.530214e-10 2
#> CV:pam            61     2.970743e-12 3
#> CV:pam            67     1.052938e-12 4
#> CV:pam            55     3.547954e-09 5
#> CV:pam            44     6.415777e-09 6
#> MAD:pam           71     7.945497e-13 2
#> MAD:pam           52     1.191102e-09 3
#> MAD:pam           64     1.568666e-10 4
#> MAD:pam           63     9.244743e-11 5
#> MAD:pam           63     4.009798e-10 6
#> ATC:pam           70     8.115728e-07 2
#> ATC:pam           70     1.088870e-06 3
#> ATC:pam           71     1.518569e-05 4
#> ATC:pam           62     9.970617e-07 5
#> ATC:pam           52     3.562270e-05 6
#> SD:hclust         72     1.946919e-16 2
#> SD:hclust         70     6.305117e-16 3
#> SD:hclust         49     2.289735e-11 4
#> SD:hclust         31     3.061396e-06 5
#> SD:hclust         32     1.913098e-06 6
#> CV:hclust         68     1.519850e-15 2
#> CV:hclust         53     3.098819e-12 3
#> CV:hclust         53     1.833362e-11 4
#> CV:hclust         41     2.687829e-08 5
#> CV:hclust         50     1.385797e-09 6
#> MAD:hclust        70     1.547171e-13 2
#> MAD:hclust        27               NA 3
#> MAD:hclust        55     6.866170e-12 4
#> MAD:hclust        31     5.584731e-06 5
#> MAD:hclust        37     6.891178e-07 6
#> ATC:hclust        66     7.528671e-05 2
#> ATC:hclust        62     4.578363e-05 3
#> ATC:hclust        55     2.554227e-02 4
#> ATC:hclust        53     1.412179e-05 5
#> ATC:hclust        53     2.455890e-05 6