Test correspondance between predicted subgroups and known factors

Test correspondance between predicted subgroups and known factors

# S4 method for DownSamplingConsensusPartition
test_to_known_factors(object, k, known = get_anno(object),
    p_cutoff = 0.05, verbose = FALSE)

Arguments

object
A DownSamplingConsensusPartition-class object.

k
Number of subgroups. It uses all k if it is not specified.

known
A vector or a data frame with known factors. By default it is the annotation table set in consensus_partition_by_down_sampling.

p_cutoff
Cutoff for p-values for the class prediction. Samples with p-value higher than it are omit.

verbose
Whether to print messages.

Details

The test is performed by test_between_factors between the predicted classes and user's annotation table.

Value

A data frame with the following columns:

number of samples used to test after filtered by p_cutoff,

p-values from the tests,

number of subgroups.

Author

Zuguang Gu <z.gu@dkfz.de>

Examples

data(golub_cola_ds)
test_to_known_factors(golub_cola_ds, k = 3)

#>             n_sample ALL.AML(p-value) k
#> ATC:skmeans       67     9.399449e-09 3

test_to_known_factors(golub_cola_ds)

#>             n_sample ALL.AML(p-value) k
#> ATC:skmeans       70     1.434142e-04 2
#> ATC:skmeans       67     9.399449e-09 3
#> ATC:skmeans       68     2.926023e-08 4
#> ATC:skmeans       49     2.534392e-02 5
#> ATC:skmeans       60     1.437048e-05 6