The ConsensusPartitionList class

The ConsensusPartitionList class

Details

The object contains results from all combinations of top-value methods and partition methods.

Methods

The ConsensusPartitionList-class provides following methods:

run_all_consensus_partition_methods:

constructor method.

top_rows_overlap,ConsensusPartitionList-method:

plot the overlaps of top rows under different top-value methods.

top_rows_heatmap,ConsensusPartitionList-method:

plot the heatmap of top rows under different top-value methods.

get_classes,ConsensusPartitionList-method:

get consensus class IDs merging from all methods.

get_matrix,ConsensusPartition-method:

get the original matrix.

get_stats,ConsensusPartitionList-method:

get metrics for a specified k.

get_membership,ConsensusPartitionList-method:

get consensus membership matrix summarized from all methods.

suggest_best_k,ConsensusPartitionList-method:

guess the best number of partitions for all methods.

collect_plots,ConsensusPartitionList-method:

collect plots from all combinations of top-value methods and partition methods with choosing a plotting function.

collect_classes,ConsensusPartitionList-method:

make a plot which contains predicted classes from all combinations of top-value methods and partition methods.

test_to_known_factors,ConsensusPartitionList-method:

test correlation between predicted classes and known annotations, if provided.

cola_report,ConsensusPartitionList-method:

generate a HTML report for the whole analysis.

functional_enrichment,ConsensusPartitionList-method:

perform functional enrichment analysis on significant genes if rows in the matrix can be corresponded to genes.

See also

Examples

# There is no example
NULL

#> NULL