Adapted PAC scores

Adapted PAC scores

aPAC(consensus_mat)

Arguments

consensus_mat
A consensus matrix.

Details

For the consensus values x, it is transformed to 1 - x if x < 0.5. After the transformation, for any pair of samples in the consensus matrix, If they are always in a same group or always in different groups, the value x is both to 1. Thus, if the consensus matrix shows stable partitions, values x will be all close to 1. Reflected in the CDF of x, the curve is shifted to the right and the area under CDF curve should be very small.

An aPAC value less than 0.05 is considered as the stable partition, which can be thought the proportion of abmiguous partitioning is less than 0.05.

Value

A numeric value.

Examples

data(cola_rl)
aPAC(get_consensus(cola_rl[1, 1], k = 2))

#> [1] 0.3051786

aPAC(get_consensus(cola_rl[1, 1], k = 3))

#> [1] 0.09455614

aPAC(get_consensus(cola_rl[1, 1], k = 4))

#> [1] 0.1459211

aPAC(get_consensus(cola_rl[1, 1], k = 5))

#> [1] 0.2022039

aPAC(get_consensus(cola_rl[1, 1], k = 6))

#> [1] 0.2388224