Get subgroup labels

Get subgroup labels

# S4 method for ConsensusPartition
get_classes(object, k = object@k)

Arguments

object
A ConsensusPartition-class object.

k
Number of subgroups.

Value

A data frame with subgroup labels and other columns which are entropy of the percent membership matrix and the silhouette scores which measure the stability of a sample to stay in its group.

If k is not specified, it returns a data frame with subgroup labels from all k.

Examples

data(golub_cola)
obj = golub_cola["ATC", "skmeans"]
get_classes(obj, k = 2)

#>           class    entropy silhouette
#> sample_39     1 0.00000000  0.9820987
#> sample_40     1 0.00000000  0.9820987
#> sample_42     2 0.00000000  0.9997442
#> sample_47     2 0.00000000  0.9997442
#> sample_48     2 0.00000000  0.9997442
#> sample_49     1 0.00000000  0.9820987
#> sample_41     2 0.00000000  0.9997442
#> sample_43     2 0.00000000  0.9997442
#> sample_44     2 0.00000000  0.9997442
#> sample_45     2 0.00000000  0.9997442
#> sample_46     2 0.00000000  0.9997442
#> sample_70     2 0.00000000  0.9997442
#> sample_71     2 0.00000000  0.9997442
#> sample_72     2 0.00000000  0.9997442
#> sample_68     2 0.00000000  0.9997442
#> sample_69     2 0.00000000  0.9997442
#> sample_67     2 0.00000000  0.9997442
#> sample_55     1 0.00000000  0.9820987
#> sample_56     1 0.00000000  0.9820987
#> sample_59     2 0.00000000  0.9997442
#> sample_52     1 0.00000000  0.9820987
#> sample_53     1 0.00000000  0.9820987
#> sample_51     1 0.00000000  0.9820987
#> sample_50     1 0.00000000  0.9820987
#> sample_54     2 0.00000000  0.9997442
#> sample_57     1 0.00000000  0.9820987
#> sample_58     1 0.00000000  0.9820987
#> sample_60     1 0.91695673  0.5149407
#> sample_61     1 0.00000000  0.9820987
#> sample_65     1 0.00000000  0.9820987
#> sample_66     2 0.00000000  0.9997442
#> sample_63     1 0.00000000  0.9820987
#> sample_64     1 0.00000000  0.9820987
#> sample_62     1 0.00000000  0.9820987
#> sample_1      1 0.00000000  0.9820987
#> sample_2      2 0.00000000  0.9997442
#> sample_3      1 0.00000000  0.9820987
#> sample_4      1 0.00000000  0.9820987
#> sample_5      2 0.00000000  0.9997442
#> sample_6      1 0.00000000  0.9820987
#> sample_7      1 0.00000000  0.9820987
#> sample_8      1 0.00000000  0.9820987
#> sample_9      2 0.00000000  0.9997442
#> sample_10     2 0.06722154  0.9917483
#> sample_11     2 0.00000000  0.9997442
#> sample_12     1 0.00000000  0.9820987
#> sample_13     2 0.00000000  0.9997442
#> sample_14     2 0.00000000  0.9997442
#> sample_15     2 0.00000000  0.9997442
#> sample_16     2 0.00000000  0.9997442
#> sample_17     2 0.00000000  0.9997442
#> sample_18     1 0.92859509  0.4887252
#> sample_19     2 0.00000000  0.9997442
#> sample_20     2 0.00000000  0.9997442
#> sample_21     2 0.00000000  0.9997442
#> sample_22     1 0.00000000  0.9820987
#> sample_23     1 0.00000000  0.9820987
#> sample_24     2 0.00000000  0.9997442
#> sample_25     1 0.00000000  0.9820987
#> sample_26     2 0.00000000  0.9997442
#> sample_27     1 0.00000000  0.9820987
#> sample_34     1 0.00000000  0.9820987
#> sample_35     1 0.00000000  0.9820987
#> sample_36     1 0.00000000  0.9820987
#> sample_37     1 0.00000000  0.9820987
#> sample_38     1 0.00000000  0.9820987
#> sample_28     1 0.00000000  0.9820987
#> sample_29     2 0.00000000  0.9997442
#> sample_30     1 0.00000000  0.9820987
#> sample_31     1 0.00000000  0.9820987
#> sample_32     1 0.00000000  0.9820987
#> sample_33     1 0.00000000  0.9820987

get_classes(obj)

#>           k=2 k=3 k=4 k=5 k=6
#> sample_39   1   3   3   3   3
#> sample_40   1   3   3   3   3
#> sample_42   2   2   4   4   6
#> sample_47   2   2   2   2   2
#> sample_48   2   2   2   2   2
#> sample_49   1   3   3   3   3
#> sample_41   2   2   2   2   2
#> sample_43   2   2   2   2   2
#> sample_44   2   2   2   2   2
#> sample_45   2   2   2   2   2
#> sample_46   2   2   2   2   2
#> sample_70   2   2   2   2   2
#> sample_71   2   2   4   4   6
#> sample_72   2   2   4   4   6
#> sample_68   2   2   2   2   2
#> sample_69   2   2   2   2   2
#> sample_67   2   2   4   4   6
#> sample_55   1   3   3   3   3
#> sample_56   1   3   3   3   3
#> sample_59   2   2   2   5   5
#> sample_52   1   1   1   1   5
#> sample_53   1   1   1   1   1
#> sample_51   1   1   1   1   1
#> sample_50   1   1   1   1   1
#> sample_54   2   2   2   5   5
#> sample_57   1   1   1   1   5
#> sample_58   1   1   1   1   1
#> sample_60   1   1   1   5   5
#> sample_61   1   1   1   1   1
#> sample_65   1   1   1   1   1
#> sample_66   2   2   2   2   2
#> sample_63   1   1   1   1   5
#> sample_64   1   1   1   1   3
#> sample_62   1   1   1   1   5
#> sample_1    1   3   3   3   3
#> sample_2    2   2   2   4   6
#> sample_3    1   3   3   3   4
#> sample_4    1   3   3   3   3
#> sample_5    2   2   2   2   2
#> sample_6    1   3   3   3   4
#> sample_7    1   3   3   3   3
#> sample_8    1   3   3   3   3
#> sample_9    2   2   2   2   2
#> sample_10   2   3   4   4   4
#> sample_11   2   2   4   4   6
#> sample_12   1   1   1   1   1
#> sample_13   2   2   2   2   2
#> sample_14   2   2   2   2   2
#> sample_15   2   2   2   2   2
#> sample_16   2   2   2   2   2
#> sample_17   2   2   2   2   2
#> sample_18   1   3   3   5   4
#> sample_19   2   2   2   2   2
#> sample_20   2   2   2   2   2
#> sample_21   2   2   2   2   2
#> sample_22   1   1   1   1   1
#> sample_23   1   3   3   3   4
#> sample_24   2   2   2   2   2
#> sample_25   1   3   3   3   1
#> sample_26   2   3   2   2   2
#> sample_27   1   3   3   3   3
#> sample_34   1   1   1   1   1
#> sample_35   1   1   1   1   3
#> sample_36   1   1   1   1   1
#> sample_37   1   1   1   1   1
#> sample_38   1   1   1   1   1
#> sample_28   1   1   1   1   1
#> sample_29   2   2   2   4   6
#> sample_30   1   1   1   1   1
#> sample_31   1   1   1   1   1
#> sample_32   1   1   1   1   1
#> sample_33   1   1   1   1   1