Get class IDs from the ConsensusPartition object

Get class IDs from the ConsensusPartition object

# S4 method for ConsensusPartition
get_classes(object, k = object@k)

Arguments

object
A ConsensusPartition-class object.

k
Number of partitions.

Value

A data frame with class IDs and other columns which are entropy of the percent membership matrix and the silhouette scores which measure the stability of a sample to stay in its group.

If k is not specified, it returns a data frame with class IDs from every k.

Examples

data(cola_rl)
obj = cola_rl["sd", "kmeans"]
get_classes(obj, k = 2)

#>    class    entropy silhouette
#> 1      1 0.63430955  0.7561385
#> 2      1 0.60143539  0.7650894
#> 3      1 0.56650951  0.7684535
#> 4      1 0.48977901  0.7711975
#> 5      1 0.63430955  0.7566534
#> 6      1 0.54822512  0.7693620
#> 7      1 0.42524461  0.7669570
#> 8      1 0.60143539  0.7650894
#> 9      1 0.60143539  0.7650894
#> 10     1 0.60143539  0.7650894
#> 11     1 0.40217919  0.7659594
#> 12     1 0.58423881  0.7672999
#> 13     1 0.61811949  0.7615428
#> 14     1 0.60143539  0.7650894
#> 15     1 0.60143539  0.7650894
#> 16     1 0.58423881  0.7672999
#> 17     1 0.58423881  0.7672999
#> 18     1 0.56650951  0.7684535
#> 19     1 0.60143539  0.7650894
#> 20     1 0.61811949  0.7615428
#> 21     2 0.00000000  0.9914726
#> 22     2 0.00000000  0.9914726
#> 23     2 0.00000000  0.9914726
#> 24     2 0.00000000  0.9914726
#> 25     2 0.00000000  0.9914726
#> 26     2 0.00000000  0.9914726
#> 27     2 0.00000000  0.9914726
#> 28     2 0.00000000  0.9914726
#> 29     2 0.00000000  0.9914726
#> 30     2 0.10215804  0.9765606
#> 31     2 0.24229219  0.9469803
#> 32     2 0.10215804  0.9818215
#> 33     2 0.00000000  0.9914726
#> 34     2 0.00000000  0.9914726
#> 35     2 0.00000000  0.9914726
#> 36     2 0.05777796  0.9876216
#> 37     2 0.05777796  0.9876216
#> 38     2 0.00000000  0.9914726
#> 39     2 0.10215804  0.9818215
#> 40     2 0.14144054  0.9742401
#> 41     1 0.81651871  0.7333465
#> 42     1 0.81651871  0.7333465
#> 43     1 0.81651871  0.7333465
#> 44     1 0.77215514  0.7450517
#> 45     1 0.79504028  0.7409190
#> 46     1 0.80595204  0.7376163
#> 47     1 0.81651871  0.7333465
#> 48     1 0.87298829  0.6863955
#> 49     1 0.81651871  0.7333465
#> 50     1 0.80595204  0.7379172
#> 51     1 0.81651871  0.7333465
#> 52     1 0.63430955  0.7550444
#> 53     1 0.80595204  0.7378319
#> 54     1 0.99679163  0.3188277
#> 55     1 0.96261471  0.5272889
#> 56     1 0.80595204  0.7376163
#> 57     1 0.79504028  0.7409620
#> 58     1 0.76016750  0.7466591
#> 59     1 0.81651871  0.7333465
#> 60     1 0.81651871  0.7333465

get_classes(obj)

#>    k=2 k=3 k=4 k=5 k=6
#> 1    1   1   1   1   1
#> 2    1   1   1   1   1
#> 3    1   1   1   1   1
#> 4    1   1   1   1   1
#> 5    1   1   1   1   1
#> 6    1   1   1   1   1
#> 7    1   1   1   1   4
#> 8    1   1   1   1   1
#> 9    1   1   1   1   1
#> 10   1   1   1   1   1
#> 11   1   1   1   1   4
#> 12   1   1   1   4   1
#> 13   1   1   1   1   1
#> 14   1   1   1   1   1
#> 15   1   1   1   1   1
#> 16   1   1   1   1   1
#> 17   1   1   1   1   1
#> 18   1   1   1   1   1
#> 19   1   1   1   1   1
#> 20   1   1   1   1   1
#> 21   2   2   2   2   2
#> 22   2   2   2   2   2
#> 23   2   2   2   2   2
#> 24   2   2   2   2   2
#> 25   2   2   2   2   2
#> 26   2   2   2   2   2
#> 27   2   2   2   2   5
#> 28   2   2   2   2   2
#> 29   2   2   2   2   5
#> 30   2   2   2   2   5
#> 31   2   2   2   2   2
#> 32   2   2   2   2   2
#> 33   2   2   2   2   2
#> 34   2   2   2   2   5
#> 35   2   2   2   2   5
#> 36   2   2   2   2   2
#> 37   2   2   2   2   2
#> 38   2   2   2   2   5
#> 39   2   2   2   2   2
#> 40   2   2   2   2   2
#> 41   1   3   3   5   3
#> 42   1   3   3   3   3
#> 43   1   3   3   3   6
#> 44   1   3   3   5   3
#> 45   1   3   3   5   3
#> 46   1   3   3   3   3
#> 47   1   3   3   3   6
#> 48   1   3   3   3   3
#> 49   1   3   3   3   3
#> 50   1   3   3   3   3
#> 51   1   3   3   3   6
#> 52   1   3   3   3   3
#> 53   1   3   3   3   3
#> 54   1   3   3   3   3
#> 55   1   3   3   3   3
#> 56   1   3   3   3   6
#> 57   1   3   3   3   6
#> 58   1   3   3   3   3
#> 59   1   3   3   3   6
#> 60   1   3   3   3   6