Get class IDs from the ConsensusPartition object

Get class IDs from the ConsensusPartition object

# S4 method for ConsensusPartition
get_classes(object, k = object@k)

Arguments

object
A ConsensusPartition-class object.

k
Number of partitions.

Value

A data frame with class IDs and other columns which are entropy of the percent membership matrix and the silhouette scores which measure the stability of a sample to stay in its group.

If k is not specified, it returns a data frame with class IDs from every k.

Examples

data(cola_rl)
obj = cola_rl["sd", "kmeans"]
get_classes(obj, k = 2)

#>    class    entropy silhouette
#> 1      1 0.63430955  0.7715207
#> 2      1 0.63430955  0.7715207
#> 3      1 0.63430955  0.7715207
#> 4      1 0.44749926  0.7756645
#> 5      1 0.65002242  0.7664377
#> 6      1 0.58423881  0.7756748
#> 7      1 0.44749926  0.7753550
#> 8      1 0.63430955  0.7715207
#> 9      1 0.63430955  0.7715207
#> 10     1 0.63430955  0.7715207
#> 11     1 0.44749926  0.7756645
#> 12     1 0.56650951  0.7763642
#> 13     1 0.63430955  0.7715207
#> 14     1 0.63430955  0.7715207
#> 15     1 0.63430955  0.7715207
#> 16     1 0.63430955  0.7715207
#> 17     1 0.61811949  0.7734234
#> 18     1 0.60143539  0.7745092
#> 19     1 0.63430955  0.7715207
#> 20     1 0.63430955  0.7715207
#> 21     2 0.05777796  0.9867899
#> 22     2 0.05777796  0.9867899
#> 23     2 0.05777796  0.9836094
#> 24     2 0.05777796  0.9867899
#> 25     2 0.00000000  0.9880458
#> 26     2 0.00000000  0.9880458
#> 27     2 0.10215804  0.9790155
#> 28     2 0.00000000  0.9880458
#> 29     2 0.05777796  0.9867899
#> 30     2 0.05777796  0.9867899
#> 31     2 0.30039142  0.9235074
#> 32     2 0.05777796  0.9847722
#> 33     2 0.05777796  0.9867899
#> 34     2 0.00000000  0.9880458
#> 35     2 0.00000000  0.9880458
#> 36     2 0.05777796  0.9847722
#> 37     2 0.10215804  0.9776535
#> 38     2 0.05777796  0.9867899
#> 39     2 0.05777796  0.9847722
#> 40     2 0.05777796  0.9847722
#> 41     1 0.77215514  0.7543333
#> 42     1 0.78377695  0.7498975
#> 43     1 0.78377695  0.7498975
#> 44     1 0.74780616  0.7603918
#> 45     1 0.76016750  0.7583337
#> 46     1 0.74780616  0.7603918
#> 47     1 0.78377695  0.7498975
#> 48     1 0.83664074  0.7136809
#> 49     1 0.78377695  0.7498975
#> 50     1 0.77215514  0.7543333
#> 51     1 0.78377695  0.7498975
#> 52     1 0.58423881  0.7682275
#> 53     1 0.73506269  0.7618119
#> 54     1 0.98442690  0.4117493
#> 55     1 0.93700983  0.5805883
#> 56     1 0.76016750  0.7583337
#> 57     1 0.76016750  0.7578507
#> 58     1 0.76016750  0.7578507
#> 59     1 0.77215514  0.7543333
#> 60     1 0.78377695  0.7498975

get_classes(obj)

#>    k=2 k=3 k=4 k=5 k=6
#> 1    1   1   1   1   1
#> 2    1   1   1   1   1
#> 3    1   1   1   1   1
#> 4    1   1   1   1   4
#> 5    1   1   1   1   1
#> 6    1   1   1   1   1
#> 7    1   1   1   1   1
#> 8    1   1   1   1   1
#> 9    1   1   1   1   1
#> 10   1   1   1   1   1
#> 11   1   1   1   1   1
#> 12   1   1   1   1   4
#> 13   1   1   1   1   1
#> 14   1   1   1   1   1
#> 15   1   1   1   1   1
#> 16   1   1   1   1   1
#> 17   1   1   1   1   1
#> 18   1   1   1   1   1
#> 19   1   1   1   1   1
#> 20   1   1   1   1   1
#> 21   2   2   2   2   2
#> 22   2   2   2   2   2
#> 23   2   2   2   2   2
#> 24   2   2   2   2   2
#> 25   2   2   2   2   2
#> 26   2   2   2   2   2
#> 27   2   2   2   2   5
#> 28   2   2   2   2   2
#> 29   2   2   2   2   5
#> 30   2   2   2   2   2
#> 31   2   2   2   2   2
#> 32   2   2   2   2   2
#> 33   2   2   2   2   2
#> 34   2   2   2   2   5
#> 35   2   2   2   2   5
#> 36   2   2   2   2   2
#> 37   2   2   2   2   2
#> 38   2   2   2   2   2
#> 39   2   2   2   2   2
#> 40   2   2   2   2   2
#> 41   1   3   3   3   6
#> 42   1   3   3   3   6
#> 43   1   3   3   4   3
#> 44   1   3   3   3   6
#> 45   1   3   3   3   6
#> 46   1   3   3   3   6
#> 47   1   3   3   4   3
#> 48   1   3   3   3   6
#> 49   1   3   3   3   6
#> 50   1   3   3   3   6
#> 51   1   3   3   3   3
#> 52   1   3   3   3   6
#> 53   1   3   3   3   6
#> 54   1   3   3   3   6
#> 55   1   3   3   3   6
#> 56   1   3   3   4   3
#> 57   1   3   3   4   3
#> 58   1   3   3   3   6
#> 59   1   3   3   3   3
#> 60   1   3   3   3   3