Get subgroup labels

Get subgroup labels

# S4 method for ConsensusPartitionList
get_classes(object, k)

Arguments

object
A ConsensusPartitionList-class object.

k
Number of subgroups.

Details

The subgroup labels are inferred by merging partitions from all methods by weighting the mean silhouette scores in each method.

Value

A data frame with subgroup labels and other columns which are entropy of the percent membership matrix and the silhouette scores which measure the stability of a sample to stay in its group.

Examples

data(golub_cola)
get_classes(golub_cola, k = 2)

#>    class    entropy silhouette
#> 1      1 0.96533165  0.7505154
#> 2      1 0.99590637  0.5666583
#> 3      1 0.39304833  0.8388608
#> 4      1 0.11591029  0.8500240
#> 5      1 0.05414058  0.8500240
#> 6      1 0.99655883  0.6277406
#> 7      1 0.05070886  0.8500240
#> 8      1 0.22002818  0.8388608
#> 9      1 0.11812529  0.8500240
#> 10     1 0.11909828  0.8500240
#> 11     1 0.16724883  0.8500240
#> 12     1 0.50792051  0.8429163
#> 13     1 0.45983587  0.8287910
#> 14     1 0.36026742  0.8388608
#> 15     1 0.05070886  0.8500240
#> 16     1 0.05234636  0.8500240
#> 17     2 0.99227801  0.3530249
#> 18     1 0.98527242  0.6277406
#> 19     1 0.95831675  0.7742325
#> 20     1 0.44223285  0.8388608
#> 21     2 0.21059316  0.9498095
#> 22     2 0.05848631  0.9498095
#> 23     2 0.05528721  0.9498095
#> 24     2 0.05528721  0.9498095
#> 25     2 0.84006967  0.7241225
#> 26     2 0.25336525  0.9498095
#> 27     2 0.06671818  0.9498095
#> 28     2 0.61669566  0.8484613
#> 29     2 0.07904313  0.9498095
#> 30     2 0.07385061  0.9498095
#> 31     1 0.85025315  0.6549276
#> 32     2 0.20970171  0.9498095
#> 33     2 0.32560409  0.9498095
#> 34     2 0.22006977  0.9498095
#> 35     1 0.80993708  0.7742325
#> 36     1 0.96182446  0.4974049
#> 37     1 0.98205724  0.5666583
#> 38     1 0.87022691  0.7742325
#> 39     1 0.05070886  0.8500240
#> 40     1 0.98585425  0.6277406
#> 41     1 0.97795404  0.6277406
#> 42     1 0.98130972  0.7048961
#> 43     1 0.39657755  0.8500240
#> 44     1 0.92002316  0.6849900
#> 45     1 0.35461758  0.8388608
#> 46     2 0.45641902  0.7548411
#> 47     1 0.05414058  0.8500240
#> 48     1 0.21945533  0.8500240
#> 49     1 0.05070886  0.8500240
#> 50     1 0.21143285  0.8500240
#> 51     1 0.35324693  0.8500240
#> 52     1 0.74209711  0.8042626
#> 53     1 0.20397989  0.8500240
#> 54     1 0.05189532  0.8500240
#> 55     1 0.09392989  0.8500240
#> 56     1 0.99136469  0.6443877
#> 57     1 0.98298797  0.5666583
#> 58     1 0.07798073  0.8500240
#> 59     1 0.94547738  0.7505154
#> 60     1 0.34779975  0.8429163
#> 61     1 0.99731563  0.6277406
#> 62     2 0.10810020  0.9498095
#> 63     2 0.23570078  0.9498095
#> 64     2 0.06706583  0.9498095
#> 65     2 0.05528721  0.9498095
#> 66     2 0.10101179  0.9498095
#> 67     2 0.11035134  0.9498095
#> 68     2 0.85664678  0.7356050
#> 69     2 0.14305025  0.9498095
#> 70     2 0.09785362  0.9498095
#> 71     2 0.09237835  0.9498095
#> 72     2 0.05528721  0.9498095