Get subgroup labels

Get subgroup labels

# S4 method for ConsensusPartitionList
get_classes(object, k)

Arguments

object
A ConsensusPartitionList-class object.

k
Number of subgroups.

Details

The subgroup labels are inferred by merging partitions from all methods by weighting the mean silhouette scores in each method.

Value

A data frame with subgroup labels and other columns which are entropy of the percent membership matrix and the silhouette scores which measure the stability of a sample to stay in its group.

Author

Zuguang Gu <z.gu@dkfz.de>

Examples

data(golub_cola)
get_classes(golub_cola, k = 2)

#>           class    entropy  silhouette
#> sample_39     2 0.83546869  0.77456368
#> sample_40     2 0.88110176  0.72177524
#> sample_42     2 0.38094115  0.87509347
#> sample_47     2 0.08659599  0.86861045
#> sample_48     2 0.03006108  0.86861045
#> sample_49     2 0.88037515  0.72177524
#> sample_41     2 0.03006108  0.86861045
#> sample_43     2 0.33458476  0.87509347
#> sample_44     2 0.03425504  0.86861045
#> sample_45     2 0.05625610  0.86861045
#> sample_46     2 0.14211153  0.86861045
#> sample_70     2 0.33728419  0.87509347
#> sample_71     2 0.40319077  0.87509347
#> sample_72     2 0.36544018  0.87509347
#> sample_68     2 0.03006108  0.86861045
#> sample_69     2 0.03834360  0.86861045
#> sample_67     1 0.89277364  0.52123495
#> sample_55     2 0.87798589  0.77687868
#> sample_56     2 0.83309360  0.77456368
#> sample_59     2 0.42633580  0.87509347
#> sample_52     1 0.10384674  0.93236179
#> sample_53     1 0.01261116  0.93236179
#> sample_51     1 0.01261116  0.93236179
#> sample_50     1 0.01261116  0.93236179
#> sample_54     1 0.79977030  0.68466844
#> sample_57     1 0.10136692  0.93236179
#> sample_58     1 0.04161273  0.93236179
#> sample_60     1 0.75757068  0.68466844
#> sample_61     1 0.04554145  0.93236179
#> sample_65     1 0.04503343  0.93236179
#> sample_66     2 0.98021214 -0.03954292
#> sample_63     1 0.10698274  0.93236179
#> sample_64     1 0.29284090  0.93236179
#> sample_62     1 0.10883934  0.93236179
#> sample_1      2 0.73433090  0.81258118
#> sample_2      1 0.95123756  0.40154933
#> sample_3      2 0.84187985  0.77687868
#> sample_4      2 0.73038992  0.81258118
#> sample_5      2 0.03006108  0.86861045
#> sample_6      2 0.86609165  0.77687868
#> sample_7      2 0.87640935  0.72177524
#> sample_8      2 0.83295898  0.77456368
#> sample_9      2 0.37780274  0.87509347
#> sample_10     2 0.67163462  0.82205795
#> sample_11     2 0.38916224  0.87509347
#> sample_12     1 0.50356469  0.82668750
#> sample_13     2 0.03006108  0.86861045
#> sample_14     2 0.39163184  0.86861045
#> sample_15     2 0.03006108  0.86861045
#> sample_16     2 0.31562152  0.87509347
#> sample_17     2 0.78051729  0.80481898
#> sample_18     2 0.53680882  0.86193165
#> sample_19     2 0.31562152  0.87509347
#> sample_20     2 0.03006108  0.86861045
#> sample_21     2 0.03904881  0.86861045
#> sample_22     2 0.84458439  0.77456368
#> sample_23     2 0.85525242  0.72177524
#> sample_24     2 0.05582276  0.86861045
#> sample_25     2 0.77025035  0.81258118
#> sample_26     2 0.37759644  0.87509347
#> sample_27     2 0.88820455  0.72177524
#> sample_34     1 0.07921029  0.93236179
#> sample_35     1 0.22330137  0.93236179
#> sample_36     1 0.01261116  0.93236179
#> sample_37     1 0.01261116  0.93236179
#> sample_38     1 0.03949932  0.93236179
#> sample_28     1 0.08361511  0.93236179
#> sample_29     1 0.75500290  0.73701828
#> sample_30     1 0.03922760  0.93236179
#> sample_31     1 0.09014840  0.93236179
#> sample_32     1 0.03760030  0.93236179
#> sample_33     1 0.01261116  0.93236179