Get subgroup labels

Get subgroup labels

# S4 method for ConsensusPartitionList
get_classes(object, k)

Arguments

object
A ConsensusPartitionList-class object.

k
Number of subgroups.

Details

The subgroup labels are inferred by merging partitions from all methods by weighting the mean silhouette scores in each method.

Value

A data frame with subgroup labels and other columns which are entropy of the percent membership matrix and the silhouette scores which measure the stability of a sample to stay in its group.

Examples

data(golub_cola)
get_classes(golub_cola, k = 2)

#>    class   entropy  silhouette
#> 1      1 0.9745816 -0.13512803
#> 2      1 0.9505309  0.15888204
#> 3      2 0.4573230  0.87602469
#> 4      2 0.2987762  0.90478075
#> 5      2 0.2866128  0.90478075
#> 6      1 0.9313801  0.05490166
#> 7      2 0.2874615  0.90478075
#> 8      2 0.3325761  0.90478075
#> 9      2 0.3092411  0.90478075
#> 10     2 0.3083871  0.90478075
#> 11     2 0.3272761  0.90478075
#> 12     2 0.6445493  0.77682292
#> 13     2 0.5248221  0.84520503
#> 14     2 0.4160908  0.87602469
#> 15     2 0.2866128  0.90478075
#> 16     2 0.2866128  0.90478075
#> 17     2 0.9702734  0.02100302
#> 18     1 0.9804771  0.05490166
#> 19     1 0.9875102 -0.22127853
#> 20     2 0.5139632  0.90478075
#> 21     1 0.3720243  0.74531984
#> 22     1 0.2910719  0.74531984
#> 23     1 0.2889676  0.74531984
#> 24     1 0.2885445  0.74531984
#> 25     1 0.9714686  0.50963550
#> 26     1 0.4103805  0.72018093
#> 27     1 0.2929520  0.74531984
#> 28     1 0.6781747  0.72018093
#> 29     1 0.3005779  0.74531984
#> 30     1 0.2998526  0.74531984
#> 31     2 0.6531613  0.77908930
#> 32     1 0.3723692  0.74531984
#> 33     1 0.3963101  0.74531984
#> 34     1 0.3842806  0.74531984
#> 35     2 0.9207998  0.63386816
#> 36     2 0.8372073  0.50145348
#> 37     1 0.9804531 -0.07848358
#> 38     2 0.9457552  0.54925398
#> 39     2 0.2866128  0.90478075
#> 40     1 0.9644586 -0.07848358
#> 41     1 0.9825241 -0.07848358
#> 42     1 0.9447433  0.03397883
#> 43     2 0.5419271  0.90478075
#> 44     2 0.9995427  0.10061725
#> 45     2 0.4715703  0.87602469
#> 46     1 0.4642694  0.69880208
#> 47     2 0.2866128  0.90478075
#> 48     2 0.3420120  0.90478075
#> 49     2 0.2866128  0.90478075
#> 50     2 0.3347742  0.90478075
#> 51     2 0.3365098  0.87264129
#> 52     2 0.8917136  0.70778725
#> 53     2 0.3318953  0.90478075
#> 54     2 0.2866128  0.90478075
#> 55     2 0.3053886  0.90478075
#> 56     1 0.9343241  0.23539996
#> 57     1 0.9653853 -0.07848358
#> 58     2 0.2971360  0.90478075
#> 59     1 0.9971391 -0.39592757
#> 60     2 0.4185503  0.90478075
#> 61     1 0.9176978  0.33413939
#> 62     1 0.3137126  0.74531984
#> 63     1 0.3744023  0.74531984
#> 64     1 0.3146950  0.74531984
#> 65     1 0.2893254  0.74531984
#> 66     1 0.3064752  0.74531984
#> 67     1 0.3083544  0.74531984
#> 68     1 0.9681667  0.42891159
#> 69     1 0.3354444  0.74531984
#> 70     1 0.3016310  0.74531984
#> 71     1 0.3026159  0.74531984
#> 72     1 0.2899651  0.74531984