Get class IDs from the ConsensusPartitionList object

Get class IDs from the ConsensusPartitionList object

# S4 method for ConsensusPartitionList
get_classes(object, k)

Arguments

object
A ConsensusPartitionList-class object.

k
Number of partitions.

Details

The class IDs are inferred by merging partitions from all methods by weighting the mean silhouette scores in each method.

Value

A data frame with class IDs and other columns which are entropy of the percent membership matrix and the silhouette scores which measure the stability of a sample to stay in its group.

Examples

data(cola_rl)
get_classes(cola_rl, k = 2)

#>    class   entropy silhouette
#> 1      1 0.8204160  0.7381513
#> 2      1 0.8176682  0.7381513
#> 3      1 0.7901168  0.7381513
#> 4      1 0.7270564  0.7409586
#> 5      1 0.8317944  0.6894413
#> 6      1 0.6269981  0.7382742
#> 7      1 0.7131667  0.7381513
#> 8      1 0.8406323  0.6894413
#> 9      1 0.8078637  0.7381513
#> 10     1 0.8026041  0.7381513
#> 11     1 0.6842217  0.7381513
#> 12     1 0.8024194  0.7381513
#> 13     1 0.8329919  0.6894413
#> 14     1 0.7834641  0.7381513
#> 15     1 0.8173511  0.6894413
#> 16     1 0.8327001  0.7381513
#> 17     1 0.8118270  0.7381513
#> 18     1 0.8136441  0.7381513
#> 19     1 0.7961281  0.7381513
#> 20     1 0.8440688  0.6894413
#> 21     2 0.2565923  0.9302765
#> 22     2 0.2870352  0.9302765
#> 23     2 0.4518025  0.9302765
#> 24     2 0.2669770  0.9302765
#> 25     2 0.2863617  0.9302765
#> 26     2 0.2788491  0.9302765
#> 27     2 0.3654281  0.9302765
#> 28     2 0.2727485  0.9302765
#> 29     2 0.2592440  0.9302765
#> 30     2 0.3931537  0.9302765
#> 31     2 0.5185257  0.9302765
#> 32     2 0.3296890  0.9302765
#> 33     2 0.2630413  0.9302765
#> 34     2 0.2863954  0.9302765
#> 35     2 0.2615932  0.9302765
#> 36     2 0.4155093  0.9302765
#> 37     2 0.5345264  0.8659778
#> 38     2 0.2985292  0.9302765
#> 39     2 0.4441485  0.9302765
#> 40     2 0.3564810  0.9302765
#> 41     1 0.9658327  0.5791372
#> 42     1 0.9674642  0.5791372
#> 43     1 0.9856709  0.4985902
#> 44     1 0.7373514  0.7016921
#> 45     1 0.8597168  0.6774079
#> 46     1 0.8729511  0.6362928
#> 47     1 0.7976466  0.7016921
#> 48     1 0.9959296  0.3958816
#> 49     1 0.9857170  0.3958816
#> 50     1 0.7384490  0.7114454
#> 51     1 0.9379168  0.5791372
#> 52     1 0.5694565  0.7114454
#> 53     1 0.8236761  0.6774079
#> 54     2 0.9995360 -0.1584008
#> 55     2 0.9974478 -0.1584008
#> 56     1 0.7120604  0.7114454
#> 57     1 0.6065071  0.7114454
#> 58     1 0.8653916  0.6775447
#> 59     1 0.8663739  0.6362928
#> 60     1 0.7719622  0.7016921