Get subgroup labels

Get subgroup labels

# S4 method for DownSamplingConsensusPartition
get_classes(object, k = object@k, p_cutoff = 0.05)

Arguments

object
A DownSamplingConsensusPartition-class object.

k
Number of subgroups.

p_cutoff
Cutoff of p-values of class label prediction. It is only used when k is a vector.

Value

If k is a scalar, it returns a data frame with two columns:

the class labels

the p-value for the prediction of class labels.

If k is a vector, it returns a data frame of class labels for each k. The class label with prediction p-value > p_cutoff is set to NA.

Examples

data(golub_cola_ds)
get_classes(golub_cola_ds, k = 3)

#>           class     p
#> sample_39     3 0.000
#> sample_40     3 0.000
#> sample_42     2 0.000
#> sample_47     2 0.000
#> sample_48     2 0.000
#> sample_49     3 0.000
#> sample_41     2 0.000
#> sample_43     2 0.000
#> sample_44     2 0.000
#> sample_45     2 0.000
#> sample_46     2 0.000
#> sample_70     3 0.000
#> sample_71     2 0.000
#> sample_72     2 0.000
#> sample_68     2 0.000
#> sample_69     2 0.000
#> sample_67     2 0.000
#> sample_55     3 0.000
#> sample_56     3 0.000
#> sample_59     3 0.394
#> sample_52     1 0.000
#> sample_53     1 0.000
#> sample_51     1 0.000
#> sample_50     1 0.000
#> sample_54     1 0.000
#> sample_57     1 0.000
#> sample_58     1 0.000
#> sample_60     1 0.000
#> sample_61     1 0.000
#> sample_65     1 0.000
#> sample_66     2 0.000
#> sample_63     1 0.000
#> sample_64     1 0.000
#> sample_62     1 0.000
#> sample_1      3 0.000
#> sample_2      2 0.000
#> sample_3      3 0.000
#> sample_4      3 0.000
#> sample_5      2 0.000
#> sample_6      3 0.000
#> sample_7      3 0.000
#> sample_8      3 0.000
#> sample_9      2 0.000
#> sample_10     3 0.827
#> sample_11     2 0.000
#> sample_12     1 0.005
#> sample_13     2 0.000
#> sample_14     2 0.000
#> sample_15     2 0.000
#> sample_16     2 0.007
#> sample_17     2 0.000
#> sample_18     3 0.000
#> sample_19     2 0.000
#> sample_20     2 0.000
#> sample_21     2 0.000
#> sample_22     3 0.015
#> sample_23     3 0.000
#> sample_24     2 0.000
#> sample_25     3 0.825
#> sample_26     3 0.906
#> sample_27     3 0.000
#> sample_34     1 0.000
#> sample_35     1 0.001
#> sample_36     1 0.000
#> sample_37     1 0.000
#> sample_38     1 0.000
#> sample_28     1 0.000
#> sample_29     1 0.590
#> sample_30     1 0.000
#> sample_31     1 0.000
#> sample_32     1 0.000
#> sample_33     1 0.000

get_classes(golub_cola_ds)

#>           k=2 k=3 k=4 k=5 k=6
#> sample_39   2   3   3   3   3
#> sample_40   2   3   3   3   3
#> sample_42   2   2   2   5  NA
#> sample_47   2   2   2   2   2
#> sample_48   2   2   2   2   2
#> sample_49   2   3   3   3   3
#> sample_41   2   2   2   2   2
#> sample_43   2   2   2   2   6
#> sample_44   2   2   2   2   6
#> sample_45   2   2   2   2   6
#> sample_46   2   2   2   2   6
#> sample_70   2   3   3   3  NA
#> sample_71   2   2  NA   5   5
#> sample_72   2   2   2   5   5
#> sample_68   2   2   2   2   2
#> sample_69   2   2   2   2   2
#> sample_67  NA   2   2   5   5
#> sample_55   2   3   3   3   3
#> sample_56   2   3   3   3   3
#> sample_59   2  NA  NA  NA   1
#> sample_52   1   1   1   1   1
#> sample_53   1   1  NA   4   4
#> sample_51   1   1  NA   4   4
#> sample_50   1   1   1   4   4
#> sample_54   1   1   1   1   1
#> sample_57   1   1   1   1   1
#> sample_58   1   1   1   1   1
#> sample_60   1   1   1   1   1
#> sample_61   1   1  NA   4   4
#> sample_65   1   1   1   4  NA
#> sample_66   2   2   2   5   5
#> sample_63   1   1   1   1   1
#> sample_64   1   1  NA  NA  NA
#> sample_62   1   1   1   1   1
#> sample_1    2   3   3   3   3
#> sample_2   NA   2   2   5   5
#> sample_3    2   3   3   3  NA
#> sample_4    2   3   3   3   3
#> sample_5    2   2   2   2   2
#> sample_6    2   3   3   3  NA
#> sample_7    2   3   3   3   3
#> sample_8    2   3   3   3   3
#> sample_9    2   2   2   5   5
#> sample_10   2  NA   3   5   5
#> sample_11   2   2   2   5   5
#> sample_12  NA   1   3   4   4
#> sample_13   2   2   2   2   2
#> sample_14   2   2   2   5   5
#> sample_15   2   2   2   2   2
#> sample_16   2   2   2  NA  NA
#> sample_17   2   2   2   5   5
#> sample_18   2   3   3   3  NA
#> sample_19   2   2   2  NA  NA
#> sample_20   2   2   2   2   2
#> sample_21   2   2   2   2   2
#> sample_22  NA   3   3  NA  NA
#> sample_23   2   3   3   3   3
#> sample_24   2   2   2   2   2
#> sample_25   2  NA   3  NA  NA
#> sample_26   2  NA  NA   3   3
#> sample_27   2   3   3   3   3
#> sample_34   1   1  NA   4   4
#> sample_35   1   1   3   4   4
#> sample_36   1   1   1   4   4
#> sample_37   1   1   1   4   4
#> sample_38   1   1   3   4   4
#> sample_28   1   1   3   4   4
#> sample_29   1  NA   2   4   4
#> sample_30   1   1   1   4   4
#> sample_31   1   1  NA   4   4
#> sample_32   1   1  NA   4   4
#> sample_33   1   1  NA   4   4