Get membership matrix

Get membership matrix

# S4 method for ConsensusPartition
get_membership(object, k, each = FALSE)

Arguments

object

A ConsensusPartition-class object.

k

Number of subgroups.

each

Whether to return the percentage membership matrix which is summarized from all partitions or the individual membership in every single partition run.

Details

If each == FALSE, the value in the membership matrix is the probability to be in one subgroup, while if each == TRUE, the membership matrix contains the subgroup labels for every single partitions which are from randomly sampling from the original matrix.

The percent membership matrix is calculated by cl_consensus.

Value

If each == FALSE, it returns a membership matrix where rows correspond to the columns from the subgroups.

If each == TRUE, it returns a membership matrix where rows correspond to the columns from the original matrix.

See also

Author

Zuguang Gu <z.gu@dkfz.de>

Examples

data(golub_cola)
obj = golub_cola["ATC", "skmeans"]
get_membership(obj, k = 2)
#>             p1   p2
#> sample_39 0.00 1.00
#> sample_40 0.00 1.00
#> sample_42 1.00 0.00
#> sample_47 1.00 0.00
#> sample_48 1.00 0.00
#> sample_49 0.00 1.00
#> sample_41 1.00 0.00
#> sample_43 1.00 0.00
#> sample_44 1.00 0.00
#> sample_45 1.00 0.00
#> sample_46 1.00 0.00
#> sample_70 1.00 0.00
#> sample_71 1.00 0.00
#> sample_72 1.00 0.00
#> sample_68 1.00 0.00
#> sample_69 1.00 0.00
#> sample_67 1.00 0.00
#> sample_55 0.00 1.00
#> sample_56 0.00 1.00
#> sample_59 1.00 0.00
#> sample_52 0.00 1.00
#> sample_53 0.00 1.00
#> sample_51 0.00 1.00
#> sample_50 0.00 1.00
#> sample_54 1.00 0.00
#> sample_57 0.00 1.00
#> sample_58 0.00 1.00
#> sample_60 1.00 0.00
#> sample_61 0.00 1.00
#> sample_65 0.00 1.00
#> sample_66 1.00 0.00
#> sample_63 0.00 1.00
#> sample_64 0.00 1.00
#> sample_62 0.00 1.00
#> sample_1  0.00 1.00
#> sample_2  1.00 0.00
#> sample_3  0.00 1.00
#> sample_4  0.06 0.94
#> sample_5  1.00 0.00
#> sample_6  0.00 1.00
#> sample_7  0.00 1.00
#> sample_8  0.00 1.00
#> sample_9  1.00 0.00
#> sample_10 1.00 0.00
#> sample_11 1.00 0.00
#> sample_12 0.00 1.00
#> sample_13 1.00 0.00
#> sample_14 1.00 0.00
#> sample_15 1.00 0.00
#> sample_16 1.00 0.00
#> sample_17 1.00 0.00
#> sample_18 0.90 0.10
#> sample_19 1.00 0.00
#> sample_20 1.00 0.00
#> sample_21 1.00 0.00
#> sample_22 0.00 1.00
#> sample_23 0.00 1.00
#> sample_24 1.00 0.00
#> sample_25 0.00 1.00
#> sample_26 1.00 0.00
#> sample_27 0.00 1.00
#> sample_34 0.00 1.00
#> sample_35 0.00 1.00
#> sample_36 0.00 1.00
#> sample_37 0.00 1.00
#> sample_38 0.00 1.00
#> sample_28 0.00 1.00
#> sample_29 1.00 0.00
#> sample_30 0.00 1.00
#> sample_31 0.00 1.00
#> sample_32 0.00 1.00
#> sample_33 0.00 1.00
get_membership(obj, k = 2, each = TRUE)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#>  [2,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#>  [3,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#>  [4,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#>  [5,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#>  [6,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#>  [7,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#>  [8,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#>  [9,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [10,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [11,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [12,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [13,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [14,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [15,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [16,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [17,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [18,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [19,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [20,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [21,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [22,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [23,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [24,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [25,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [26,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [27,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [28,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [29,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [30,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [31,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [32,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [33,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [34,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [35,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [36,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [37,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [38,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [39,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [40,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [41,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [42,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [43,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [44,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [45,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [46,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [47,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [48,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [49,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [50,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [51,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [52,]    1    1    1    1    1    2    2    2    1     1     1     1     1
#> [53,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [54,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [55,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [56,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [57,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [58,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [59,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [60,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [61,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [62,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [63,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [64,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [65,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [66,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [67,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [68,]    1    1    1    1    1    1    1    1    1     1     1     1     1
#> [69,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [70,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [71,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#> [72,]    2    2    2    2    2    2    2    2    2     2     2     2     2
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [2,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [3,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [4,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [5,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [6,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [7,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [8,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [9,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [10,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [11,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [12,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [13,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [14,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [15,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [16,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [17,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [18,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [19,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [20,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [21,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [22,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [23,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [24,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [25,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [26,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [27,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [28,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [29,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [30,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [31,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [32,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [33,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [34,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [35,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [36,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [37,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [38,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [39,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [40,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [41,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [42,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [43,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [44,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [45,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [46,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [47,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [48,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [49,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [50,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [51,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [52,]     2     1     1     1     1     1     1     1     1     1     1     1
#> [53,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [54,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [55,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [56,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [57,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [58,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [59,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [60,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [61,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [62,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [63,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [64,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [65,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [66,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [67,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [68,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [69,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [70,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [71,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [72,]     2     2     2     2     2     2     2     2     2     2     2     2
#>       [,26] [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36] [,37]
#>  [1,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [2,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [3,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [4,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [5,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [6,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [7,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [8,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [9,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [10,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [11,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [12,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [13,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [14,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [15,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [16,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [17,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [18,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [19,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [20,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [21,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [22,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [23,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [24,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [25,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [26,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [27,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [28,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [29,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [30,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [31,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [32,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [33,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [34,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [35,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [36,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [37,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [38,]     2     2     2     2     2     1     2     2     2     2     2     2
#> [39,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [40,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [41,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [42,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [43,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [44,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [45,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [46,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [47,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [48,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [49,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [50,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [51,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [52,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [53,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [54,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [55,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [56,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [57,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [58,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [59,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [60,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [61,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [62,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [63,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [64,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [65,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [66,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [67,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [68,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [69,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [70,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [71,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [72,]     2     2     2     2     2     2     2     2     2     2     2     2
#>       [,38] [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46] [,47] [,48] [,49]
#>  [1,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [2,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [3,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [4,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [5,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [6,]     2     2     2     2     2     2     2     2     2     2     2     2
#>  [7,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [8,]     1     1     1     1     1     1     1     1     1     1     1     1
#>  [9,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [10,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [11,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [12,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [13,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [14,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [15,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [16,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [17,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [18,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [19,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [20,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [21,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [22,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [23,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [24,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [25,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [26,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [27,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [28,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [29,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [30,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [31,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [32,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [33,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [34,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [35,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [36,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [37,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [38,]     2     2     2     1     2     2     2     2     2     2     2     2
#> [39,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [40,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [41,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [42,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [43,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [44,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [45,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [46,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [47,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [48,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [49,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [50,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [51,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [52,]     1     1     1     1     1     1     1     1     2     1     1     1
#> [53,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [54,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [55,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [56,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [57,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [58,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [59,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [60,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [61,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [62,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [63,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [64,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [65,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [66,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [67,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [68,]     1     1     1     1     1     1     1     1     1     1     1     1
#> [69,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [70,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [71,]     2     2     2     2     2     2     2     2     2     2     2     2
#> [72,]     2     2     2     2     2     2     2     2     2     2     2     2
#>       [,50]
#>  [1,]     2
#>  [2,]     2
#>  [3,]     1
#>  [4,]     1
#>  [5,]     1
#>  [6,]     2
#>  [7,]     1
#>  [8,]     1
#>  [9,]     1
#> [10,]     1
#> [11,]     1
#> [12,]     1
#> [13,]     1
#> [14,]     1
#> [15,]     1
#> [16,]     1
#> [17,]     1
#> [18,]     2
#> [19,]     2
#> [20,]     1
#> [21,]     2
#> [22,]     2
#> [23,]     2
#> [24,]     2
#> [25,]     1
#> [26,]     2
#> [27,]     2
#> [28,]     1
#> [29,]     2
#> [30,]     2
#> [31,]     1
#> [32,]     2
#> [33,]     2
#> [34,]     2
#> [35,]     2
#> [36,]     1
#> [37,]     2
#> [38,]     1
#> [39,]     1
#> [40,]     2
#> [41,]     2
#> [42,]     2
#> [43,]     1
#> [44,]     1
#> [45,]     1
#> [46,]     2
#> [47,]     1
#> [48,]     1
#> [49,]     1
#> [50,]     1
#> [51,]     1
#> [52,]     1
#> [53,]     1
#> [54,]     1
#> [55,]     1
#> [56,]     2
#> [57,]     2
#> [58,]     1
#> [59,]     2
#> [60,]     1
#> [61,]     2
#> [62,]     2
#> [63,]     2
#> [64,]     2
#> [65,]     2
#> [66,]     2
#> [67,]     2
#> [68,]     1
#> [69,]     2
#> [70,]     2
#> [71,]     2
#> [72,]     2