Get membership matrix

Get membership matrix

# S4 method for ConsensusPartitionList
get_membership(object, k)

Arguments

object
A ConsensusPartitionList-class object.

k
Number of partitions.

Details

The membership matrix (the probability of each sample to be in one group, if assuming columns represent samples) is inferred from the consensus partition of every combination of methods, weighted by the mean silhouette score of the partition for each method. So methods which give instable partitions have lower weights when summarizing membership matrix from all methods.

Value

A membership matrix where rows correspond to the columns in the original matrix.

See also

Examples

data(cola_rl)
get_membership(cola_rl, k = 2)

#>               p1        p2
#>  [1,] 0.74415208 0.2558479
#>  [2,] 0.74592820 0.2540718
#>  [3,] 0.76294007 0.2370599
#>  [4,] 0.79742091 0.2025791
#>  [5,] 0.73662659 0.2633734
#>  [6,] 0.84303544 0.1569646
#>  [7,] 0.80433803 0.1956620
#>  [8,] 0.73057918 0.2694208
#>  [9,] 0.75214306 0.2478569
#> [10,] 0.75540155 0.2445984
#> [11,] 0.81809734 0.1819027
#> [12,] 0.75551505 0.2444850
#> [13,] 0.73581791 0.2641821
#> [14,] 0.76684885 0.2331511
#> [15,] 0.74613219 0.2538678
#> [16,] 0.73601526 0.2639847
#> [17,] 0.74965338 0.2503466
#> [18,] 0.74850178 0.2514982
#> [19,] 0.75934485 0.2406551
#> [20,] 0.72817674 0.2718233
#> [21,] 0.04315879 0.9568412
#> [22,] 0.05015033 0.9498497
#> [23,] 0.09464973 0.9053503
#> [24,] 0.04550283 0.9544972
#> [25,] 0.04999170 0.9500083
#> [26,] 0.04823444 0.9517656
#> [27,] 0.06986727 0.9301327
#> [28,] 0.04682383 0.9531762
#> [29,] 0.04375333 0.9562467
#> [30,] 0.07745671 0.9225433
#> [31,] 0.11626401 0.8837360
#> [32,] 0.06056634 0.9394337
#> [33,] 0.04460949 0.9553905
#> [34,] 0.04999963 0.9500004
#> [35,] 0.04428233 0.9557177
#> [36,] 0.08382322 0.9161768
#> [37,] 0.12180476 0.8781952
#> [38,] 0.05288527 0.9471147
#> [39,] 0.09231451 0.9076855
#> [40,] 0.06748861 0.9325114
#> [41,] 0.60838684 0.3916132
#> [42,] 0.60578767 0.3942123
#> [43,] 0.57035365 0.4296464
#> [44,] 0.79215081 0.2078492
#> [45,] 0.71684303 0.2831570
#> [46,] 0.70669598 0.2933040
#> [47,] 0.75842682 0.2415732
#> [48,] 0.53754164 0.4624584
#> [49,] 0.57024059 0.4297594
#> [50,] 0.79158144 0.2084186
#> [51,] 0.64562264 0.3543774
#> [52,] 0.86557260 0.1344274
#> [53,] 0.74202445 0.2579756
#> [54,] 0.48731946 0.5126805
#> [55,] 0.47026793 0.5297321
#> [56,] 0.80487983 0.1951202
#> [57,] 0.85132795 0.1486720
#> [58,] 0.71255977 0.2874402
#> [59,] 0.71180830 0.2881917
#> [60,] 0.77344226 0.2265577