Get parameters

Get parameters

# S4 method for ConsensusPartition
get_param(object, k = object@k, unique = TRUE)

Arguments

object
A ConsensusPartition-class object.

k
Number of subgroups.

unique
Whether to apply unique to rows of the returned data frame.

Details

It is mainly used internally.

Value

A data frame of parameters corresponding to the current k. In the data frame, each row corresponds to a partition run.

Examples

data(golub_cola)
obj = golub_cola["ATC", "skmeans"]
get_param(obj)

#>    top_n k n_row n_col
#> 1    412 2   330    72
#> 2    824 2   659    72
#> 3   1235 2   988    72
#> 4    412 3   330    72
#> 5    824 3   659    72
#> 6   1235 3   988    72
#> 7    412 4   330    72
#> 8    824 4   659    72
#> 9   1235 4   988    72
#> 10   412 5   330    72
#> 11   824 5   659    72
#> 12  1235 5   988    72
#> 13   412 6   330    72
#> 14   824 6   659    72
#> 15  1235 6   988    72

get_param(obj, k = 2)

#>   top_n k n_row n_col
#> 1   412 2   330    72
#> 2   824 2   659    72
#> 3  1235 2   988    72

get_param(obj, unique = FALSE)

#>     top_n k n_row n_col
#> 1     412 2   330    72
#> 2     412 2   330    72
#> 3     412 2   330    72
#> 4     412 2   330    72
#> 5     412 2   330    72
#> 6     412 2   330    72
#> 7     412 2   330    72
#> 8     412 2   330    72
#> 9     412 2   330    72
#> 10    412 2   330    72
#> 11    412 2   330    72
#> 12    412 2   330    72
#> 13    412 2   330    72
#> 14    412 2   330    72
#> 15    412 2   330    72
#> 16    412 2   330    72
#> 17    412 2   330    72
#> 18    412 2   330    72
#> 19    412 2   330    72
#> 20    412 2   330    72
#> 21    412 2   330    72
#> 22    412 2   330    72
#> 23    412 2   330    72
#> 24    412 2   330    72
#> 25    412 2   330    72
#> 26    412 2   330    72
#> 27    412 2   330    72
#> 28    412 2   330    72
#> 29    412 2   330    72
#> 30    412 2   330    72
#> 31    412 2   330    72
#> 32    412 2   330    72
#> 33    412 2   330    72
#> 34    412 2   330    72
#> 35    412 2   330    72
#> 36    412 2   330    72
#> 37    412 2   330    72
#> 38    412 2   330    72
#> 39    412 2   330    72
#> 40    412 2   330    72
#> 41    412 2   330    72
#> 42    412 2   330    72
#> 43    412 2   330    72
#> 44    412 2   330    72
#> 45    412 2   330    72
#> 46    412 2   330    72
#> 47    412 2   330    72
#> 48    412 2   330    72
#> 49    412 2   330    72
#> 50    412 2   330    72
#> 51    824 2   659    72
#> 52    824 2   659    72
#> 53    824 2   659    72
#> 54    824 2   659    72
#> 55    824 2   659    72
#> 56    824 2   659    72
#> 57    824 2   659    72
#> 58    824 2   659    72
#> 59    824 2   659    72
#> 60    824 2   659    72
#> 61    824 2   659    72
#> 62    824 2   659    72
#> 63    824 2   659    72
#> 64    824 2   659    72
#> 65    824 2   659    72
#> 66    824 2   659    72
#> 67    824 2   659    72
#> 68    824 2   659    72
#> 69    824 2   659    72
#> 70    824 2   659    72
#> 71    824 2   659    72
#> 72    824 2   659    72
#> 73    824 2   659    72
#> 74    824 2   659    72
#> 75    824 2   659    72
#> 76    824 2   659    72
#> 77    824 2   659    72
#> 78    824 2   659    72
#> 79    824 2   659    72
#> 80    824 2   659    72
#> 81    824 2   659    72
#> 82    824 2   659    72
#> 83    824 2   659    72
#> 84    824 2   659    72
#> 85    824 2   659    72
#> 86    824 2   659    72
#> 87    824 2   659    72
#> 88    824 2   659    72
#> 89    824 2   659    72
#> 90    824 2   659    72
#> 91    824 2   659    72
#> 92    824 2   659    72
#> 93    824 2   659    72
#> 94    824 2   659    72
#> 95    824 2   659    72
#> 96    824 2   659    72
#> 97    824 2   659    72
#> 98    824 2   659    72
#> 99    824 2   659    72
#> 100   824 2   659    72
#> 101  1235 2   988    72
#> 102  1235 2   988    72
#> 103  1235 2   988    72
#> 104  1235 2   988    72
#> 105  1235 2   988    72
#> 106  1235 2   988    72
#> 107  1235 2   988    72
#> 108  1235 2   988    72
#> 109  1235 2   988    72
#> 110  1235 2   988    72
#> 111  1235 2   988    72
#> 112  1235 2   988    72
#> 113  1235 2   988    72
#> 114  1235 2   988    72
#> 115  1235 2   988    72
#> 116  1235 2   988    72
#> 117  1235 2   988    72
#> 118  1235 2   988    72
#> 119  1235 2   988    72
#> 120  1235 2   988    72
#> 121  1235 2   988    72
#> 122  1235 2   988    72
#> 123  1235 2   988    72
#> 124  1235 2   988    72
#> 125  1235 2   988    72
#> 126  1235 2   988    72
#> 127  1235 2   988    72
#> 128  1235 2   988    72
#> 129  1235 2   988    72
#> 130  1235 2   988    72
#> 131  1235 2   988    72
#> 132  1235 2   988    72
#> 133  1235 2   988    72
#> 134  1235 2   988    72
#> 135  1235 2   988    72
#> 136  1235 2   988    72
#> 137  1235 2   988    72
#> 138  1235 2   988    72
#> 139  1235 2   988    72
#> 140  1235 2   988    72
#> 141  1235 2   988    72
#> 142  1235 2   988    72
#> 143  1235 2   988    72
#> 144  1235 2   988    72
#> 145  1235 2   988    72
#> 146  1235 2   988    72
#> 147  1235 2   988    72
#> 148  1235 2   988    72
#> 149  1235 2   988    72
#> 150  1235 2   988    72
#> 151   412 3   330    72
#> 152   412 3   330    72
#> 153   412 3   330    72
#> 154   412 3   330    72
#> 155   412 3   330    72
#> 156   412 3   330    72
#> 157   412 3   330    72
#> 158   412 3   330    72
#> 159   412 3   330    72
#> 160   412 3   330    72
#> 161   412 3   330    72
#> 162   412 3   330    72
#> 163   412 3   330    72
#> 164   412 3   330    72
#> 165   412 3   330    72
#> 166   412 3   330    72
#> 167   412 3   330    72
#> 168   412 3   330    72
#> 169   412 3   330    72
#> 170   412 3   330    72
#> 171   412 3   330    72
#> 172   412 3   330    72
#> 173   412 3   330    72
#> 174   412 3   330    72
#> 175   412 3   330    72
#> 176   412 3   330    72
#> 177   412 3   330    72
#> 178   412 3   330    72
#> 179   412 3   330    72
#> 180   412 3   330    72
#> 181   412 3   330    72
#> 182   412 3   330    72
#> 183   412 3   330    72
#> 184   412 3   330    72
#> 185   412 3   330    72
#> 186   412 3   330    72
#> 187   412 3   330    72
#> 188   412 3   330    72
#> 189   412 3   330    72
#> 190   412 3   330    72
#> 191   412 3   330    72
#> 192   412 3   330    72
#> 193   412 3   330    72
#> 194   412 3   330    72
#> 195   412 3   330    72
#> 196   412 3   330    72
#> 197   412 3   330    72
#> 198   412 3   330    72
#> 199   412 3   330    72
#> 200   412 3   330    72
#> 201   824 3   659    72
#> 202   824 3   659    72
#> 203   824 3   659    72
#> 204   824 3   659    72
#> 205   824 3   659    72
#> 206   824 3   659    72
#> 207   824 3   659    72
#> 208   824 3   659    72
#> 209   824 3   659    72
#> 210   824 3   659    72
#> 211   824 3   659    72
#> 212   824 3   659    72
#> 213   824 3   659    72
#> 214   824 3   659    72
#> 215   824 3   659    72
#> 216   824 3   659    72
#> 217   824 3   659    72
#> 218   824 3   659    72
#> 219   824 3   659    72
#> 220   824 3   659    72
#> 221   824 3   659    72
#> 222   824 3   659    72
#> 223   824 3   659    72
#> 224   824 3   659    72
#> 225   824 3   659    72
#> 226   824 3   659    72
#> 227   824 3   659    72
#> 228   824 3   659    72
#> 229   824 3   659    72
#> 230   824 3   659    72
#> 231   824 3   659    72
#> 232   824 3   659    72
#> 233   824 3   659    72
#> 234   824 3   659    72
#> 235   824 3   659    72
#> 236   824 3   659    72
#> 237   824 3   659    72
#> 238   824 3   659    72
#> 239   824 3   659    72
#> 240   824 3   659    72
#> 241   824 3   659    72
#> 242   824 3   659    72
#> 243   824 3   659    72
#> 244   824 3   659    72
#> 245   824 3   659    72
#> 246   824 3   659    72
#> 247   824 3   659    72
#> 248   824 3   659    72
#> 249   824 3   659    72
#> 250   824 3   659    72
#> 251  1235 3   988    72
#> 252  1235 3   988    72
#> 253  1235 3   988    72
#> 254  1235 3   988    72
#> 255  1235 3   988    72
#> 256  1235 3   988    72
#> 257  1235 3   988    72
#> 258  1235 3   988    72
#> 259  1235 3   988    72
#> 260  1235 3   988    72
#> 261  1235 3   988    72
#> 262  1235 3   988    72
#> 263  1235 3   988    72
#> 264  1235 3   988    72
#> 265  1235 3   988    72
#> 266  1235 3   988    72
#> 267  1235 3   988    72
#> 268  1235 3   988    72
#> 269  1235 3   988    72
#> 270  1235 3   988    72
#> 271  1235 3   988    72
#> 272  1235 3   988    72
#> 273  1235 3   988    72
#> 274  1235 3   988    72
#> 275  1235 3   988    72
#> 276  1235 3   988    72
#> 277  1235 3   988    72
#> 278  1235 3   988    72
#> 279  1235 3   988    72
#> 280  1235 3   988    72
#> 281  1235 3   988    72
#> 282  1235 3   988    72
#> 283  1235 3   988    72
#> 284  1235 3   988    72
#> 285  1235 3   988    72
#> 286  1235 3   988    72
#> 287  1235 3   988    72
#> 288  1235 3   988    72
#> 289  1235 3   988    72
#> 290  1235 3   988    72
#> 291  1235 3   988    72
#> 292  1235 3   988    72
#> 293  1235 3   988    72
#> 294  1235 3   988    72
#> 295  1235 3   988    72
#> 296  1235 3   988    72
#> 297  1235 3   988    72
#> 298  1235 3   988    72
#> 299  1235 3   988    72
#> 300  1235 3   988    72
#> 301   412 4   330    72
#> 302   412 4   330    72
#> 303   412 4   330    72
#> 304   412 4   330    72
#> 305   412 4   330    72
#> 306   412 4   330    72
#> 307   412 4   330    72
#> 308   412 4   330    72
#> 309   412 4   330    72
#> 310   412 4   330    72
#> 311   412 4   330    72
#> 312   412 4   330    72
#> 313   412 4   330    72
#> 314   412 4   330    72
#> 315   412 4   330    72
#> 316   412 4   330    72
#> 317   412 4   330    72
#> 318   412 4   330    72
#> 319   412 4   330    72
#> 320   412 4   330    72
#> 321   412 4   330    72
#> 322   412 4   330    72
#> 323   412 4   330    72
#> 324   412 4   330    72
#> 325   412 4   330    72
#> 326   412 4   330    72
#> 327   412 4   330    72
#> 328   412 4   330    72
#> 329   412 4   330    72
#> 330   412 4   330    72
#> 331   412 4   330    72
#> 332   412 4   330    72
#> 333   412 4   330    72
#> 334   412 4   330    72
#> 335   412 4   330    72
#> 336   412 4   330    72
#> 337   412 4   330    72
#> 338   412 4   330    72
#> 339   412 4   330    72
#> 340   412 4   330    72
#> 341   412 4   330    72
#> 342   412 4   330    72
#> 343   412 4   330    72
#> 344   412 4   330    72
#> 345   412 4   330    72
#> 346   412 4   330    72
#> 347   412 4   330    72
#> 348   412 4   330    72
#> 349   412 4   330    72
#> 350   412 4   330    72
#> 351   824 4   659    72
#> 352   824 4   659    72
#> 353   824 4   659    72
#> 354   824 4   659    72
#> 355   824 4   659    72
#> 356   824 4   659    72
#> 357   824 4   659    72
#> 358   824 4   659    72
#> 359   824 4   659    72
#> 360   824 4   659    72
#> 361   824 4   659    72
#> 362   824 4   659    72
#> 363   824 4   659    72
#> 364   824 4   659    72
#> 365   824 4   659    72
#> 366   824 4   659    72
#> 367   824 4   659    72
#> 368   824 4   659    72
#> 369   824 4   659    72
#> 370   824 4   659    72
#> 371   824 4   659    72
#> 372   824 4   659    72
#> 373   824 4   659    72
#> 374   824 4   659    72
#> 375   824 4   659    72
#> 376   824 4   659    72
#> 377   824 4   659    72
#> 378   824 4   659    72
#> 379   824 4   659    72
#> 380   824 4   659    72
#> 381   824 4   659    72
#> 382   824 4   659    72
#> 383   824 4   659    72
#> 384   824 4   659    72
#> 385   824 4   659    72
#> 386   824 4   659    72
#> 387   824 4   659    72
#> 388   824 4   659    72
#> 389   824 4   659    72
#> 390   824 4   659    72
#> 391   824 4   659    72
#> 392   824 4   659    72
#> 393   824 4   659    72
#> 394   824 4   659    72
#> 395   824 4   659    72
#> 396   824 4   659    72
#> 397   824 4   659    72
#> 398   824 4   659    72
#> 399   824 4   659    72
#> 400   824 4   659    72
#> 401  1235 4   988    72
#> 402  1235 4   988    72
#> 403  1235 4   988    72
#> 404  1235 4   988    72
#> 405  1235 4   988    72
#> 406  1235 4   988    72
#> 407  1235 4   988    72
#> 408  1235 4   988    72
#> 409  1235 4   988    72
#> 410  1235 4   988    72
#> 411  1235 4   988    72
#> 412  1235 4   988    72
#> 413  1235 4   988    72
#> 414  1235 4   988    72
#> 415  1235 4   988    72
#> 416  1235 4   988    72
#> 417  1235 4   988    72
#> 418  1235 4   988    72
#> 419  1235 4   988    72
#> 420  1235 4   988    72
#> 421  1235 4   988    72
#> 422  1235 4   988    72
#> 423  1235 4   988    72
#> 424  1235 4   988    72
#> 425  1235 4   988    72
#> 426  1235 4   988    72
#> 427  1235 4   988    72
#> 428  1235 4   988    72
#> 429  1235 4   988    72
#> 430  1235 4   988    72
#> 431  1235 4   988    72
#> 432  1235 4   988    72
#> 433  1235 4   988    72
#> 434  1235 4   988    72
#> 435  1235 4   988    72
#> 436  1235 4   988    72
#> 437  1235 4   988    72
#> 438  1235 4   988    72
#> 439  1235 4   988    72
#> 440  1235 4   988    72
#> 441  1235 4   988    72
#> 442  1235 4   988    72
#> 443  1235 4   988    72
#> 444  1235 4   988    72
#> 445  1235 4   988    72
#> 446  1235 4   988    72
#> 447  1235 4   988    72
#> 448  1235 4   988    72
#> 449  1235 4   988    72
#> 450  1235 4   988    72
#> 451   412 5   330    72
#> 452   412 5   330    72
#> 453   412 5   330    72
#> 454   412 5   330    72
#> 455   412 5   330    72
#> 456   412 5   330    72
#> 457   412 5   330    72
#> 458   412 5   330    72
#> 459   412 5   330    72
#> 460   412 5   330    72
#> 461   412 5   330    72
#> 462   412 5   330    72
#> 463   412 5   330    72
#> 464   412 5   330    72
#> 465   412 5   330    72
#> 466   412 5   330    72
#> 467   412 5   330    72
#> 468   412 5   330    72
#> 469   412 5   330    72
#> 470   412 5   330    72
#> 471   412 5   330    72
#> 472   412 5   330    72
#> 473   412 5   330    72
#> 474   412 5   330    72
#> 475   412 5   330    72
#> 476   412 5   330    72
#> 477   412 5   330    72
#> 478   412 5   330    72
#> 479   412 5   330    72
#> 480   412 5   330    72
#> 481   412 5   330    72
#> 482   412 5   330    72
#> 483   412 5   330    72
#> 484   412 5   330    72
#> 485   412 5   330    72
#> 486   412 5   330    72
#> 487   412 5   330    72
#> 488   412 5   330    72
#> 489   412 5   330    72
#> 490   412 5   330    72
#> 491   412 5   330    72
#> 492   412 5   330    72
#> 493   412 5   330    72
#> 494   412 5   330    72
#> 495   412 5   330    72
#> 496   412 5   330    72
#> 497   412 5   330    72
#> 498   412 5   330    72
#> 499   412 5   330    72
#> 500   412 5   330    72
#> 501   824 5   659    72
#> 502   824 5   659    72
#> 503   824 5   659    72
#> 504   824 5   659    72
#> 505   824 5   659    72
#> 506   824 5   659    72
#> 507   824 5   659    72
#> 508   824 5   659    72
#> 509   824 5   659    72
#> 510   824 5   659    72
#> 511   824 5   659    72
#> 512   824 5   659    72
#> 513   824 5   659    72
#> 514   824 5   659    72
#> 515   824 5   659    72
#> 516   824 5   659    72
#> 517   824 5   659    72
#> 518   824 5   659    72
#> 519   824 5   659    72
#> 520   824 5   659    72
#> 521   824 5   659    72
#> 522   824 5   659    72
#> 523   824 5   659    72
#> 524   824 5   659    72
#> 525   824 5   659    72
#> 526   824 5   659    72
#> 527   824 5   659    72
#> 528   824 5   659    72
#> 529   824 5   659    72
#> 530   824 5   659    72
#> 531   824 5   659    72
#> 532   824 5   659    72
#> 533   824 5   659    72
#> 534   824 5   659    72
#> 535   824 5   659    72
#> 536   824 5   659    72
#> 537   824 5   659    72
#> 538   824 5   659    72
#> 539   824 5   659    72
#> 540   824 5   659    72
#> 541   824 5   659    72
#> 542   824 5   659    72
#> 543   824 5   659    72
#> 544   824 5   659    72
#> 545   824 5   659    72
#> 546   824 5   659    72
#> 547   824 5   659    72
#> 548   824 5   659    72
#> 549   824 5   659    72
#> 550   824 5   659    72
#> 551  1235 5   988    72
#> 552  1235 5   988    72
#> 553  1235 5   988    72
#> 554  1235 5   988    72
#> 555  1235 5   988    72
#> 556  1235 5   988    72
#> 557  1235 5   988    72
#> 558  1235 5   988    72
#> 559  1235 5   988    72
#> 560  1235 5   988    72
#> 561  1235 5   988    72
#> 562  1235 5   988    72
#> 563  1235 5   988    72
#> 564  1235 5   988    72
#> 565  1235 5   988    72
#> 566  1235 5   988    72
#> 567  1235 5   988    72
#> 568  1235 5   988    72
#> 569  1235 5   988    72
#> 570  1235 5   988    72
#> 571  1235 5   988    72
#> 572  1235 5   988    72
#> 573  1235 5   988    72
#> 574  1235 5   988    72
#> 575  1235 5   988    72
#> 576  1235 5   988    72
#> 577  1235 5   988    72
#> 578  1235 5   988    72
#> 579  1235 5   988    72
#> 580  1235 5   988    72
#> 581  1235 5   988    72
#> 582  1235 5   988    72
#> 583  1235 5   988    72
#> 584  1235 5   988    72
#> 585  1235 5   988    72
#> 586  1235 5   988    72
#> 587  1235 5   988    72
#> 588  1235 5   988    72
#> 589  1235 5   988    72
#> 590  1235 5   988    72
#> 591  1235 5   988    72
#> 592  1235 5   988    72
#> 593  1235 5   988    72
#> 594  1235 5   988    72
#> 595  1235 5   988    72
#> 596  1235 5   988    72
#> 597  1235 5   988    72
#> 598  1235 5   988    72
#> 599  1235 5   988    72
#> 600  1235 5   988    72
#> 601   412 6   330    72
#> 602   412 6   330    72
#> 603   412 6   330    72
#> 604   412 6   330    72
#> 605   412 6   330    72
#> 606   412 6   330    72
#> 607   412 6   330    72
#> 608   412 6   330    72
#> 609   412 6   330    72
#> 610   412 6   330    72
#> 611   412 6   330    72
#> 612   412 6   330    72
#> 613   412 6   330    72
#> 614   412 6   330    72
#> 615   412 6   330    72
#> 616   412 6   330    72
#> 617   412 6   330    72
#> 618   412 6   330    72
#> 619   412 6   330    72
#> 620   412 6   330    72
#> 621   412 6   330    72
#> 622   412 6   330    72
#> 623   412 6   330    72
#> 624   412 6   330    72
#> 625   412 6   330    72
#> 626   412 6   330    72
#> 627   412 6   330    72
#> 628   412 6   330    72
#> 629   412 6   330    72
#> 630   412 6   330    72
#> 631   412 6   330    72
#> 632   412 6   330    72
#> 633   412 6   330    72
#> 634   412 6   330    72
#> 635   412 6   330    72
#> 636   412 6   330    72
#> 637   412 6   330    72
#> 638   412 6   330    72
#> 639   412 6   330    72
#> 640   412 6   330    72
#> 641   412 6   330    72
#> 642   412 6   330    72
#> 643   412 6   330    72
#> 644   412 6   330    72
#> 645   412 6   330    72
#> 646   412 6   330    72
#> 647   412 6   330    72
#> 648   412 6   330    72
#> 649   412 6   330    72
#> 650   412 6   330    72
#> 651   824 6   659    72
#> 652   824 6   659    72
#> 653   824 6   659    72
#> 654   824 6   659    72
#> 655   824 6   659    72
#> 656   824 6   659    72
#> 657   824 6   659    72
#> 658   824 6   659    72
#> 659   824 6   659    72
#> 660   824 6   659    72
#> 661   824 6   659    72
#> 662   824 6   659    72
#> 663   824 6   659    72
#> 664   824 6   659    72
#> 665   824 6   659    72
#> 666   824 6   659    72
#> 667   824 6   659    72
#> 668   824 6   659    72
#> 669   824 6   659    72
#> 670   824 6   659    72
#> 671   824 6   659    72
#> 672   824 6   659    72
#> 673   824 6   659    72
#> 674   824 6   659    72
#> 675   824 6   659    72
#> 676   824 6   659    72
#> 677   824 6   659    72
#> 678   824 6   659    72
#> 679   824 6   659    72
#> 680   824 6   659    72
#> 681   824 6   659    72
#> 682   824 6   659    72
#> 683   824 6   659    72
#> 684   824 6   659    72
#> 685   824 6   659    72
#> 686   824 6   659    72
#> 687   824 6   659    72
#> 688   824 6   659    72
#> 689   824 6   659    72
#> 690   824 6   659    72
#> 691   824 6   659    72
#> 692   824 6   659    72
#> 693   824 6   659    72
#> 694   824 6   659    72
#> 695   824 6   659    72
#> 696   824 6   659    72
#> 697   824 6   659    72
#> 698   824 6   659    72
#> 699   824 6   659    72
#> 700   824 6   659    72
#> 701  1235 6   988    72
#> 702  1235 6   988    72
#> 703  1235 6   988    72
#> 704  1235 6   988    72
#> 705  1235 6   988    72
#> 706  1235 6   988    72
#> 707  1235 6   988    72
#> 708  1235 6   988    72
#> 709  1235 6   988    72
#> 710  1235 6   988    72
#> 711  1235 6   988    72
#> 712  1235 6   988    72
#> 713  1235 6   988    72
#> 714  1235 6   988    72
#> 715  1235 6   988    72
#> 716  1235 6   988    72
#> 717  1235 6   988    72
#> 718  1235 6   988    72
#> 719  1235 6   988    72
#> 720  1235 6   988    72
#> 721  1235 6   988    72
#> 722  1235 6   988    72
#> 723  1235 6   988    72
#> 724  1235 6   988    72
#> 725  1235 6   988    72
#> 726  1235 6   988    72
#> 727  1235 6   988    72
#> 728  1235 6   988    72
#> 729  1235 6   988    72
#> 730  1235 6   988    72
#> 731  1235 6   988    72
#> 732  1235 6   988    72
#> 733  1235 6   988    72
#> 734  1235 6   988    72
#> 735  1235 6   988    72
#> 736  1235 6   988    72
#> 737  1235 6   988    72
#> 738  1235 6   988    72
#> 739  1235 6   988    72
#> 740  1235 6   988    72
#> 741  1235 6   988    72
#> 742  1235 6   988    72
#> 743  1235 6   988    72
#> 744  1235 6   988    72
#> 745  1235 6   988    72
#> 746  1235 6   988    72
#> 747  1235 6   988    72
#> 748  1235 6   988    72
#> 749  1235 6   988    72
#> 750  1235 6   988    72