Get parameters

Get parameters

# S4 method for ConsensusPartition
get_param(object, k = object@k, unique = TRUE)

Arguments

object
A ConsensusPartition-class object.

k
Number of subgroups.

unique
Whether to apply unique to rows of the returned data frame.

Details

It is mainly used internally.

Value

A data frame of parameters corresponding to the current k. In the data frame, each row corresponds to a partition run.

Examples

data(golub_cola)
obj = golub_cola["ATC", "skmeans"]
get_param(obj)

#>    top_n k n_row n_col
#> 1    412 2   330    72
#> 2    824 2   659    72
#> 3   1235 2   988    72
#> 4   1646 2  1317    72
#> 5   2058 2  1646    72
#> 6    412 3   330    72
#> 7    824 3   659    72
#> 8   1235 3   988    72
#> 9   1646 3  1317    72
#> 10  2058 3  1646    72
#> 11   412 4   330    72
#> 12   824 4   659    72
#> 13  1235 4   988    72
#> 14  1646 4  1317    72
#> 15  2058 4  1646    72
#> 16   412 5   330    72
#> 17   824 5   659    72
#> 18  1235 5   988    72
#> 19  1646 5  1317    72
#> 20  2058 5  1646    72
#> 21   412 6   330    72
#> 22   824 6   659    72
#> 23  1235 6   988    72
#> 24  1646 6  1317    72
#> 25  2058 6  1646    72

get_param(obj, k = 2)

#>   top_n k n_row n_col
#> 1   412 2   330    72
#> 2   824 2   659    72
#> 3  1235 2   988    72
#> 4  1646 2  1317    72
#> 5  2058 2  1646    72

get_param(obj, unique = FALSE)

#>      top_n k n_row n_col
#> 1      412 2   330    72
#> 2      412 2   330    72
#> 3      412 2   330    72
#> 4      412 2   330    72
#> 5      412 2   330    72
#> 6      412 2   330    72
#> 7      412 2   330    72
#> 8      412 2   330    72
#> 9      412 2   330    72
#> 10     412 2   330    72
#> 11     412 2   330    72
#> 12     412 2   330    72
#> 13     412 2   330    72
#> 14     412 2   330    72
#> 15     412 2   330    72
#> 16     412 2   330    72
#> 17     412 2   330    72
#> 18     412 2   330    72
#> 19     412 2   330    72
#> 20     412 2   330    72
#> 21     412 2   330    72
#> 22     412 2   330    72
#> 23     412 2   330    72
#> 24     412 2   330    72
#> 25     412 2   330    72
#> 26     412 2   330    72
#> 27     412 2   330    72
#> 28     412 2   330    72
#> 29     412 2   330    72
#> 30     412 2   330    72
#> 31     412 2   330    72
#> 32     412 2   330    72
#> 33     412 2   330    72
#> 34     412 2   330    72
#> 35     412 2   330    72
#> 36     412 2   330    72
#> 37     412 2   330    72
#> 38     412 2   330    72
#> 39     412 2   330    72
#> 40     412 2   330    72
#> 41     412 2   330    72
#> 42     412 2   330    72
#> 43     412 2   330    72
#> 44     412 2   330    72
#> 45     412 2   330    72
#> 46     412 2   330    72
#> 47     412 2   330    72
#> 48     412 2   330    72
#> 49     412 2   330    72
#> 50     412 2   330    72
#> 51     824 2   659    72
#> 52     824 2   659    72
#> 53     824 2   659    72
#> 54     824 2   659    72
#> 55     824 2   659    72
#> 56     824 2   659    72
#> 57     824 2   659    72
#> 58     824 2   659    72
#> 59     824 2   659    72
#> 60     824 2   659    72
#> 61     824 2   659    72
#> 62     824 2   659    72
#> 63     824 2   659    72
#> 64     824 2   659    72
#> 65     824 2   659    72
#> 66     824 2   659    72
#> 67     824 2   659    72
#> 68     824 2   659    72
#> 69     824 2   659    72
#> 70     824 2   659    72
#> 71     824 2   659    72
#> 72     824 2   659    72
#> 73     824 2   659    72
#> 74     824 2   659    72
#> 75     824 2   659    72
#> 76     824 2   659    72
#> 77     824 2   659    72
#> 78     824 2   659    72
#> 79     824 2   659    72
#> 80     824 2   659    72
#> 81     824 2   659    72
#> 82     824 2   659    72
#> 83     824 2   659    72
#> 84     824 2   659    72
#> 85     824 2   659    72
#> 86     824 2   659    72
#> 87     824 2   659    72
#> 88     824 2   659    72
#> 89     824 2   659    72
#> 90     824 2   659    72
#> 91     824 2   659    72
#> 92     824 2   659    72
#> 93     824 2   659    72
#> 94     824 2   659    72
#> 95     824 2   659    72
#> 96     824 2   659    72
#> 97     824 2   659    72
#> 98     824 2   659    72
#> 99     824 2   659    72
#> 100    824 2   659    72
#> 101   1235 2   988    72
#> 102   1235 2   988    72
#> 103   1235 2   988    72
#> 104   1235 2   988    72
#> 105   1235 2   988    72
#> 106   1235 2   988    72
#> 107   1235 2   988    72
#> 108   1235 2   988    72
#> 109   1235 2   988    72
#> 110   1235 2   988    72
#> 111   1235 2   988    72
#> 112   1235 2   988    72
#> 113   1235 2   988    72
#> 114   1235 2   988    72
#> 115   1235 2   988    72
#> 116   1235 2   988    72
#> 117   1235 2   988    72
#> 118   1235 2   988    72
#> 119   1235 2   988    72
#> 120   1235 2   988    72
#> 121   1235 2   988    72
#> 122   1235 2   988    72
#> 123   1235 2   988    72
#> 124   1235 2   988    72
#> 125   1235 2   988    72
#> 126   1235 2   988    72
#> 127   1235 2   988    72
#> 128   1235 2   988    72
#> 129   1235 2   988    72
#> 130   1235 2   988    72
#> 131   1235 2   988    72
#> 132   1235 2   988    72
#> 133   1235 2   988    72
#> 134   1235 2   988    72
#> 135   1235 2   988    72
#> 136   1235 2   988    72
#> 137   1235 2   988    72
#> 138   1235 2   988    72
#> 139   1235 2   988    72
#> 140   1235 2   988    72
#> 141   1235 2   988    72
#> 142   1235 2   988    72
#> 143   1235 2   988    72
#> 144   1235 2   988    72
#> 145   1235 2   988    72
#> 146   1235 2   988    72
#> 147   1235 2   988    72
#> 148   1235 2   988    72
#> 149   1235 2   988    72
#> 150   1235 2   988    72
#> 151   1646 2  1317    72
#> 152   1646 2  1317    72
#> 153   1646 2  1317    72
#> 154   1646 2  1317    72
#> 155   1646 2  1317    72
#> 156   1646 2  1317    72
#> 157   1646 2  1317    72
#> 158   1646 2  1317    72
#> 159   1646 2  1317    72
#> 160   1646 2  1317    72
#> 161   1646 2  1317    72
#> 162   1646 2  1317    72
#> 163   1646 2  1317    72
#> 164   1646 2  1317    72
#> 165   1646 2  1317    72
#> 166   1646 2  1317    72
#> 167   1646 2  1317    72
#> 168   1646 2  1317    72
#> 169   1646 2  1317    72
#> 170   1646 2  1317    72
#> 171   1646 2  1317    72
#> 172   1646 2  1317    72
#> 173   1646 2  1317    72
#> 174   1646 2  1317    72
#> 175   1646 2  1317    72
#> 176   1646 2  1317    72
#> 177   1646 2  1317    72
#> 178   1646 2  1317    72
#> 179   1646 2  1317    72
#> 180   1646 2  1317    72
#> 181   1646 2  1317    72
#> 182   1646 2  1317    72
#> 183   1646 2  1317    72
#> 184   1646 2  1317    72
#> 185   1646 2  1317    72
#> 186   1646 2  1317    72
#> 187   1646 2  1317    72
#> 188   1646 2  1317    72
#> 189   1646 2  1317    72
#> 190   1646 2  1317    72
#> 191   1646 2  1317    72
#> 192   1646 2  1317    72
#> 193   1646 2  1317    72
#> 194   1646 2  1317    72
#> 195   1646 2  1317    72
#> 196   1646 2  1317    72
#> 197   1646 2  1317    72
#> 198   1646 2  1317    72
#> 199   1646 2  1317    72
#> 200   1646 2  1317    72
#> 201   2058 2  1646    72
#> 202   2058 2  1646    72
#> 203   2058 2  1646    72
#> 204   2058 2  1646    72
#> 205   2058 2  1646    72
#> 206   2058 2  1646    72
#> 207   2058 2  1646    72
#> 208   2058 2  1646    72
#> 209   2058 2  1646    72
#> 210   2058 2  1646    72
#> 211   2058 2  1646    72
#> 212   2058 2  1646    72
#> 213   2058 2  1646    72
#> 214   2058 2  1646    72
#> 215   2058 2  1646    72
#> 216   2058 2  1646    72
#> 217   2058 2  1646    72
#> 218   2058 2  1646    72
#> 219   2058 2  1646    72
#> 220   2058 2  1646    72
#> 221   2058 2  1646    72
#> 222   2058 2  1646    72
#> 223   2058 2  1646    72
#> 224   2058 2  1646    72
#> 225   2058 2  1646    72
#> 226   2058 2  1646    72
#> 227   2058 2  1646    72
#> 228   2058 2  1646    72
#> 229   2058 2  1646    72
#> 230   2058 2  1646    72
#> 231   2058 2  1646    72
#> 232   2058 2  1646    72
#> 233   2058 2  1646    72
#> 234   2058 2  1646    72
#> 235   2058 2  1646    72
#> 236   2058 2  1646    72
#> 237   2058 2  1646    72
#> 238   2058 2  1646    72
#> 239   2058 2  1646    72
#> 240   2058 2  1646    72
#> 241   2058 2  1646    72
#> 242   2058 2  1646    72
#> 243   2058 2  1646    72
#> 244   2058 2  1646    72
#> 245   2058 2  1646    72
#> 246   2058 2  1646    72
#> 247   2058 2  1646    72
#> 248   2058 2  1646    72
#> 249   2058 2  1646    72
#> 250   2058 2  1646    72
#> 251    412 3   330    72
#> 252    412 3   330    72
#> 253    412 3   330    72
#> 254    412 3   330    72
#> 255    412 3   330    72
#> 256    412 3   330    72
#> 257    412 3   330    72
#> 258    412 3   330    72
#> 259    412 3   330    72
#> 260    412 3   330    72
#> 261    412 3   330    72
#> 262    412 3   330    72
#> 263    412 3   330    72
#> 264    412 3   330    72
#> 265    412 3   330    72
#> 266    412 3   330    72
#> 267    412 3   330    72
#> 268    412 3   330    72
#> 269    412 3   330    72
#> 270    412 3   330    72
#> 271    412 3   330    72
#> 272    412 3   330    72
#> 273    412 3   330    72
#> 274    412 3   330    72
#> 275    412 3   330    72
#> 276    412 3   330    72
#> 277    412 3   330    72
#> 278    412 3   330    72
#> 279    412 3   330    72
#> 280    412 3   330    72
#> 281    412 3   330    72
#> 282    412 3   330    72
#> 283    412 3   330    72
#> 284    412 3   330    72
#> 285    412 3   330    72
#> 286    412 3   330    72
#> 287    412 3   330    72
#> 288    412 3   330    72
#> 289    412 3   330    72
#> 290    412 3   330    72
#> 291    412 3   330    72
#> 292    412 3   330    72
#> 293    412 3   330    72
#> 294    412 3   330    72
#> 295    412 3   330    72
#> 296    412 3   330    72
#> 297    412 3   330    72
#> 298    412 3   330    72
#> 299    412 3   330    72
#> 300    412 3   330    72
#> 301    824 3   659    72
#> 302    824 3   659    72
#> 303    824 3   659    72
#> 304    824 3   659    72
#> 305    824 3   659    72
#> 306    824 3   659    72
#> 307    824 3   659    72
#> 308    824 3   659    72
#> 309    824 3   659    72
#> 310    824 3   659    72
#> 311    824 3   659    72
#> 312    824 3   659    72
#> 313    824 3   659    72
#> 314    824 3   659    72
#> 315    824 3   659    72
#> 316    824 3   659    72
#> 317    824 3   659    72
#> 318    824 3   659    72
#> 319    824 3   659    72
#> 320    824 3   659    72
#> 321    824 3   659    72
#> 322    824 3   659    72
#> 323    824 3   659    72
#> 324    824 3   659    72
#> 325    824 3   659    72
#> 326    824 3   659    72
#> 327    824 3   659    72
#> 328    824 3   659    72
#> 329    824 3   659    72
#> 330    824 3   659    72
#> 331    824 3   659    72
#> 332    824 3   659    72
#> 333    824 3   659    72
#> 334    824 3   659    72
#> 335    824 3   659    72
#> 336    824 3   659    72
#> 337    824 3   659    72
#> 338    824 3   659    72
#> 339    824 3   659    72
#> 340    824 3   659    72
#> 341    824 3   659    72
#> 342    824 3   659    72
#> 343    824 3   659    72
#> 344    824 3   659    72
#> 345    824 3   659    72
#> 346    824 3   659    72
#> 347    824 3   659    72
#> 348    824 3   659    72
#> 349    824 3   659    72
#> 350    824 3   659    72
#> 351   1235 3   988    72
#> 352   1235 3   988    72
#> 353   1235 3   988    72
#> 354   1235 3   988    72
#> 355   1235 3   988    72
#> 356   1235 3   988    72
#> 357   1235 3   988    72
#> 358   1235 3   988    72
#> 359   1235 3   988    72
#> 360   1235 3   988    72
#> 361   1235 3   988    72
#> 362   1235 3   988    72
#> 363   1235 3   988    72
#> 364   1235 3   988    72
#> 365   1235 3   988    72
#> 366   1235 3   988    72
#> 367   1235 3   988    72
#> 368   1235 3   988    72
#> 369   1235 3   988    72
#> 370   1235 3   988    72
#> 371   1235 3   988    72
#> 372   1235 3   988    72
#> 373   1235 3   988    72
#> 374   1235 3   988    72
#> 375   1235 3   988    72
#> 376   1235 3   988    72
#> 377   1235 3   988    72
#> 378   1235 3   988    72
#> 379   1235 3   988    72
#> 380   1235 3   988    72
#> 381   1235 3   988    72
#> 382   1235 3   988    72
#> 383   1235 3   988    72
#> 384   1235 3   988    72
#> 385   1235 3   988    72
#> 386   1235 3   988    72
#> 387   1235 3   988    72
#> 388   1235 3   988    72
#> 389   1235 3   988    72
#> 390   1235 3   988    72
#> 391   1235 3   988    72
#> 392   1235 3   988    72
#> 393   1235 3   988    72
#> 394   1235 3   988    72
#> 395   1235 3   988    72
#> 396   1235 3   988    72
#> 397   1235 3   988    72
#> 398   1235 3   988    72
#> 399   1235 3   988    72
#> 400   1235 3   988    72
#> 401   1646 3  1317    72
#> 402   1646 3  1317    72
#> 403   1646 3  1317    72
#> 404   1646 3  1317    72
#> 405   1646 3  1317    72
#> 406   1646 3  1317    72
#> 407   1646 3  1317    72
#> 408   1646 3  1317    72
#> 409   1646 3  1317    72
#> 410   1646 3  1317    72
#> 411   1646 3  1317    72
#> 412   1646 3  1317    72
#> 413   1646 3  1317    72
#> 414   1646 3  1317    72
#> 415   1646 3  1317    72
#> 416   1646 3  1317    72
#> 417   1646 3  1317    72
#> 418   1646 3  1317    72
#> 419   1646 3  1317    72
#> 420   1646 3  1317    72
#> 421   1646 3  1317    72
#> 422   1646 3  1317    72
#> 423   1646 3  1317    72
#> 424   1646 3  1317    72
#> 425   1646 3  1317    72
#> 426   1646 3  1317    72
#> 427   1646 3  1317    72
#> 428   1646 3  1317    72
#> 429   1646 3  1317    72
#> 430   1646 3  1317    72
#> 431   1646 3  1317    72
#> 432   1646 3  1317    72
#> 433   1646 3  1317    72
#> 434   1646 3  1317    72
#> 435   1646 3  1317    72
#> 436   1646 3  1317    72
#> 437   1646 3  1317    72
#> 438   1646 3  1317    72
#> 439   1646 3  1317    72
#> 440   1646 3  1317    72
#> 441   1646 3  1317    72
#> 442   1646 3  1317    72
#> 443   1646 3  1317    72
#> 444   1646 3  1317    72
#> 445   1646 3  1317    72
#> 446   1646 3  1317    72
#> 447   1646 3  1317    72
#> 448   1646 3  1317    72
#> 449   1646 3  1317    72
#> 450   1646 3  1317    72
#> 451   2058 3  1646    72
#> 452   2058 3  1646    72
#> 453   2058 3  1646    72
#> 454   2058 3  1646    72
#> 455   2058 3  1646    72
#> 456   2058 3  1646    72
#> 457   2058 3  1646    72
#> 458   2058 3  1646    72
#> 459   2058 3  1646    72
#> 460   2058 3  1646    72
#> 461   2058 3  1646    72
#> 462   2058 3  1646    72
#> 463   2058 3  1646    72
#> 464   2058 3  1646    72
#> 465   2058 3  1646    72
#> 466   2058 3  1646    72
#> 467   2058 3  1646    72
#> 468   2058 3  1646    72
#> 469   2058 3  1646    72
#> 470   2058 3  1646    72
#> 471   2058 3  1646    72
#> 472   2058 3  1646    72
#> 473   2058 3  1646    72
#> 474   2058 3  1646    72
#> 475   2058 3  1646    72
#> 476   2058 3  1646    72
#> 477   2058 3  1646    72
#> 478   2058 3  1646    72
#> 479   2058 3  1646    72
#> 480   2058 3  1646    72
#> 481   2058 3  1646    72
#> 482   2058 3  1646    72
#> 483   2058 3  1646    72
#> 484   2058 3  1646    72
#> 485   2058 3  1646    72
#> 486   2058 3  1646    72
#> 487   2058 3  1646    72
#> 488   2058 3  1646    72
#> 489   2058 3  1646    72
#> 490   2058 3  1646    72
#> 491   2058 3  1646    72
#> 492   2058 3  1646    72
#> 493   2058 3  1646    72
#> 494   2058 3  1646    72
#> 495   2058 3  1646    72
#> 496   2058 3  1646    72
#> 497   2058 3  1646    72
#> 498   2058 3  1646    72
#> 499   2058 3  1646    72
#> 500   2058 3  1646    72
#> 501    412 4   330    72
#> 502    412 4   330    72
#> 503    412 4   330    72
#> 504    412 4   330    72
#> 505    412 4   330    72
#> 506    412 4   330    72
#> 507    412 4   330    72
#> 508    412 4   330    72
#> 509    412 4   330    72
#> 510    412 4   330    72
#> 511    412 4   330    72
#> 512    412 4   330    72
#> 513    412 4   330    72
#> 514    412 4   330    72
#> 515    412 4   330    72
#> 516    412 4   330    72
#> 517    412 4   330    72
#> 518    412 4   330    72
#> 519    412 4   330    72
#> 520    412 4   330    72
#> 521    412 4   330    72
#> 522    412 4   330    72
#> 523    412 4   330    72
#> 524    412 4   330    72
#> 525    412 4   330    72
#> 526    412 4   330    72
#> 527    412 4   330    72
#> 528    412 4   330    72
#> 529    412 4   330    72
#> 530    412 4   330    72
#> 531    412 4   330    72
#> 532    412 4   330    72
#> 533    412 4   330    72
#> 534    412 4   330    72
#> 535    412 4   330    72
#> 536    412 4   330    72
#> 537    412 4   330    72
#> 538    412 4   330    72
#> 539    412 4   330    72
#> 540    412 4   330    72
#> 541    412 4   330    72
#> 542    412 4   330    72
#> 543    412 4   330    72
#> 544    412 4   330    72
#> 545    412 4   330    72
#> 546    412 4   330    72
#> 547    412 4   330    72
#> 548    412 4   330    72
#> 549    412 4   330    72
#> 550    412 4   330    72
#> 551    824 4   659    72
#> 552    824 4   659    72
#> 553    824 4   659    72
#> 554    824 4   659    72
#> 555    824 4   659    72
#> 556    824 4   659    72
#> 557    824 4   659    72
#> 558    824 4   659    72
#> 559    824 4   659    72
#> 560    824 4   659    72
#> 561    824 4   659    72
#> 562    824 4   659    72
#> 563    824 4   659    72
#> 564    824 4   659    72
#> 565    824 4   659    72
#> 566    824 4   659    72
#> 567    824 4   659    72
#> 568    824 4   659    72
#> 569    824 4   659    72
#> 570    824 4   659    72
#> 571    824 4   659    72
#> 572    824 4   659    72
#> 573    824 4   659    72
#> 574    824 4   659    72
#> 575    824 4   659    72
#> 576    824 4   659    72
#> 577    824 4   659    72
#> 578    824 4   659    72
#> 579    824 4   659    72
#> 580    824 4   659    72
#> 581    824 4   659    72
#> 582    824 4   659    72
#> 583    824 4   659    72
#> 584    824 4   659    72
#> 585    824 4   659    72
#> 586    824 4   659    72
#> 587    824 4   659    72
#> 588    824 4   659    72
#> 589    824 4   659    72
#> 590    824 4   659    72
#> 591    824 4   659    72
#> 592    824 4   659    72
#> 593    824 4   659    72
#> 594    824 4   659    72
#> 595    824 4   659    72
#> 596    824 4   659    72
#> 597    824 4   659    72
#> 598    824 4   659    72
#> 599    824 4   659    72
#> 600    824 4   659    72
#> 601   1235 4   988    72
#> 602   1235 4   988    72
#> 603   1235 4   988    72
#> 604   1235 4   988    72
#> 605   1235 4   988    72
#> 606   1235 4   988    72
#> 607   1235 4   988    72
#> 608   1235 4   988    72
#> 609   1235 4   988    72
#> 610   1235 4   988    72
#> 611   1235 4   988    72
#> 612   1235 4   988    72
#> 613   1235 4   988    72
#> 614   1235 4   988    72
#> 615   1235 4   988    72
#> 616   1235 4   988    72
#> 617   1235 4   988    72
#> 618   1235 4   988    72
#> 619   1235 4   988    72
#> 620   1235 4   988    72
#> 621   1235 4   988    72
#> 622   1235 4   988    72
#> 623   1235 4   988    72
#> 624   1235 4   988    72
#> 625   1235 4   988    72
#> 626   1235 4   988    72
#> 627   1235 4   988    72
#> 628   1235 4   988    72
#> 629   1235 4   988    72
#> 630   1235 4   988    72
#> 631   1235 4   988    72
#> 632   1235 4   988    72
#> 633   1235 4   988    72
#> 634   1235 4   988    72
#> 635   1235 4   988    72
#> 636   1235 4   988    72
#> 637   1235 4   988    72
#> 638   1235 4   988    72
#> 639   1235 4   988    72
#> 640   1235 4   988    72
#> 641   1235 4   988    72
#> 642   1235 4   988    72
#> 643   1235 4   988    72
#> 644   1235 4   988    72
#> 645   1235 4   988    72
#> 646   1235 4   988    72
#> 647   1235 4   988    72
#> 648   1235 4   988    72
#> 649   1235 4   988    72
#> 650   1235 4   988    72
#> 651   1646 4  1317    72
#> 652   1646 4  1317    72
#> 653   1646 4  1317    72
#> 654   1646 4  1317    72
#> 655   1646 4  1317    72
#> 656   1646 4  1317    72
#> 657   1646 4  1317    72
#> 658   1646 4  1317    72
#> 659   1646 4  1317    72
#> 660   1646 4  1317    72
#> 661   1646 4  1317    72
#> 662   1646 4  1317    72
#> 663   1646 4  1317    72
#> 664   1646 4  1317    72
#> 665   1646 4  1317    72
#> 666   1646 4  1317    72
#> 667   1646 4  1317    72
#> 668   1646 4  1317    72
#> 669   1646 4  1317    72
#> 670   1646 4  1317    72
#> 671   1646 4  1317    72
#> 672   1646 4  1317    72
#> 673   1646 4  1317    72
#> 674   1646 4  1317    72
#> 675   1646 4  1317    72
#> 676   1646 4  1317    72
#> 677   1646 4  1317    72
#> 678   1646 4  1317    72
#> 679   1646 4  1317    72
#> 680   1646 4  1317    72
#> 681   1646 4  1317    72
#> 682   1646 4  1317    72
#> 683   1646 4  1317    72
#> 684   1646 4  1317    72
#> 685   1646 4  1317    72
#> 686   1646 4  1317    72
#> 687   1646 4  1317    72
#> 688   1646 4  1317    72
#> 689   1646 4  1317    72
#> 690   1646 4  1317    72
#> 691   1646 4  1317    72
#> 692   1646 4  1317    72
#> 693   1646 4  1317    72
#> 694   1646 4  1317    72
#> 695   1646 4  1317    72
#> 696   1646 4  1317    72
#> 697   1646 4  1317    72
#> 698   1646 4  1317    72
#> 699   1646 4  1317    72
#> 700   1646 4  1317    72
#> 701   2058 4  1646    72
#> 702   2058 4  1646    72
#> 703   2058 4  1646    72
#> 704   2058 4  1646    72
#> 705   2058 4  1646    72
#> 706   2058 4  1646    72
#> 707   2058 4  1646    72
#> 708   2058 4  1646    72
#> 709   2058 4  1646    72
#> 710   2058 4  1646    72
#> 711   2058 4  1646    72
#> 712   2058 4  1646    72
#> 713   2058 4  1646    72
#> 714   2058 4  1646    72
#> 715   2058 4  1646    72
#> 716   2058 4  1646    72
#> 717   2058 4  1646    72
#> 718   2058 4  1646    72
#> 719   2058 4  1646    72
#> 720   2058 4  1646    72
#> 721   2058 4  1646    72
#> 722   2058 4  1646    72
#> 723   2058 4  1646    72
#> 724   2058 4  1646    72
#> 725   2058 4  1646    72
#> 726   2058 4  1646    72
#> 727   2058 4  1646    72
#> 728   2058 4  1646    72
#> 729   2058 4  1646    72
#> 730   2058 4  1646    72
#> 731   2058 4  1646    72
#> 732   2058 4  1646    72
#> 733   2058 4  1646    72
#> 734   2058 4  1646    72
#> 735   2058 4  1646    72
#> 736   2058 4  1646    72
#> 737   2058 4  1646    72
#> 738   2058 4  1646    72
#> 739   2058 4  1646    72
#> 740   2058 4  1646    72
#> 741   2058 4  1646    72
#> 742   2058 4  1646    72
#> 743   2058 4  1646    72
#> 744   2058 4  1646    72
#> 745   2058 4  1646    72
#> 746   2058 4  1646    72
#> 747   2058 4  1646    72
#> 748   2058 4  1646    72
#> 749   2058 4  1646    72
#> 750   2058 4  1646    72
#> 751    412 5   330    72
#> 752    412 5   330    72
#> 753    412 5   330    72
#> 754    412 5   330    72
#> 755    412 5   330    72
#> 756    412 5   330    72
#> 757    412 5   330    72
#> 758    412 5   330    72
#> 759    412 5   330    72
#> 760    412 5   330    72
#> 761    412 5   330    72
#> 762    412 5   330    72
#> 763    412 5   330    72
#> 764    412 5   330    72
#> 765    412 5   330    72
#> 766    412 5   330    72
#> 767    412 5   330    72
#> 768    412 5   330    72
#> 769    412 5   330    72
#> 770    412 5   330    72
#> 771    412 5   330    72
#> 772    412 5   330    72
#> 773    412 5   330    72
#> 774    412 5   330    72
#> 775    412 5   330    72
#> 776    412 5   330    72
#> 777    412 5   330    72
#> 778    412 5   330    72
#> 779    412 5   330    72
#> 780    412 5   330    72
#> 781    412 5   330    72
#> 782    412 5   330    72
#> 783    412 5   330    72
#> 784    412 5   330    72
#> 785    412 5   330    72
#> 786    412 5   330    72
#> 787    412 5   330    72
#> 788    412 5   330    72
#> 789    412 5   330    72
#> 790    412 5   330    72
#> 791    412 5   330    72
#> 792    412 5   330    72
#> 793    412 5   330    72
#> 794    412 5   330    72
#> 795    412 5   330    72
#> 796    412 5   330    72
#> 797    412 5   330    72
#> 798    412 5   330    72
#> 799    412 5   330    72
#> 800    412 5   330    72
#> 801    824 5   659    72
#> 802    824 5   659    72
#> 803    824 5   659    72
#> 804    824 5   659    72
#> 805    824 5   659    72
#> 806    824 5   659    72
#> 807    824 5   659    72
#> 808    824 5   659    72
#> 809    824 5   659    72
#> 810    824 5   659    72
#> 811    824 5   659    72
#> 812    824 5   659    72
#> 813    824 5   659    72
#> 814    824 5   659    72
#> 815    824 5   659    72
#> 816    824 5   659    72
#> 817    824 5   659    72
#> 818    824 5   659    72
#> 819    824 5   659    72
#> 820    824 5   659    72
#> 821    824 5   659    72
#> 822    824 5   659    72
#> 823    824 5   659    72
#> 824    824 5   659    72
#> 825    824 5   659    72
#> 826    824 5   659    72
#> 827    824 5   659    72
#> 828    824 5   659    72
#> 829    824 5   659    72
#> 830    824 5   659    72
#> 831    824 5   659    72
#> 832    824 5   659    72
#> 833    824 5   659    72
#> 834    824 5   659    72
#> 835    824 5   659    72
#> 836    824 5   659    72
#> 837    824 5   659    72
#> 838    824 5   659    72
#> 839    824 5   659    72
#> 840    824 5   659    72
#> 841    824 5   659    72
#> 842    824 5   659    72
#> 843    824 5   659    72
#> 844    824 5   659    72
#> 845    824 5   659    72
#> 846    824 5   659    72
#> 847    824 5   659    72
#> 848    824 5   659    72
#> 849    824 5   659    72
#> 850    824 5   659    72
#> 851   1235 5   988    72
#> 852   1235 5   988    72
#> 853   1235 5   988    72
#> 854   1235 5   988    72
#> 855   1235 5   988    72
#> 856   1235 5   988    72
#> 857   1235 5   988    72
#> 858   1235 5   988    72
#> 859   1235 5   988    72
#> 860   1235 5   988    72
#> 861   1235 5   988    72
#> 862   1235 5   988    72
#> 863   1235 5   988    72
#> 864   1235 5   988    72
#> 865   1235 5   988    72
#> 866   1235 5   988    72
#> 867   1235 5   988    72
#> 868   1235 5   988    72
#> 869   1235 5   988    72
#> 870   1235 5   988    72
#> 871   1235 5   988    72
#> 872   1235 5   988    72
#> 873   1235 5   988    72
#> 874   1235 5   988    72
#> 875   1235 5   988    72
#> 876   1235 5   988    72
#> 877   1235 5   988    72
#> 878   1235 5   988    72
#> 879   1235 5   988    72
#> 880   1235 5   988    72
#> 881   1235 5   988    72
#> 882   1235 5   988    72
#> 883   1235 5   988    72
#> 884   1235 5   988    72
#> 885   1235 5   988    72
#> 886   1235 5   988    72
#> 887   1235 5   988    72
#> 888   1235 5   988    72
#> 889   1235 5   988    72
#> 890   1235 5   988    72
#> 891   1235 5   988    72
#> 892   1235 5   988    72
#> 893   1235 5   988    72
#> 894   1235 5   988    72
#> 895   1235 5   988    72
#> 896   1235 5   988    72
#> 897   1235 5   988    72
#> 898   1235 5   988    72
#> 899   1235 5   988    72
#> 900   1235 5   988    72
#> 901   1646 5  1317    72
#> 902   1646 5  1317    72
#> 903   1646 5  1317    72
#> 904   1646 5  1317    72
#> 905   1646 5  1317    72
#> 906   1646 5  1317    72
#> 907   1646 5  1317    72
#> 908   1646 5  1317    72
#> 909   1646 5  1317    72
#> 910   1646 5  1317    72
#> 911   1646 5  1317    72
#> 912   1646 5  1317    72
#> 913   1646 5  1317    72
#> 914   1646 5  1317    72
#> 915   1646 5  1317    72
#> 916   1646 5  1317    72
#> 917   1646 5  1317    72
#> 918   1646 5  1317    72
#> 919   1646 5  1317    72
#> 920   1646 5  1317    72
#> 921   1646 5  1317    72
#> 922   1646 5  1317    72
#> 923   1646 5  1317    72
#> 924   1646 5  1317    72
#> 925   1646 5  1317    72
#> 926   1646 5  1317    72
#> 927   1646 5  1317    72
#> 928   1646 5  1317    72
#> 929   1646 5  1317    72
#> 930   1646 5  1317    72
#> 931   1646 5  1317    72
#> 932   1646 5  1317    72
#> 933   1646 5  1317    72
#> 934   1646 5  1317    72
#> 935   1646 5  1317    72
#> 936   1646 5  1317    72
#> 937   1646 5  1317    72
#> 938   1646 5  1317    72
#> 939   1646 5  1317    72
#> 940   1646 5  1317    72
#> 941   1646 5  1317    72
#> 942   1646 5  1317    72
#> 943   1646 5  1317    72
#> 944   1646 5  1317    72
#> 945   1646 5  1317    72
#> 946   1646 5  1317    72
#> 947   1646 5  1317    72
#> 948   1646 5  1317    72
#> 949   1646 5  1317    72
#> 950   1646 5  1317    72
#> 951   2058 5  1646    72
#> 952   2058 5  1646    72
#> 953   2058 5  1646    72
#> 954   2058 5  1646    72
#> 955   2058 5  1646    72
#> 956   2058 5  1646    72
#> 957   2058 5  1646    72
#> 958   2058 5  1646    72
#> 959   2058 5  1646    72
#> 960   2058 5  1646    72
#> 961   2058 5  1646    72
#> 962   2058 5  1646    72
#> 963   2058 5  1646    72
#> 964   2058 5  1646    72
#> 965   2058 5  1646    72
#> 966   2058 5  1646    72
#> 967   2058 5  1646    72
#> 968   2058 5  1646    72
#> 969   2058 5  1646    72
#> 970   2058 5  1646    72
#> 971   2058 5  1646    72
#> 972   2058 5  1646    72
#> 973   2058 5  1646    72
#> 974   2058 5  1646    72
#> 975   2058 5  1646    72
#> 976   2058 5  1646    72
#> 977   2058 5  1646    72
#> 978   2058 5  1646    72
#> 979   2058 5  1646    72
#> 980   2058 5  1646    72
#> 981   2058 5  1646    72
#> 982   2058 5  1646    72
#> 983   2058 5  1646    72
#> 984   2058 5  1646    72
#> 985   2058 5  1646    72
#> 986   2058 5  1646    72
#> 987   2058 5  1646    72
#> 988   2058 5  1646    72
#> 989   2058 5  1646    72
#> 990   2058 5  1646    72
#> 991   2058 5  1646    72
#> 992   2058 5  1646    72
#> 993   2058 5  1646    72
#> 994   2058 5  1646    72
#> 995   2058 5  1646    72
#> 996   2058 5  1646    72
#> 997   2058 5  1646    72
#> 998   2058 5  1646    72
#> 999   2058 5  1646    72
#> 1000  2058 5  1646    72
#> 1001   412 6   330    72
#> 1002   412 6   330    72
#> 1003   412 6   330    72
#> 1004   412 6   330    72
#> 1005   412 6   330    72
#> 1006   412 6   330    72
#> 1007   412 6   330    72
#> 1008   412 6   330    72
#> 1009   412 6   330    72
#> 1010   412 6   330    72
#> 1011   412 6   330    72
#> 1012   412 6   330    72
#> 1013   412 6   330    72
#> 1014   412 6   330    72
#> 1015   412 6   330    72
#> 1016   412 6   330    72
#> 1017   412 6   330    72
#> 1018   412 6   330    72
#> 1019   412 6   330    72
#> 1020   412 6   330    72
#> 1021   412 6   330    72
#> 1022   412 6   330    72
#> 1023   412 6   330    72
#> 1024   412 6   330    72
#> 1025   412 6   330    72
#> 1026   412 6   330    72
#> 1027   412 6   330    72
#> 1028   412 6   330    72
#> 1029   412 6   330    72
#> 1030   412 6   330    72
#> 1031   412 6   330    72
#> 1032   412 6   330    72
#> 1033   412 6   330    72
#> 1034   412 6   330    72
#> 1035   412 6   330    72
#> 1036   412 6   330    72
#> 1037   412 6   330    72
#> 1038   412 6   330    72
#> 1039   412 6   330    72
#> 1040   412 6   330    72
#> 1041   412 6   330    72
#> 1042   412 6   330    72
#> 1043   412 6   330    72
#> 1044   412 6   330    72
#> 1045   412 6   330    72
#> 1046   412 6   330    72
#> 1047   412 6   330    72
#> 1048   412 6   330    72
#> 1049   412 6   330    72
#> 1050   412 6   330    72
#> 1051   824 6   659    72
#> 1052   824 6   659    72
#> 1053   824 6   659    72
#> 1054   824 6   659    72
#> 1055   824 6   659    72
#> 1056   824 6   659    72
#> 1057   824 6   659    72
#> 1058   824 6   659    72
#> 1059   824 6   659    72
#> 1060   824 6   659    72
#> 1061   824 6   659    72
#> 1062   824 6   659    72
#> 1063   824 6   659    72
#> 1064   824 6   659    72
#> 1065   824 6   659    72
#> 1066   824 6   659    72
#> 1067   824 6   659    72
#> 1068   824 6   659    72
#> 1069   824 6   659    72
#> 1070   824 6   659    72
#> 1071   824 6   659    72
#> 1072   824 6   659    72
#> 1073   824 6   659    72
#> 1074   824 6   659    72
#> 1075   824 6   659    72
#> 1076   824 6   659    72
#> 1077   824 6   659    72
#> 1078   824 6   659    72
#> 1079   824 6   659    72
#> 1080   824 6   659    72
#> 1081   824 6   659    72
#> 1082   824 6   659    72
#> 1083   824 6   659    72
#> 1084   824 6   659    72
#> 1085   824 6   659    72
#> 1086   824 6   659    72
#> 1087   824 6   659    72
#> 1088   824 6   659    72
#> 1089   824 6   659    72
#> 1090   824 6   659    72
#> 1091   824 6   659    72
#> 1092   824 6   659    72
#> 1093   824 6   659    72
#> 1094   824 6   659    72
#> 1095   824 6   659    72
#> 1096   824 6   659    72
#> 1097   824 6   659    72
#> 1098   824 6   659    72
#> 1099   824 6   659    72
#> 1100   824 6   659    72
#> 1101  1235 6   988    72
#> 1102  1235 6   988    72
#> 1103  1235 6   988    72
#> 1104  1235 6   988    72
#> 1105  1235 6   988    72
#> 1106  1235 6   988    72
#> 1107  1235 6   988    72
#> 1108  1235 6   988    72
#> 1109  1235 6   988    72
#> 1110  1235 6   988    72
#> 1111  1235 6   988    72
#> 1112  1235 6   988    72
#> 1113  1235 6   988    72
#> 1114  1235 6   988    72
#> 1115  1235 6   988    72
#> 1116  1235 6   988    72
#> 1117  1235 6   988    72
#> 1118  1235 6   988    72
#> 1119  1235 6   988    72
#> 1120  1235 6   988    72
#> 1121  1235 6   988    72
#> 1122  1235 6   988    72
#> 1123  1235 6   988    72
#> 1124  1235 6   988    72
#> 1125  1235 6   988    72
#> 1126  1235 6   988    72
#> 1127  1235 6   988    72
#> 1128  1235 6   988    72
#> 1129  1235 6   988    72
#> 1130  1235 6   988    72
#> 1131  1235 6   988    72
#> 1132  1235 6   988    72
#> 1133  1235 6   988    72
#> 1134  1235 6   988    72
#> 1135  1235 6   988    72
#> 1136  1235 6   988    72
#> 1137  1235 6   988    72
#> 1138  1235 6   988    72
#> 1139  1235 6   988    72
#> 1140  1235 6   988    72
#> 1141  1235 6   988    72
#> 1142  1235 6   988    72
#> 1143  1235 6   988    72
#> 1144  1235 6   988    72
#> 1145  1235 6   988    72
#> 1146  1235 6   988    72
#> 1147  1235 6   988    72
#> 1148  1235 6   988    72
#> 1149  1235 6   988    72
#> 1150  1235 6   988    72
#> 1151  1646 6  1317    72
#> 1152  1646 6  1317    72
#> 1153  1646 6  1317    72
#> 1154  1646 6  1317    72
#> 1155  1646 6  1317    72
#> 1156  1646 6  1317    72
#> 1157  1646 6  1317    72
#> 1158  1646 6  1317    72
#> 1159  1646 6  1317    72
#> 1160  1646 6  1317    72
#> 1161  1646 6  1317    72
#> 1162  1646 6  1317    72
#> 1163  1646 6  1317    72
#> 1164  1646 6  1317    72
#> 1165  1646 6  1317    72
#> 1166  1646 6  1317    72
#> 1167  1646 6  1317    72
#> 1168  1646 6  1317    72
#> 1169  1646 6  1317    72
#> 1170  1646 6  1317    72
#> 1171  1646 6  1317    72
#> 1172  1646 6  1317    72
#> 1173  1646 6  1317    72
#> 1174  1646 6  1317    72
#> 1175  1646 6  1317    72
#> 1176  1646 6  1317    72
#> 1177  1646 6  1317    72
#> 1178  1646 6  1317    72
#> 1179  1646 6  1317    72
#> 1180  1646 6  1317    72
#> 1181  1646 6  1317    72
#> 1182  1646 6  1317    72
#> 1183  1646 6  1317    72
#> 1184  1646 6  1317    72
#> 1185  1646 6  1317    72
#> 1186  1646 6  1317    72
#> 1187  1646 6  1317    72
#> 1188  1646 6  1317    72
#> 1189  1646 6  1317    72
#> 1190  1646 6  1317    72
#> 1191  1646 6  1317    72
#> 1192  1646 6  1317    72
#> 1193  1646 6  1317    72
#> 1194  1646 6  1317    72
#> 1195  1646 6  1317    72
#> 1196  1646 6  1317    72
#> 1197  1646 6  1317    72
#> 1198  1646 6  1317    72
#> 1199  1646 6  1317    72
#> 1200  1646 6  1317    72
#> 1201  2058 6  1646    72
#> 1202  2058 6  1646    72
#> 1203  2058 6  1646    72
#> 1204  2058 6  1646    72
#> 1205  2058 6  1646    72
#> 1206  2058 6  1646    72
#> 1207  2058 6  1646    72
#> 1208  2058 6  1646    72
#> 1209  2058 6  1646    72
#> 1210  2058 6  1646    72
#> 1211  2058 6  1646    72
#> 1212  2058 6  1646    72
#> 1213  2058 6  1646    72
#> 1214  2058 6  1646    72
#> 1215  2058 6  1646    72
#> 1216  2058 6  1646    72
#> 1217  2058 6  1646    72
#> 1218  2058 6  1646    72
#> 1219  2058 6  1646    72
#> 1220  2058 6  1646    72
#> 1221  2058 6  1646    72
#> 1222  2058 6  1646    72
#> 1223  2058 6  1646    72
#> 1224  2058 6  1646    72
#> 1225  2058 6  1646    72
#> 1226  2058 6  1646    72
#> 1227  2058 6  1646    72
#> 1228  2058 6  1646    72
#> 1229  2058 6  1646    72
#> 1230  2058 6  1646    72
#> 1231  2058 6  1646    72
#> 1232  2058 6  1646    72
#> 1233  2058 6  1646    72
#> 1234  2058 6  1646    72
#> 1235  2058 6  1646    72
#> 1236  2058 6  1646    72
#> 1237  2058 6  1646    72
#> 1238  2058 6  1646    72
#> 1239  2058 6  1646    72
#> 1240  2058 6  1646    72
#> 1241  2058 6  1646    72
#> 1242  2058 6  1646    72
#> 1243  2058 6  1646    72
#> 1244  2058 6  1646    72
#> 1245  2058 6  1646    72
#> 1246  2058 6  1646    72
#> 1247  2058 6  1646    72
#> 1248  2058 6  1646    72
#> 1249  2058 6  1646    72
#> 1250  2058 6  1646    72