Get parameters

Get parameters

# S4 method for ConsensusPartition
get_param(object, k = object@k, unique = TRUE)

Arguments

object
A ConsensusPartition-class object.

k
Number of partitions.

unique
Whether apply unique to rows of the returned data frame.

Details

It is mainly used internally.

Value

A data frame of parameters corresponding to the current k. In the data frame, each row corresponds to a partition run.

Examples

data(cola_rl)
obj = cola_rl["sd", "kmeans"]
get_param(obj)

#>    top_n k n_row n_col
#> 1     20 2    16    60
#> 2     30 2    24    60
#> 3     40 2    32    60
#> 4     20 3    16    60
#> 5     30 3    24    60
#> 6     40 3    32    60
#> 7     20 4    16    60
#> 8     30 4    24    60
#> 9     40 4    32    60
#> 10    20 5    16    60
#> 11    30 5    24    60
#> 12    40 5    32    60
#> 13    20 6    16    60
#> 14    30 6    24    60
#> 15    40 6    32    60

get_param(obj, k = 2)

#>   top_n k n_row n_col
#> 1    20 2    16    60
#> 2    30 2    24    60
#> 3    40 2    32    60

get_param(obj, unique = FALSE)

#>     top_n k n_row n_col
#> 1      20 2    16    60
#> 2      20 2    16    60
#> 3      20 2    16    60
#> 4      20 2    16    60
#> 5      20 2    16    60
#> 6      20 2    16    60
#> 7      20 2    16    60
#> 8      20 2    16    60
#> 9      20 2    16    60
#> 10     20 2    16    60
#> 11     20 2    16    60
#> 12     20 2    16    60
#> 13     20 2    16    60
#> 14     20 2    16    60
#> 15     20 2    16    60
#> 16     20 2    16    60
#> 17     20 2    16    60
#> 18     20 2    16    60
#> 19     20 2    16    60
#> 20     20 2    16    60
#> 21     20 2    16    60
#> 22     20 2    16    60
#> 23     20 2    16    60
#> 24     20 2    16    60
#> 25     20 2    16    60
#> 26     20 2    16    60
#> 27     20 2    16    60
#> 28     20 2    16    60
#> 29     20 2    16    60
#> 30     20 2    16    60
#> 31     20 2    16    60
#> 32     20 2    16    60
#> 33     20 2    16    60
#> 34     20 2    16    60
#> 35     20 2    16    60
#> 36     20 2    16    60
#> 37     20 2    16    60
#> 38     20 2    16    60
#> 39     20 2    16    60
#> 40     20 2    16    60
#> 41     20 2    16    60
#> 42     20 2    16    60
#> 43     20 2    16    60
#> 44     20 2    16    60
#> 45     20 2    16    60
#> 46     20 2    16    60
#> 47     20 2    16    60
#> 48     20 2    16    60
#> 49     20 2    16    60
#> 50     20 2    16    60
#> 51     30 2    24    60
#> 52     30 2    24    60
#> 53     30 2    24    60
#> 54     30 2    24    60
#> 55     30 2    24    60
#> 56     30 2    24    60
#> 57     30 2    24    60
#> 58     30 2    24    60
#> 59     30 2    24    60
#> 60     30 2    24    60
#> 61     30 2    24    60
#> 62     30 2    24    60
#> 63     30 2    24    60
#> 64     30 2    24    60
#> 65     30 2    24    60
#> 66     30 2    24    60
#> 67     30 2    24    60
#> 68     30 2    24    60
#> 69     30 2    24    60
#> 70     30 2    24    60
#> 71     30 2    24    60
#> 72     30 2    24    60
#> 73     30 2    24    60
#> 74     30 2    24    60
#> 75     30 2    24    60
#> 76     30 2    24    60
#> 77     30 2    24    60
#> 78     30 2    24    60
#> 79     30 2    24    60
#> 80     30 2    24    60
#> 81     30 2    24    60
#> 82     30 2    24    60
#> 83     30 2    24    60
#> 84     30 2    24    60
#> 85     30 2    24    60
#> 86     30 2    24    60
#> 87     30 2    24    60
#> 88     30 2    24    60
#> 89     30 2    24    60
#> 90     30 2    24    60
#> 91     30 2    24    60
#> 92     30 2    24    60
#> 93     30 2    24    60
#> 94     30 2    24    60
#> 95     30 2    24    60
#> 96     30 2    24    60
#> 97     30 2    24    60
#> 98     30 2    24    60
#> 99     30 2    24    60
#> 100    30 2    24    60
#> 101    40 2    32    60
#> 102    40 2    32    60
#> 103    40 2    32    60
#> 104    40 2    32    60
#> 105    40 2    32    60
#> 106    40 2    32    60
#> 107    40 2    32    60
#> 108    40 2    32    60
#> 109    40 2    32    60
#> 110    40 2    32    60
#> 111    40 2    32    60
#> 112    40 2    32    60
#> 113    40 2    32    60
#> 114    40 2    32    60
#> 115    40 2    32    60
#> 116    40 2    32    60
#> 117    40 2    32    60
#> 118    40 2    32    60
#> 119    40 2    32    60
#> 120    40 2    32    60
#> 121    40 2    32    60
#> 122    40 2    32    60
#> 123    40 2    32    60
#> 124    40 2    32    60
#> 125    40 2    32    60
#> 126    40 2    32    60
#> 127    40 2    32    60
#> 128    40 2    32    60
#> 129    40 2    32    60
#> 130    40 2    32    60
#> 131    40 2    32    60
#> 132    40 2    32    60
#> 133    40 2    32    60
#> 134    40 2    32    60
#> 135    40 2    32    60
#> 136    40 2    32    60
#> 137    40 2    32    60
#> 138    40 2    32    60
#> 139    40 2    32    60
#> 140    40 2    32    60
#> 141    40 2    32    60
#> 142    40 2    32    60
#> 143    40 2    32    60
#> 144    40 2    32    60
#> 145    40 2    32    60
#> 146    40 2    32    60
#> 147    40 2    32    60
#> 148    40 2    32    60
#> 149    40 2    32    60
#> 150    40 2    32    60
#> 151    20 3    16    60
#> 152    20 3    16    60
#> 153    20 3    16    60
#> 154    20 3    16    60
#> 155    20 3    16    60
#> 156    20 3    16    60
#> 157    20 3    16    60
#> 158    20 3    16    60
#> 159    20 3    16    60
#> 160    20 3    16    60
#> 161    20 3    16    60
#> 162    20 3    16    60
#> 163    20 3    16    60
#> 164    20 3    16    60
#> 165    20 3    16    60
#> 166    20 3    16    60
#> 167    20 3    16    60
#> 168    20 3    16    60
#> 169    20 3    16    60
#> 170    20 3    16    60
#> 171    20 3    16    60
#> 172    20 3    16    60
#> 173    20 3    16    60
#> 174    20 3    16    60
#> 175    20 3    16    60
#> 176    20 3    16    60
#> 177    20 3    16    60
#> 178    20 3    16    60
#> 179    20 3    16    60
#> 180    20 3    16    60
#> 181    20 3    16    60
#> 182    20 3    16    60
#> 183    20 3    16    60
#> 184    20 3    16    60
#> 185    20 3    16    60
#> 186    20 3    16    60
#> 187    20 3    16    60
#> 188    20 3    16    60
#> 189    20 3    16    60
#> 190    20 3    16    60
#> 191    20 3    16    60
#> 192    20 3    16    60
#> 193    20 3    16    60
#> 194    20 3    16    60
#> 195    20 3    16    60
#> 196    20 3    16    60
#> 197    20 3    16    60
#> 198    20 3    16    60
#> 199    20 3    16    60
#> 200    20 3    16    60
#> 201    30 3    24    60
#> 202    30 3    24    60
#> 203    30 3    24    60
#> 204    30 3    24    60
#> 205    30 3    24    60
#> 206    30 3    24    60
#> 207    30 3    24    60
#> 208    30 3    24    60
#> 209    30 3    24    60
#> 210    30 3    24    60
#> 211    30 3    24    60
#> 212    30 3    24    60
#> 213    30 3    24    60
#> 214    30 3    24    60
#> 215    30 3    24    60
#> 216    30 3    24    60
#> 217    30 3    24    60
#> 218    30 3    24    60
#> 219    30 3    24    60
#> 220    30 3    24    60
#> 221    30 3    24    60
#> 222    30 3    24    60
#> 223    30 3    24    60
#> 224    30 3    24    60
#> 225    30 3    24    60
#> 226    30 3    24    60
#> 227    30 3    24    60
#> 228    30 3    24    60
#> 229    30 3    24    60
#> 230    30 3    24    60
#> 231    30 3    24    60
#> 232    30 3    24    60
#> 233    30 3    24    60
#> 234    30 3    24    60
#> 235    30 3    24    60
#> 236    30 3    24    60
#> 237    30 3    24    60
#> 238    30 3    24    60
#> 239    30 3    24    60
#> 240    30 3    24    60
#> 241    30 3    24    60
#> 242    30 3    24    60
#> 243    30 3    24    60
#> 244    30 3    24    60
#> 245    30 3    24    60
#> 246    30 3    24    60
#> 247    30 3    24    60
#> 248    30 3    24    60
#> 249    30 3    24    60
#> 250    30 3    24    60
#> 251    40 3    32    60
#> 252    40 3    32    60
#> 253    40 3    32    60
#> 254    40 3    32    60
#> 255    40 3    32    60
#> 256    40 3    32    60
#> 257    40 3    32    60
#> 258    40 3    32    60
#> 259    40 3    32    60
#> 260    40 3    32    60
#> 261    40 3    32    60
#> 262    40 3    32    60
#> 263    40 3    32    60
#> 264    40 3    32    60
#> 265    40 3    32    60
#> 266    40 3    32    60
#> 267    40 3    32    60
#> 268    40 3    32    60
#> 269    40 3    32    60
#> 270    40 3    32    60
#> 271    40 3    32    60
#> 272    40 3    32    60
#> 273    40 3    32    60
#> 274    40 3    32    60
#> 275    40 3    32    60
#> 276    40 3    32    60
#> 277    40 3    32    60
#> 278    40 3    32    60
#> 279    40 3    32    60
#> 280    40 3    32    60
#> 281    40 3    32    60
#> 282    40 3    32    60
#> 283    40 3    32    60
#> 284    40 3    32    60
#> 285    40 3    32    60
#> 286    40 3    32    60
#> 287    40 3    32    60
#> 288    40 3    32    60
#> 289    40 3    32    60
#> 290    40 3    32    60
#> 291    40 3    32    60
#> 292    40 3    32    60
#> 293    40 3    32    60
#> 294    40 3    32    60
#> 295    40 3    32    60
#> 296    40 3    32    60
#> 297    40 3    32    60
#> 298    40 3    32    60
#> 299    40 3    32    60
#> 300    40 3    32    60
#> 301    20 4    16    60
#> 302    20 4    16    60
#> 303    20 4    16    60
#> 304    20 4    16    60
#> 305    20 4    16    60
#> 306    20 4    16    60
#> 307    20 4    16    60
#> 308    20 4    16    60
#> 309    20 4    16    60
#> 310    20 4    16    60
#> 311    20 4    16    60
#> 312    20 4    16    60
#> 313    20 4    16    60
#> 314    20 4    16    60
#> 315    20 4    16    60
#> 316    20 4    16    60
#> 317    20 4    16    60
#> 318    20 4    16    60
#> 319    20 4    16    60
#> 320    20 4    16    60
#> 321    20 4    16    60
#> 322    20 4    16    60
#> 323    20 4    16    60
#> 324    20 4    16    60
#> 325    20 4    16    60
#> 326    20 4    16    60
#> 327    20 4    16    60
#> 328    20 4    16    60
#> 329    20 4    16    60
#> 330    20 4    16    60
#> 331    20 4    16    60
#> 332    20 4    16    60
#> 333    20 4    16    60
#> 334    20 4    16    60
#> 335    20 4    16    60
#> 336    20 4    16    60
#> 337    20 4    16    60
#> 338    20 4    16    60
#> 339    20 4    16    60
#> 340    20 4    16    60
#> 341    20 4    16    60
#> 342    20 4    16    60
#> 343    20 4    16    60
#> 344    20 4    16    60
#> 345    20 4    16    60
#> 346    20 4    16    60
#> 347    20 4    16    60
#> 348    20 4    16    60
#> 349    20 4    16    60
#> 350    20 4    16    60
#> 351    30 4    24    60
#> 352    30 4    24    60
#> 353    30 4    24    60
#> 354    30 4    24    60
#> 355    30 4    24    60
#> 356    30 4    24    60
#> 357    30 4    24    60
#> 358    30 4    24    60
#> 359    30 4    24    60
#> 360    30 4    24    60
#> 361    30 4    24    60
#> 362    30 4    24    60
#> 363    30 4    24    60
#> 364    30 4    24    60
#> 365    30 4    24    60
#> 366    30 4    24    60
#> 367    30 4    24    60
#> 368    30 4    24    60
#> 369    30 4    24    60
#> 370    30 4    24    60
#> 371    30 4    24    60
#> 372    30 4    24    60
#> 373    30 4    24    60
#> 374    30 4    24    60
#> 375    30 4    24    60
#> 376    30 4    24    60
#> 377    30 4    24    60
#> 378    30 4    24    60
#> 379    30 4    24    60
#> 380    30 4    24    60
#> 381    30 4    24    60
#> 382    30 4    24    60
#> 383    30 4    24    60
#> 384    30 4    24    60
#> 385    30 4    24    60
#> 386    30 4    24    60
#> 387    30 4    24    60
#> 388    30 4    24    60
#> 389    30 4    24    60
#> 390    30 4    24    60
#> 391    30 4    24    60
#> 392    30 4    24    60
#> 393    30 4    24    60
#> 394    30 4    24    60
#> 395    30 4    24    60
#> 396    30 4    24    60
#> 397    30 4    24    60
#> 398    30 4    24    60
#> 399    30 4    24    60
#> 400    30 4    24    60
#> 401    40 4    32    60
#> 402    40 4    32    60
#> 403    40 4    32    60
#> 404    40 4    32    60
#> 405    40 4    32    60
#> 406    40 4    32    60
#> 407    40 4    32    60
#> 408    40 4    32    60
#> 409    40 4    32    60
#> 410    40 4    32    60
#> 411    40 4    32    60
#> 412    40 4    32    60
#> 413    40 4    32    60
#> 414    40 4    32    60
#> 415    40 4    32    60
#> 416    40 4    32    60
#> 417    40 4    32    60
#> 418    40 4    32    60
#> 419    40 4    32    60
#> 420    40 4    32    60
#> 421    40 4    32    60
#> 422    40 4    32    60
#> 423    40 4    32    60
#> 424    40 4    32    60
#> 425    40 4    32    60
#> 426    40 4    32    60
#> 427    40 4    32    60
#> 428    40 4    32    60
#> 429    40 4    32    60
#> 430    40 4    32    60
#> 431    40 4    32    60
#> 432    40 4    32    60
#> 433    40 4    32    60
#> 434    40 4    32    60
#> 435    40 4    32    60
#> 436    40 4    32    60
#> 437    40 4    32    60
#> 438    40 4    32    60
#> 439    40 4    32    60
#> 440    40 4    32    60
#> 441    40 4    32    60
#> 442    40 4    32    60
#> 443    40 4    32    60
#> 444    40 4    32    60
#> 445    40 4    32    60
#> 446    40 4    32    60
#> 447    40 4    32    60
#> 448    40 4    32    60
#> 449    40 4    32    60
#> 450    40 4    32    60
#> 451    20 5    16    60
#> 452    20 5    16    60
#> 453    20 5    16    60
#> 454    20 5    16    60
#> 455    20 5    16    60
#> 456    20 5    16    60
#> 457    20 5    16    60
#> 458    20 5    16    60
#> 459    20 5    16    60
#> 460    20 5    16    60
#> 461    20 5    16    60
#> 462    20 5    16    60
#> 463    20 5    16    60
#> 464    20 5    16    60
#> 465    20 5    16    60
#> 466    20 5    16    60
#> 467    20 5    16    60
#> 468    20 5    16    60
#> 469    20 5    16    60
#> 470    20 5    16    60
#> 471    20 5    16    60
#> 472    20 5    16    60
#> 473    20 5    16    60
#> 474    20 5    16    60
#> 475    20 5    16    60
#> 476    20 5    16    60
#> 477    20 5    16    60
#> 478    20 5    16    60
#> 479    20 5    16    60
#> 480    20 5    16    60
#> 481    20 5    16    60
#> 482    20 5    16    60
#> 483    20 5    16    60
#> 484    20 5    16    60
#> 485    20 5    16    60
#> 486    20 5    16    60
#> 487    20 5    16    60
#> 488    20 5    16    60
#> 489    20 5    16    60
#> 490    20 5    16    60
#> 491    20 5    16    60
#> 492    20 5    16    60
#> 493    20 5    16    60
#> 494    20 5    16    60
#> 495    20 5    16    60
#> 496    20 5    16    60
#> 497    20 5    16    60
#> 498    20 5    16    60
#> 499    20 5    16    60
#> 500    20 5    16    60
#> 501    30 5    24    60
#> 502    30 5    24    60
#> 503    30 5    24    60
#> 504    30 5    24    60
#> 505    30 5    24    60
#> 506    30 5    24    60
#> 507    30 5    24    60
#> 508    30 5    24    60
#> 509    30 5    24    60
#> 510    30 5    24    60
#> 511    30 5    24    60
#> 512    30 5    24    60
#> 513    30 5    24    60
#> 514    30 5    24    60
#> 515    30 5    24    60
#> 516    30 5    24    60
#> 517    30 5    24    60
#> 518    30 5    24    60
#> 519    30 5    24    60
#> 520    30 5    24    60
#> 521    30 5    24    60
#> 522    30 5    24    60
#> 523    30 5    24    60
#> 524    30 5    24    60
#> 525    30 5    24    60
#> 526    30 5    24    60
#> 527    30 5    24    60
#> 528    30 5    24    60
#> 529    30 5    24    60
#> 530    30 5    24    60
#> 531    30 5    24    60
#> 532    30 5    24    60
#> 533    30 5    24    60
#> 534    30 5    24    60
#> 535    30 5    24    60
#> 536    30 5    24    60
#> 537    30 5    24    60
#> 538    30 5    24    60
#> 539    30 5    24    60
#> 540    30 5    24    60
#> 541    30 5    24    60
#> 542    30 5    24    60
#> 543    30 5    24    60
#> 544    30 5    24    60
#> 545    30 5    24    60
#> 546    30 5    24    60
#> 547    30 5    24    60
#> 548    30 5    24    60
#> 549    30 5    24    60
#> 550    30 5    24    60
#> 551    40 5    32    60
#> 552    40 5    32    60
#> 553    40 5    32    60
#> 554    40 5    32    60
#> 555    40 5    32    60
#> 556    40 5    32    60
#> 557    40 5    32    60
#> 558    40 5    32    60
#> 559    40 5    32    60
#> 560    40 5    32    60
#> 561    40 5    32    60
#> 562    40 5    32    60
#> 563    40 5    32    60
#> 564    40 5    32    60
#> 565    40 5    32    60
#> 566    40 5    32    60
#> 567    40 5    32    60
#> 568    40 5    32    60
#> 569    40 5    32    60
#> 570    40 5    32    60
#> 571    40 5    32    60
#> 572    40 5    32    60
#> 573    40 5    32    60
#> 574    40 5    32    60
#> 575    40 5    32    60
#> 576    40 5    32    60
#> 577    40 5    32    60
#> 578    40 5    32    60
#> 579    40 5    32    60
#> 580    40 5    32    60
#> 581    40 5    32    60
#> 582    40 5    32    60
#> 583    40 5    32    60
#> 584    40 5    32    60
#> 585    40 5    32    60
#> 586    40 5    32    60
#> 587    40 5    32    60
#> 588    40 5    32    60
#> 589    40 5    32    60
#> 590    40 5    32    60
#> 591    40 5    32    60
#> 592    40 5    32    60
#> 593    40 5    32    60
#> 594    40 5    32    60
#> 595    40 5    32    60
#> 596    40 5    32    60
#> 597    40 5    32    60
#> 598    40 5    32    60
#> 599    40 5    32    60
#> 600    40 5    32    60
#> 601    20 6    16    60
#> 602    20 6    16    60
#> 603    20 6    16    60
#> 604    20 6    16    60
#> 605    20 6    16    60
#> 606    20 6    16    60
#> 607    20 6    16    60
#> 608    20 6    16    60
#> 609    20 6    16    60
#> 610    20 6    16    60
#> 611    20 6    16    60
#> 612    20 6    16    60
#> 613    20 6    16    60
#> 614    20 6    16    60
#> 615    20 6    16    60
#> 616    20 6    16    60
#> 617    20 6    16    60
#> 618    20 6    16    60
#> 619    20 6    16    60
#> 620    20 6    16    60
#> 621    20 6    16    60
#> 622    20 6    16    60
#> 623    20 6    16    60
#> 624    20 6    16    60
#> 625    20 6    16    60
#> 626    20 6    16    60
#> 627    20 6    16    60
#> 628    20 6    16    60
#> 629    20 6    16    60
#> 630    20 6    16    60
#> 631    20 6    16    60
#> 632    20 6    16    60
#> 633    20 6    16    60
#> 634    20 6    16    60
#> 635    20 6    16    60
#> 636    20 6    16    60
#> 637    20 6    16    60
#> 638    20 6    16    60
#> 639    20 6    16    60
#> 640    20 6    16    60
#> 641    20 6    16    60
#> 642    20 6    16    60
#> 643    20 6    16    60
#> 644    20 6    16    60
#> 645    20 6    16    60
#> 646    20 6    16    60
#> 647    20 6    16    60
#> 648    20 6    16    60
#> 649    20 6    16    60
#> 650    20 6    16    60
#> 651    30 6    24    60
#> 652    30 6    24    60
#> 653    30 6    24    60
#> 654    30 6    24    60
#> 655    30 6    24    60
#> 656    30 6    24    60
#> 657    30 6    24    60
#> 658    30 6    24    60
#> 659    30 6    24    60
#> 660    30 6    24    60
#> 661    30 6    24    60
#> 662    30 6    24    60
#> 663    30 6    24    60
#> 664    30 6    24    60
#> 665    30 6    24    60
#> 666    30 6    24    60
#> 667    30 6    24    60
#> 668    30 6    24    60
#> 669    30 6    24    60
#> 670    30 6    24    60
#> 671    30 6    24    60
#> 672    30 6    24    60
#> 673    30 6    24    60
#> 674    30 6    24    60
#> 675    30 6    24    60
#> 676    30 6    24    60
#> 677    30 6    24    60
#> 678    30 6    24    60
#> 679    30 6    24    60
#> 680    30 6    24    60
#> 681    30 6    24    60
#> 682    30 6    24    60
#> 683    30 6    24    60
#> 684    30 6    24    60
#> 685    30 6    24    60
#> 686    30 6    24    60
#> 687    30 6    24    60
#> 688    30 6    24    60
#> 689    30 6    24    60
#> 690    30 6    24    60
#> 691    30 6    24    60
#> 692    30 6    24    60
#> 693    30 6    24    60
#> 694    30 6    24    60
#> 695    30 6    24    60
#> 696    30 6    24    60
#> 697    30 6    24    60
#> 698    30 6    24    60
#> 699    30 6    24    60
#> 700    30 6    24    60
#> 701    40 6    32    60
#> 702    40 6    32    60
#> 703    40 6    32    60
#> 704    40 6    32    60
#> 705    40 6    32    60
#> 706    40 6    32    60
#> 707    40 6    32    60
#> 708    40 6    32    60
#> 709    40 6    32    60
#> 710    40 6    32    60
#> 711    40 6    32    60
#> 712    40 6    32    60
#> 713    40 6    32    60
#> 714    40 6    32    60
#> 715    40 6    32    60
#> 716    40 6    32    60
#> 717    40 6    32    60
#> 718    40 6    32    60
#> 719    40 6    32    60
#> 720    40 6    32    60
#> 721    40 6    32    60
#> 722    40 6    32    60
#> 723    40 6    32    60
#> 724    40 6    32    60
#> 725    40 6    32    60
#> 726    40 6    32    60
#> 727    40 6    32    60
#> 728    40 6    32    60
#> 729    40 6    32    60
#> 730    40 6    32    60
#> 731    40 6    32    60
#> 732    40 6    32    60
#> 733    40 6    32    60
#> 734    40 6    32    60
#> 735    40 6    32    60
#> 736    40 6    32    60
#> 737    40 6    32    60
#> 738    40 6    32    60
#> 739    40 6    32    60
#> 740    40 6    32    60
#> 741    40 6    32    60
#> 742    40 6    32    60
#> 743    40 6    32    60
#> 744    40 6    32    60
#> 745    40 6    32    60
#> 746    40 6    32    60
#> 747    40 6    32    60
#> 748    40 6    32    60
#> 749    40 6    32    60
#> 750    40 6    32    60