Get statistics

Get statistics

# S4 method for ConsensusPartitionList
get_stats(object, k, all_stats = FALSE)

Arguments

object
A ConsensusPartitionList-class object.

k
Number of subgroups. The value can only be a single value.

all_stats
Whether to show all statistics that were calculated. Used internally.

Value

A matrix of partition statistics for a selected k. Rows in the matrix correspond to combinations of top-value methods and partitioning methods.

Examples

data(golub_cola)
get_stats(golub_cola, k = 2)

#>             k     1-PAC mean_silhouette concordance area_increased      Rand
#> SD:NMF      2 0.6334504       0.8232491   0.9275000      0.4892486 0.5121283
#> CV:NMF      2 0.6334504       0.7872886   0.9157222      0.4909393 0.5070423
#> MAD:NMF     2 0.6799824       0.8109892   0.9250556      0.4898557 0.5070423
#> ATC:NMF     2 0.9719052       0.9266103   0.9734444      0.5047301 0.4945227
#> SD:skmeans  2 0.6167691       0.6592361   0.8646667      0.4970037 0.5121283
#> CV:skmeans  2 0.6286216       0.7477727   0.8901111      0.4969645 0.4945227
#> MAD:skmeans 2 0.6395961       0.7279295   0.8836667      0.4967373 0.4933490
#> ATC:skmeans 2 1.0000000       0.9767344   0.9905000      0.5051585 0.4964789
#> SD:mclust   2 0.3730849       0.8011169   0.8177778      0.4100016 0.5179969
#> CV:mclust   2 0.4890269       0.9083927   0.8949444      0.4313367 0.5320814
#> MAD:mclust  2 0.8235294       0.9265781   0.9318889      0.4416306 0.5320814
#> ATC:mclust  2 0.3595259       0.7109781   0.8275000      0.4257086 0.4929577
#> SD:kmeans   2 0.6624232       0.8355983   0.9114444      0.4793021 0.5402973
#> CV:kmeans   2 0.6527656       0.8059867   0.9048889      0.4836249 0.5320814
#> MAD:kmeans  2 0.6231884       0.8054912   0.9036111      0.4827648 0.5402973
#> ATC:kmeans  2 1.0000000       0.9815350   0.9915000      0.5025009 0.4992175
#> SD:pam      2 0.5478490       0.6864620   0.8758889      0.4897065 0.4929577
#> CV:pam      2 0.5460931       0.7840100   0.9035556      0.4957916 0.4964789
#> MAD:pam     2 0.6514486       0.8236402   0.9253889      0.4956635 0.4992175
#> ATC:pam     2 0.5974539       0.8083550   0.9176111      0.4754645 0.5179969
#> SD:hclust   2 0.2357331       0.7767183   0.8512778      0.4347693 0.5590767
#> CV:hclust   2 0.2139812       0.5697315   0.8053333      0.4537550 0.5121283
#> MAD:hclust  2 0.2265145       0.7367577   0.8377222      0.4437254 0.5492958
#> ATC:hclust  2 0.5737489       0.8133644   0.9105556      0.4863788 0.4945227
#>               Jaccard
#> SD:NMF      0.5121283
#> CV:NMF      0.5070423
#> MAD:NMF     0.5070423
#> ATC:NMF     0.4945227
#> SD:skmeans  0.5121283
#> CV:skmeans  0.4945227
#> MAD:skmeans 0.4933490
#> ATC:skmeans 0.4964789
#> SD:mclust   0.5179969
#> CV:mclust   0.5320814
#> MAD:mclust  0.5320814
#> ATC:mclust  0.4929577
#> SD:kmeans   0.5402973
#> CV:kmeans   0.5320814
#> MAD:kmeans  0.5402973
#> ATC:kmeans  0.4992175
#> SD:pam      0.4929577
#> CV:pam      0.4964789
#> MAD:pam     0.4992175
#> ATC:pam     0.5179969
#> SD:hclust   0.5590767
#> CV:hclust   0.5121283
#> MAD:hclust  0.5492958
#> ATC:hclust  0.4945227