Hierarchical partition

Hierarchical partition

hierarchical_partition(data,
    top_value_method = "ATC",
    partition_method = "skmeans",
    PAC_cutoff = 0.2, min_samples = 6, subset = Inf,
    max_k = 4, verbose = TRUE, mc.cores = 1, ...)

Arguments

data
a numeric matrix where subgroups are found by columns.

top_value_method
a single top-value method. Available methods are in all_top_value_methods.

partition_method
a single partition method. Available methods are in all_partition_methods.

PAC_cutoff
the cutoff of PAC scores to determine whether to continue looking for subgroups.

min_samples
the cutoff of number of samples to determine whether to continue looking for subgroups.

subset
Number of columns to randomly sample.

max_k
maximal number of partitions to try. The function will try 2:max_k partitions. Note this is the number of partitions that will be tried out on each node of the hierarchical partition. Since more subgroups will be found in the whole partition hierarchy, on each node, max_k should not be set to a large value.

verbose
whether print message.

mc.cores
multiple cores to use.

...
pass to consensus_partition

Details

The function looks for subgroups in a hierarchical way.

There is a special way to encode the node in the hierarchy. The length of the node name is the depth of the node in the hierarchy and the substring excluding the last digit is the name node of the parent node. E.g. for the node 0011, the depth is 4 and the parent node is 001.

Value

A HierarchicalPartition-class object. Simply type object in the interactive R session to see which functions can be applied on it.

Examples

# There is no example
NULL

#> NULL