Consensus partition for all combinations of methods

Consensus partition for all combinations of methods

run_all_consensus_partition_methods(data,
    top_value_method = all_top_value_methods(),
    partition_method = all_partition_methods(),
    max_k = 6,
    top_n = seq(min(1000, round(nrow(data)*0.1)),
    min(5000, round(nrow(data)*0.5)),
    length.out = 5),
    mc.cores = 1, anno = NULL, anno_col = NULL,
    sample_by = "row", p_sampling = 0.8, partition_repeat = 50,
    scale_rows = NULL, verbose = TRUE)

Arguments

data
A numeric matrix where subgroups are found by columns.

top_value_method
Method which are used to extract top n rows. Allowed methods are in all_top_value_methods and can be self-added by register_top_value_methods.

partition_method
Method which are used to do partition on samples. Allowed methods are in all_partition_methods and can be self-added by register_partition_methods.

max_k
Maximal number of partitions to try. The function will try 2:max_k partitions.

top_n
Number of rows with top values. The value can be a vector with length > 1. When n > 5000, the function only randomly sample 5000 rows from top n rows. If top_n is a vector, paritition will be applied to every values in top_n and consensus partition is summarized from all partitions.

mc.cores
Number of cores to use.

anno
A data frame with known annotation of columns.

anno_col
A list of colors (color is defined as a named vector) for the annotations. If anno is a data frame, anno_col should be a named list where names correspond to the column names in anno.

sample_by
Should randomly sample the matrix by rows or by columns?

p_sampling
Proportion of the top n rows to sample.

partition_repeat
Number of repeats for the random sampling.

scale_rows
Whether to scale rows. If it is TRUE, scaling method defined in register_partition_methods is used.

verbose
Ahether to print messages.

Details

The function runs consensus partitioning by consensus_partition for all combinations of top-value methods and partition methods.

It also adjsuts the class IDs for all methods and for all k to make them as consistent as possible.

Value

A ConsensusPartitionList-class object. Simply type object in the interactive R session to see which functions can be applied on it.

Examples

if (FALSE) {
set.seed(123)
m = cbind(rbind(matrix(rnorm(20*20, mean = 1), nr = 20),
                matrix(rnorm(20*20, mean = -1), nr = 20)),
          rbind(matrix(rnorm(20*20, mean = -1), nr = 20),
                matrix(rnorm(20*20, mean = 1), nr = 20))
         ) + matrix(rnorm(40*40), nr = 40)
rl = run_all_consensus_partition_methods(data = m, top_n = c(20, 30, 40))
}
data(cola_rl)
cola_rl

#> A 'ConsensusPartitionList' object with 20 methods.
#>   On a matrix with 60 rows and 60 columns.
#>   Top rows are extracted by 'sd, cv, MAD, ATC' methods.
#>   Subgroups are detected by 'hclust, kmeans, skmeans, pam, mclust' method.
#>   Number of partitions are tried for k = 2, 3, 4, 5, 6.
#>   Performed in total 15000 partitions by row resampling.
#> 
#> Following methods can be applied to this 'ConsensusPartitionList' object:
#>  [1] "cola_report"           "collect_classes"       "collect_plots"        
#>  [4] "collect_stats"         "colnames"              "functional_enrichment"
#>  [7] "get_anno_col"          "get_anno"              "get_classes"          
#> [10] "get_matrix"            "get_membership"        "get_stats"            
#> [13] "is_best_k"             "is_stable_k"           "ncol"                 
#> [16] "nrow"                  "rownames"              "show"                 
#> [19] "suggest_best_k"        "test_to_known_factors" "top_rows_heatmap"     
#> [22] "top_rows_overlap"     
#> 
#> You can get result for a single method by, e.g. object["sd", "hclust"] or object["sd:hclust"]
#> or a subset of methods by object[c("sd", "cv")], c("hclust", "kmeans")]