Suggest the best number of subgroups

Suggest the best number of subgroups

# S4 method for ConsensusPartitionList
suggest_best_k(object, jaccard_index_cutoff = 0.95)

Arguments

object
A ConsensusPartitionList-class object.

jaccard_index_cutoff
The cutoff for Jaccard index for comparing to previous k.

Details

It basically gives the best k for each combination of top-value method and partitioning method by calling suggest_best_k,ConsensusPartition-method.

1-PAC score higher than 0.95 is treated as very stable partition (marked by **) and higher than 0.9 is treated as stable partition (marked by *).

Value

A data frame with the best k and other statistics for each combination of methods.

Examples

data(golub_cola)
suggest_best_k(golub_cola)

#>             best_k     1-PAC mean_silhouette concordance    optional_k
#> ATC:kmeans       2 1.0000000       0.9815350   0.9915000 **           
#> ATC:NMF          2 0.9719052       0.9266103   0.9734444 **           
#> ATC:skmeans      3 0.9609306       0.9318403   0.9731667 **          2
#> MAD:skmeans      3 0.8656716       0.8722804   0.9462778              
#> CV:skmeans       3 0.8371378       0.8565841   0.9377778              
#> MAD:mclust       2 0.8235294       0.9265781   0.9318889              
#> SD:skmeans       3 0.7445127       0.8364565   0.9178889              
#> CV:kmeans        3 0.7388060       0.8643745   0.9012778              
#> MAD:NMF          2 0.6799824       0.8109892   0.9250556              
#> SD:kmeans        2 0.6624232       0.8355983   0.9114444              
#> MAD:pam          2 0.6514486       0.8236402   0.9253889              
#> SD:mclust        5 0.6338894       0.7209468   0.7903333              
#> SD:NMF           2 0.6334504       0.8232491   0.9275000              
#> CV:NMF           2 0.6334504       0.7872886   0.9157222              
#> MAD:kmeans       2 0.6231884       0.8054912   0.9036111              
#> ATC:pam          2 0.5974539       0.8083550   0.9176111              
#> ATC:hclust       2 0.5737489       0.8133644   0.9105556              
#> SD:pam           2 0.5478490       0.6864620   0.8758889              
#> CV:pam           2 0.5460931       0.7840100   0.9035556              
#> CV:mclust        2 0.4890269       0.9083927   0.8949444              
#> ATC:mclust       2 0.3595259       0.7109781   0.8275000              
#> SD:hclust        2 0.2357331       0.7767183   0.8512778              
#> MAD:hclust       2 0.2265145       0.7367577   0.8377222              
#> CV:hclust        2 0.2139812       0.5697315   0.8053333