Suggest the best number of partitions

Suggest the best number of partitions

# S4 method for ConsensusPartitionList
suggest_best_k(object, jaccard_index_cutoff = 0.95)

Arguments

object
A ConsensusPartitionList-class object.

jaccard_index_cutoff
The cutoff for Jaccard index compared to previous k.

Details

It basically gives the best k for each combination of top-value method and partition method by calling suggest_best_k,ConsensusPartition-method.

1-PAC score higher than 0.95 is treated as very stable partition and higher than 0.9 is treated as stable partition.

Value

A data frame with the best k and other statistics for each combination of methods.

Examples

data(cola_rl)
suggest_best_k(cola_rl)

#>             best_k     1-PAC mean_silhouette concordance   
#> sd:kmeans        3 1.0000000       0.9842509   0.9786667 **
#> sd:skmeans       3 1.0000000       0.9884698   0.9941111 **
#> sd:mclust        3 1.0000000       0.9853272   0.9926667 **
#> MAD:skmeans      3 1.0000000       0.9884390   0.9938889 **
#> MAD:mclust       3 1.0000000       0.9764124   0.9864444 **
#> MAD:kmeans       3 0.9993734       0.9822821   0.9842222 **
#> cv:kmeans        3 0.9586466       0.9633184   0.9530000 **
#> cv:skmeans       3 0.9548872       0.9650830   0.9810000 **
#> cv:mclust        3 0.9411028       0.9593721   0.9720000  *
#> sd:hclust        3 0.9229323       0.9421395   0.9608889  *
#> MAD:hclust       3 0.8208020       0.9303737   0.9433333   
#> ATC:skmeans      2 0.7650376       0.8553168   0.9441111   
#> cv:hclust        3 0.7305764       0.9212745   0.9271111   
#> ATC:kmeans       2 0.7105263       0.8228562   0.9275556   
#> sd:pam           3 0.6748120       0.8736248   0.9153333   
#> MAD:pam          3 0.6591479       0.8278094   0.9061111   
#> ATC:mclust       3 0.6541353       0.8519580   0.9005556   
#> cv:pam           2 0.6165414       0.8826227   0.9270000   
#> ATC:pam          2 0.4899749       0.8106902   0.9093333   
#> ATC:hclust       2 0.4028822       0.7172447   0.8663333