Test correspondance between predicted classes and known factors

Test correspondance between predicted classes and known factors

# S4 method for ConsensusPartitionList
test_to_known_factors(object, k, known = get_anno(object),
    silhouette_cutoff = 0.5, verbose = FALSE)

Arguments

object
A ConsensusPartitionList-class object.

k
Number of subgroups. It uses all k if it is not set.

known
A vector or a data frame with known factors. By default it is the annotation table set in consensus_partition or run_all_consensus_partition_methods.

silhouette_cutoff
Cutoff for sihouette scores. Samples with value less than this are omit.

verbose
Whether to print messages.

Details

The function basically sends each ConsensusPartition-class object to test_to_known_factors,ConsensusPartition-method and merges results afterwards.

Value

A data frame with the following columns:

number of samples used to test after filtered by silhouette_cutoff,

p-values from the tests,

number of subgroups.

If there are NA values, basically it means there are no efficient data points to perform the test.

See also

Examples

data(golub_cola)
test_to_known_factors(golub_cola)

#>             n_sample ALL.AML(p-value) k
#> SD:skmeans        71     9.787012e-14 2
#> SD:skmeans        68     9.758604e-14 3
#> SD:skmeans        64     3.746889e-12 4
#> SD:skmeans        51     1.833845e-09 5
#> SD:skmeans        50     4.716429e-08 6
#> CV:skmeans        72     6.197475e-14 2
#> CV:skmeans        66     2.515866e-13 3
#> CV:skmeans        65     2.342532e-12 4
#> CV:skmeans        51     1.535107e-09 5
#> CV:skmeans        46     5.489004e-08 6
#> MAD:skmeans       68     7.462969e-14 2
#> MAD:skmeans       67     1.129838e-12 3
#> MAD:skmeans       64     2.331861e-11 4
#> MAD:skmeans       62     2.661906e-10 5
#> MAD:skmeans       57     1.701241e-09 6
#> ATC:skmeans       70     8.087426e-05 2
#> ATC:skmeans       70     1.513389e-11 3
#> ATC:skmeans       65     1.287644e-10 4
#> ATC:skmeans       65     1.248501e-10 5
#> ATC:skmeans       50     1.813889e-07 6
#> SD:mclust         71     1.101449e-13 2
#> SD:mclust         64     9.817743e-14 3
#> SD:mclust         63     9.953511e-13 4
#> SD:mclust         53     8.521753e-11 5
#> SD:mclust         59     1.153528e-10 6
#> CV:mclust         70     2.935936e-14 2
#> CV:mclust         62     2.603259e-13 3
#> CV:mclust         55     6.866170e-12 4
#> CV:mclust         67     6.406547e-13 5
#> CV:mclust         62     2.646219e-11 6
#> MAD:mclust        71     1.376488e-13 2
#> MAD:mclust        62     2.603259e-13 3
#> MAD:mclust        42     5.158558e-09 4
#> MAD:mclust        58     5.414219e-11 5
#> MAD:mclust        54     1.350221e-09 6
#> ATC:mclust        66     3.089800e-08 2
#> ATC:mclust        48     1.394425e-09 3
#> ATC:mclust        43     6.854721e-07 4
#> ATC:mclust        61     1.386070e-11 5
#> ATC:mclust        49     1.391932e-08 6
#> SD:kmeans         72     8.753640e-14 2
#> SD:kmeans         66     2.515866e-13 3
#> SD:kmeans         64     8.208053e-14 4
#> SD:kmeans         56     1.439947e-10 5
#> SD:kmeans         54     2.095826e-10 6
#> CV:kmeans         72     8.753640e-14 2
#> CV:kmeans         67     1.125075e-12 3
#> CV:kmeans         59     1.542811e-13 4
#> CV:kmeans         59     2.948276e-11 5
#> CV:kmeans         67     4.308865e-13 6
#> MAD:kmeans        72     8.753640e-14 2
#> MAD:kmeans        68     7.078512e-13 3
#> MAD:kmeans        62     4.410533e-11 4
#> MAD:kmeans        43     6.086963e-08 5
#> MAD:kmeans        64     1.344445e-11 6
#> ATC:kmeans        71     1.048918e-04 2
#> ATC:kmeans        57     4.130662e-06 3
#> ATC:kmeans        44     8.460215e-07 4
#> ATC:kmeans        69     5.835515e-08 5
#> ATC:kmeans        64     4.024347e-08 6
#> SD:pam            56     3.783178e-12 2
#> SD:pam            52     3.625007e-11 3
#> SD:pam            56     1.044420e-09 4
#> SD:pam            47     7.788677e-08 5
#> SD:pam            47     1.730812e-07 6
#> CV:pam            70     1.442614e-08 2
#> CV:pam            60     7.338267e-13 3
#> CV:pam            64     4.233462e-12 4
#> CV:pam            41     1.616661e-01 5
#> CV:pam            34     9.215267e-06 6
#> MAD:pam           65     5.986142e-09 2
#> MAD:pam           59     2.492884e-10 3
#> MAD:pam           65     2.342532e-12 4
#> MAD:pam           40     1.709504e-01 5
#> MAD:pam           55     9.163279e-09 6
#> ATC:pam           60     3.098991e-05 2
#> ATC:pam           65     1.662539e-06 3
#> ATC:pam           64     4.820634e-07 4
#> ATC:pam           47     7.052487e-06 5
#> ATC:pam           27     3.967360e-01 6
#> SD:hclust         70     5.693071e-16 2
#> SD:hclust         29     1.440395e-06 3
#> SD:hclust         27     1.370959e-06 4
#> SD:hclust         51     2.232218e-10 5
#> SD:hclust         43     1.034787e-08 6
#> CV:hclust         67     2.607731e-15 2
#> CV:hclust         51     8.423464e-12 3
#> CV:hclust         41     6.539203e-09 4
#> CV:hclust         39     1.735665e-08 5
#> CV:hclust         33     4.230072e-06 6
#> MAD:hclust        69     7.559128e-15 2
#> MAD:hclust        54     1.879529e-12 3
#> MAD:hclust        51     4.892086e-11 4
#> MAD:hclust        52     1.379454e-10 5
#> MAD:hclust        44     6.415777e-09 6
#> ATC:hclust        67     2.650907e-05 2
#> ATC:hclust        36     2.918909e-04 3
#> ATC:hclust        59     1.595357e-03 4
#> ATC:hclust        63     1.896250e-06 5
#> ATC:hclust        47     9.919402e-07 6