Test correspondance between predicted classes and known factors

Test correspondance between predicted classes and known factors

# S4 method for ConsensusPartitionList
test_to_known_factors(object, k, known = get_anno(object),
    silhouette_cutoff = 0.5, verbose = FALSE)

Arguments

object
A ConsensusPartitionList-class object.

k
Number of subgroups. It uses all k if it is not set.

known
A vector or a data frame with known factors. By default it is the annotation table set in consensus_partition or run_all_consensus_partition_methods.

silhouette_cutoff
Cutoff for sihouette scores. Samples with value less than this are omit.

verbose
Whether to print messages.

Details

The function basically sends each ConsensusPartition-class object to test_to_known_factors,ConsensusPartition-method and merges results afterwards.

Value

A data frame with the following columns:

number of samples used to test after filtered by silhouette_cutoff,

p-values from the tests,

number of subgroups.

If there are NA values, basically it means there are no efficient data points to perform the test.

See also

Examples

data(golub_cola)
test_to_known_factors(golub_cola)

#>             n_sample ALL.AML(p-value) k
#> SD:NMF            66     1.345569e-12 2
#> SD:NMF            59     7.380186e-12 3
#> SD:NMF            38     5.336125e-09 4
#> SD:NMF            49     9.126510e-09 5
#> SD:NMF            33     3.220673e-07 6
#> CV:NMF            63     7.819991e-13 2
#> CV:NMF            61     1.738550e-11 3
#> CV:NMF            48     3.775135e-11 4
#> CV:NMF            47     4.063260e-09 5
#> CV:NMF            37     1.801303e-07 6
#> MAD:NMF           65     2.131136e-12 2
#> MAD:NMF           67     8.262374e-12 3
#> MAD:NMF           48     3.775135e-11 4
#> MAD:NMF           47     2.266453e-08 5
#> MAD:NMF           38     2.024822e-07 6
#> ATC:NMF           69     5.413259e-05 2
#> ATC:NMF           65     2.973163e-08 3
#> ATC:NMF           62     1.001843e-09 4
#> ATC:NMF           48     4.167716e-08 5
#> ATC:NMF           44     1.867310e-07 6
#> SD:skmeans        54     1.014943e-11 2
#> SD:skmeans        65     4.402878e-13 3
#> SD:skmeans        64     6.162488e-13 4
#> SD:skmeans        51     4.266678e-09 5
#> SD:skmeans        48     2.142307e-08 6
#> CV:skmeans        55     3.766408e-11 2
#> CV:skmeans        67     1.706206e-13 3
#> CV:skmeans        60     4.200936e-12 4
#> CV:skmeans        45     9.252702e-10 5
#> CV:skmeans        47     9.432819e-09 6
#> MAD:skmeans       67     2.386428e-06 2
#> MAD:skmeans       67     1.712295e-13 3
#> MAD:skmeans       60     4.200936e-12 4
#> MAD:skmeans       44     1.509182e-09 5
#> MAD:skmeans       56     5.177647e-10 6
#> ATC:skmeans       71     5.229760e-05 2
#> ATC:skmeans       70     1.850005e-12 3
#> ATC:skmeans       63     8.358601e-12 4
#> ATC:skmeans       60     2.598166e-10 5
#> ATC:skmeans       51     1.218846e-07 6
#> SD:mclust         70     2.935936e-14 2
#> SD:mclust         60     6.914400e-13 3
#> SD:mclust         35     2.994405e-07 4
#> SD:mclust         64     2.684745e-12 5
#> SD:mclust         58     1.740816e-10 6
#> CV:mclust         70     2.935936e-14 2
#> CV:mclust         62     2.603259e-13 3
#> CV:mclust         62     1.779177e-12 4
#> CV:mclust         66     9.879568e-13 5
#> CV:mclust         53     1.779079e-09 6
#> MAD:mclust        72     8.753640e-14 2
#> MAD:mclust        61     4.154801e-13 3
#> MAD:mclust        63     1.060773e-12 4
#> MAD:mclust        63     4.793373e-12 5
#> MAD:mclust        47     9.398221e-09 6
#> ATC:mclust        62     1.255391e-12 2
#> ATC:mclust        46     5.783239e-10 3
#> ATC:mclust        46     6.398594e-09 4
#> ATC:mclust        51     1.678413e-09 5
#> ATC:mclust        49     2.219252e-09 6
#> SD:kmeans         71     2.076259e-14 2
#> SD:kmeans         66     2.723130e-13 3
#> SD:kmeans         63     9.953511e-13 4
#> SD:kmeans         50     7.989179e-11 5
#> SD:kmeans         56     5.626716e-10 6
#> CV:kmeans         66     2.145933e-13 2
#> CV:kmeans         68     1.057032e-13 3
#> CV:kmeans         57     1.778561e-11 4
#> CV:kmeans         57     7.592622e-11 5
#> CV:kmeans         61     5.204540e-11 6
#> MAD:kmeans        65     3.433859e-13 2
#> MAD:kmeans        68     1.061159e-13 3
#> MAD:kmeans        60     4.200936e-12 4
#> MAD:kmeans        58     5.314248e-11 5
#> MAD:kmeans        60     9.089505e-11 6
#> ATC:kmeans        72     1.496779e-04 2
#> ATC:kmeans        60     2.702128e-11 3
#> ATC:kmeans        56     7.621262e-09 4
#> ATC:kmeans        56     9.681021e-08 5
#> ATC:kmeans        62     8.552577e-08 6
#> SD:pam            54     1.014943e-11 2
#> SD:pam            58     1.939484e-12 3
#> SD:pam            63     6.737013e-12 4
#> SD:pam            45     1.805686e-08 5
#> SD:pam            31     3.961440e-01 6
#> CV:pam            65     1.035297e-07 2
#> CV:pam            61     4.365043e-13 3
#> CV:pam            64     4.109809e-12 4
#> CV:pam            50     3.201917e-09 5
#> CV:pam            38     2.029612e-07 6
#> MAD:pam           63     2.894708e-07 2
#> MAD:pam           24               NA 3
#> MAD:pam           58     6.372531e-11 4
#> MAD:pam           37     4.816509e-07 5
#> MAD:pam           49     2.888254e-08 6
#> ATC:pam           67     1.646876e-04 2
#> ATC:pam           59     3.743474e-06 3
#> ATC:pam           57     5.469317e-07 4
#> ATC:pam           46     1.047418e-05 5
#> ATC:pam           47     1.078180e-06 6
#> SD:hclust         70     4.696330e-15 2
#> SD:hclust         61     5.675685e-14 3
#> SD:hclust         51     3.345563e-10 4
#> SD:hclust         55     2.430419e-10 5
#> SD:hclust         48     9.437836e-10 6
#> CV:hclust         49     1.249196e-10 2
#> CV:hclust         59     1.231506e-12 3
#> CV:hclust         54     1.060706e-10 4
#> CV:hclust         35     4.645348e-07 5
#> CV:hclust         30     1.380057e-06 6
#> MAD:hclust        69     5.205581e-14 2
#> MAD:hclust        53     1.003667e-10 3
#> MAD:hclust        48     7.210093e-09 4
#> MAD:hclust        51     1.573967e-09 5
#> MAD:hclust        52     4.991274e-09 6
#> ATC:hclust        66     4.160902e-05 2
#> ATC:hclust        62     3.035429e-03 3
#> ATC:hclust        30     8.731023e-01 4
#> ATC:hclust        52     8.514077e-06 5
#> ATC:hclust        52     1.573431e-06 6