Test correspondance between predicted classes and known factors

Test correspondance between predicted classes and known factors

# S4 method for ConsensusPartitionList
test_to_known_factors(object, k, known = get_anno(object),
    silhouette_cutoff = 0.5, verbose = FALSE)

Arguments

object

A ConsensusPartitionList-class object.

k

Number of subgroups. It uses all k if it is not set.

known

A vector or a data frame with known factors. By default it is the annotation table set in consensus_partition or run_all_consensus_partition_methods.

silhouette_cutoff

Cutoff for sihouette scores. Samples with value less than this are omit.

verbose

Whether to print messages.

Details

The function basically sends each ConsensusPartition-class object to test_to_known_factors,ConsensusPartition-method and merges results afterwards.

Value

A data frame with the following columns:

number of samples used to test after filtered by silhouette_cutoff,

p-values from the tests,

number of subgroups.

If there are NA values, basically it means there are no efficient data points to perform the test.

See also

Author

Zuguang Gu <z.gu@dkfz.de>

Examples

data(golub_cola)
test_to_known_factors(golub_cola)
#>             n_sample ALL.AML(p-value) k
#> SD:skmeans        72     3.256935e-13 2
#> SD:skmeans        69     6.082538e-14 3
#> SD:skmeans        57     6.112061e-10 4
#> SD:skmeans        60     6.592515e-10 5
#> SD:skmeans        60     4.259907e-10 6
#> CV:skmeans        71     9.787012e-14 2
#> CV:skmeans        66     1.481309e-12 3
#> CV:skmeans        58     6.372531e-11 4
#> CV:skmeans        49     3.517737e-09 5
#> CV:skmeans        51     5.083683e-09 6
#> MAD:skmeans       71     9.787012e-14 2
#> MAD:skmeans       69     3.422182e-13 3
#> MAD:skmeans       65     1.432104e-11 4
#> MAD:skmeans       51     1.066152e-08 5
#> MAD:skmeans       61     4.989276e-09 6
#> ATC:skmeans       72     1.505160e-04 2
#> ATC:skmeans       72     1.826171e-10 3
#> ATC:skmeans       70     2.170533e-09 4
#> ATC:skmeans       59     1.351324e-08 5
#> ATC:skmeans       46     7.476557e-07 6
#> SD:mclust         70     1.547171e-13 2
#> SD:mclust         72     8.542092e-14 3
#> SD:mclust         61     8.286055e-11 4
#> SD:mclust         65     5.349855e-11 5
#> SD:mclust         54     5.262681e-11 6
#> CV:mclust         71     1.101449e-13 2
#> CV:mclust         66     1.481309e-12 3
#> CV:mclust         55     1.139992e-12 4
#> CV:mclust         58     3.660309e-11 5
#> CV:mclust         65     4.130359e-11 6
#> MAD:mclust        70     1.733409e-13 2
#> MAD:mclust        68     5.882548e-13 3
#> MAD:mclust        66     8.880807e-12 4
#> MAD:mclust        65     6.031869e-11 5
#> MAD:mclust        60     5.508749e-10 6
#> ATC:mclust        72     1.574505e-02 2
#> ATC:mclust        67     6.537953e-04 3
#> ATC:mclust        58     7.968710e-06 4
#> ATC:mclust        65     6.286670e-10 5
#> ATC:mclust        60     4.838478e-09 6
#> SD:kmeans         72     6.197475e-14 2
#> SD:kmeans         70     3.557696e-14 3
#> SD:kmeans         50     5.607336e-10 4
#> SD:kmeans         35     2.028995e-06 5
#> SD:kmeans         45     3.975960e-09 6
#> CV:kmeans         72     5.108677e-13 2
#> CV:kmeans         68     6.926395e-13 3
#> CV:kmeans         34     4.139938e-08 4
#> CV:kmeans         55     1.882447e-10 5
#> CV:kmeans         59     1.256845e-10 6
#> MAD:kmeans        72     5.108677e-13 2
#> MAD:kmeans        64     3.743085e-12 3
#> MAD:kmeans        44     9.575551e-09 4
#> MAD:kmeans        59     6.096370e-10 5
#> MAD:kmeans        56     8.594749e-09 6
#> ATC:kmeans        70     2.718960e-04 2
#> ATC:kmeans        57     1.973871e-05 3
#> ATC:kmeans        60     1.155630e-06 4
#> ATC:kmeans        70     6.522482e-08 5
#> ATC:kmeans        66     2.039385e-07 6
#> SD:pam            72     5.108677e-13 2
#> SD:pam            63     7.411494e-12 3
#> SD:pam            62     1.060516e-11 4
#> SD:pam            59     1.929880e-10 5
#> SD:pam            52     4.256044e-09 6
#> CV:pam            69     1.757005e-10 2
#> CV:pam            58     1.840146e-12 3
#> CV:pam            67     9.171718e-13 4
#> CV:pam            57     1.034494e-10 5
#> CV:pam            26     1.608709e-04 6
#> MAD:pam           69     3.414214e-13 2
#> MAD:pam           60     4.847143e-12 3
#> MAD:pam           66     1.140267e-11 4
#> MAD:pam           66     3.563719e-11 5
#> MAD:pam           61     2.460744e-10 6
#> ATC:pam           70     8.115728e-07 2
#> ATC:pam           69     1.509330e-06 3
#> ATC:pam           71     1.518569e-05 4
#> ATC:pam           64     3.510948e-06 5
#> ATC:pam           48     7.882661e-06 6
#> SD:hclust         70     5.435049e-16 2
#> SD:hclust         68     1.713908e-15 3
#> SD:hclust         62     2.197099e-13 4
#> SD:hclust         53     8.521753e-11 5
#> SD:hclust         45     3.975960e-09 6
#> CV:hclust         67     2.755841e-15 2
#> CV:hclust         55     1.139992e-12 3
#> CV:hclust         43     4.599055e-10 4
#> CV:hclust         40     1.065509e-08 5
#> CV:hclust         36     2.893696e-07 6
#> MAD:hclust        69     6.975197e-15 2
#> MAD:hclust        59     1.542811e-13 3
#> MAD:hclust        29     5.043477e-07 4
#> MAD:hclust        33     3.220673e-07 5
#> MAD:hclust        51     8.648732e-10 6
#> ATC:hclust        66     7.528671e-05 2
#> ATC:hclust        57     8.433404e-05 3
#> ATC:hclust        57     7.720606e-03 4
#> ATC:hclust        54     1.021776e-05 5
#> ATC:hclust        54     1.021776e-05 6