Test correspondance between predicted classes and known factors

Test correspondance between predicted classes and known factors

# S4 method for ConsensusPartitionList
test_to_known_factors(object, k, known = get_anno(object),
    silhouette_cutoff = 0.5, verbose = FALSE)

Arguments

object
A ConsensusPartitionList-class object.

k
Number of partitions. It uses all k if it is not set.

known
A vector or a data frame with known factors. By default it is the annotation table set in consensus_partition or run_all_consensus_partition_methods.

silhouette_cutoff
Cutoff for sihouette scores. Samples with value less than this are omit.

verbose
Whether to print messages.

Details

The function basically sends each ConsensusPartition-class object to test_to_known_factors,ConsensusPartition-method and merges results afterwards.

Value

A data frame with columns:

number of samples used to test after filtered by silhouette_cutoff

p-values from the tests

number of partitions

See also

Examples

data(cola_rl)
test_to_known_factors(cola_rl, known = 1:40)

#>              n     known(p) k
#> sd:skmeans  58 5.168296e-13 2
#> sd:skmeans  60 5.168296e-13 3
#> sd:skmeans  57 5.168296e-13 4
#> sd:skmeans  45 6.428704e-05 5
#> sd:skmeans  20 2.615879e-02 6
#> cv:skmeans  57           NA 2
#> cv:skmeans  60 5.168296e-13 3
#> cv:skmeans  60 5.168296e-13 4
#> cv:skmeans  42 2.006264e-02 5
#> cv:skmeans   0           NA 6
#> MAD:skmeans 54 5.168296e-13 2
#> MAD:skmeans 60 5.168296e-13 3
#> MAD:skmeans 49 1.441904e-12 4
#> MAD:skmeans 31           NA 5
#> MAD:skmeans 18           NA 6
#> ATC:skmeans 55 5.168296e-13 2
#> ATC:skmeans 56 5.168296e-13 3
#> ATC:skmeans 42 1.712029e-12 4
#> ATC:skmeans 38 5.209289e-12 5
#> ATC:skmeans 34 3.375703e-10 6
#> sd:mclust   60 5.168296e-13 2
#> sd:mclust   60 5.168296e-13 3
#> sd:mclust   60 5.168296e-13 4
#> sd:mclust   35           NA 5
#> sd:mclust   34 5.118827e-06 6
#> cv:mclust   60           NA 2
#> cv:mclust   60 5.168296e-13 3
#> cv:mclust   60 5.168296e-13 4
#> cv:mclust   57 5.168296e-13 5
#> cv:mclust   24 4.865012e-03 6
#> MAD:mclust  59 5.168296e-13 2
#> MAD:mclust  60 5.168296e-13 3
#> MAD:mclust  60 5.168296e-13 4
#> MAD:mclust  53 1.305343e-08 5
#> MAD:mclust  26 1.047882e-01 6
#> ATC:mclust  49 5.168296e-13 2
#> ATC:mclust  58 5.168296e-13 3
#> ATC:mclust  56 4.581443e-13 4
#> ATC:mclust  39 1.488786e-12 5
#> ATC:mclust  16 9.010221e-01 6
#> sd:pam      45 2.290483e-07 2
#> sd:pam      59 1.770675e-12 3
#> sd:pam      43 4.893648e-12 4
#> sd:pam      30 2.754290e-03 5
#> sd:pam      16           NA 6
#> cv:pam      57           NA 2
#> cv:pam      57 1.426193e-11 3
#> cv:pam      48 1.928910e-08 4
#> cv:pam      37 7.646741e-05 5
#> cv:pam      26 6.766401e-02 6
#> MAD:pam     58 3.567443e-12 2
#> MAD:pam     56 7.338144e-12 3
#> MAD:pam     38 2.896696e-11 4
#> MAD:pam     32 2.750913e-06 5
#> MAD:pam      1           NA 6
#> ATC:pam     54 4.581443e-13 2
#> ATC:pam     53           NA 3
#> ATC:pam     33           NA 4
#> ATC:pam     18           NA 5
#> ATC:pam     12 7.730048e-01 6
#> sd:hclust   60 5.168296e-13 2
#> sd:hclust   60 5.168296e-13 3
#> sd:hclust   60 5.168296e-13 4
#> sd:hclust   43 4.617864e-03 5
#> sd:hclust   35           NA 6
#> cv:hclust   39           NA 2
#> cv:hclust   60 5.168296e-13 3
#> cv:hclust   60 5.168296e-13 4
#> cv:hclust   40 9.365686e-07 5
#> cv:hclust   33 9.974796e-03 6
#> MAD:hclust  60 5.168296e-13 2
#> MAD:hclust  60 5.168296e-13 3
#> MAD:hclust  60 5.168296e-13 4
#> MAD:hclust  32           NA 5
#> MAD:hclust  29           NA 6
#> ATC:hclust  52 5.168296e-13 2
#> ATC:hclust  53 1.543376e-11 3
#> ATC:hclust  49 9.317917e-11 4
#> ATC:hclust  30 1.629250e-09 5
#> ATC:hclust  31 1.313136e-09 6
#> sd:kmeans   59 5.168296e-13 2
#> sd:kmeans   60 5.168296e-13 3
#> sd:kmeans   60 5.168296e-13 4
#> sd:kmeans   45 5.168296e-13 5
#> sd:kmeans   16           NA 6
#> cv:kmeans   60           NA 2
#> cv:kmeans   60 5.168296e-13 3
#> cv:kmeans   60 5.168296e-13 4
#> cv:kmeans   40 5.168296e-13 5
#> cv:kmeans    6           NA 6
#> MAD:kmeans  39 1.441904e-12 2
#> MAD:kmeans  60 5.168296e-13 3
#> MAD:kmeans  60 5.168296e-13 4
#> MAD:kmeans  42 4.695351e-08 5
#> MAD:kmeans  40 1.644612e-09 6
#> ATC:kmeans  52 5.168296e-13 2
#> ATC:kmeans  59 5.168296e-13 3
#> ATC:kmeans  56 4.519811e-12 4
#> ATC:kmeans  29 3.591702e-04 5
#> ATC:kmeans   8 6.119776e-01 6